% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{actual_vapor_pressure}
\alias{actual_vapor_pressure}
\title{Calculate actual vapor pressure from dewpoint temperature}
\usage{
actual_vapor_pressure(Tdewpoint)
}
\arguments{
\item{Tdewpoint}{dewpoint temperature (C)}
}
\value{
actual vapor pressure, e_a (kPa)
}
\description{
Calculate actual vapor pressure from dewpoint temperature
}
\details{
Calculate actual vapor pressure from dewpoint temperature based on Stull 2000. Source: Riddell EA. 2017. Physical calculations of resistance to water loss improve predictions of species range models. Ecological Monographs 87: 21-23.
}
\examples{
\dontrun{
actual_vapor_pressure(Tdewpoint=20)
}

}
\author{
Eric Riddell
}
\keyword{actual}
\keyword{pressure}
\keyword{vapor}
