% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempWindProfileFunctions.R
\name{air_temp_profile}
\alias{air_temp_profile}
\title{Estimate air temperature profile as in NicheMapR}
\usage{
air_temp_profile(T_r, u_r, zr, z0, z, T_s)
}
\arguments{
\item{T_r}{is temperature at reference height in C.}

\item{u_r}{is windspeed at reference height in m/s.}

\item{zr}{is initial reference height in m.}

\item{z0}{is surface roughness in m.}

\item{z}{is height to scale in m.}

\item{T_s}{is surface temperatures in C.}
}
\value{
temperature (C)
}
\description{
This function allows you to estimate temperature (C) at a specified height (m).  
Estimates a single, unsegmented temperature profile using the MICRO routine from NicheMapR. Source: Kearney MR and Porter WP. 2016. NicheMapR – an R package for biophysical modelling: the microclimate model. Ecography 40:664-674.
}
\details{
Estimate temperature at a specified height
}
\examples{
\dontrun{
air_temp_profile(T_r=20, u_r=0.1, zr=0.1, z0=0.2, z=0.15, T_s=25)
}

}
\seealso{
Other microclimate functions: 
\code{\link{air_temp_profile_neutral}()},
\code{\link{air_temp_profile_segment}()},
\code{\link{degree_days}()},
\code{\link{direct_solar_radiation}()},
\code{\link{diurnal_radiation_variation}()},
\code{\link{diurnal_temp_variation_sineexp}()},
\code{\link{diurnal_temp_variation_sinesqrt}()},
\code{\link{diurnal_temp_variation_sine}()},
\code{\link{estimate_radiation}()},
\code{\link{monthly_solar_radiation}()},
\code{\link{partition_solar_radiation}()},
\code{\link{proportion_diffuse_solar_radiation}()},
\code{\link{surface_roughness}()},
\code{\link{wind_speed_profile_neutral}()},
\code{\link{wind_speed_profile_segment}()}
}
\concept{microclimate functions}
\keyword{profile}
\keyword{temperature}
