% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_To.R
\docType{data}
\name{alpha_L}
\alias{alpha_L}
\title{Calculate Operative Temperature of Lizard using model by Fei et al.}
\format{An object of class \code{numeric} of length 1.}
\usage{
alpha_L
}
\arguments{
\item{TAH}{Temperature in canopy.}

\item{Tsurface}{Surface Temperature.}

\item{Tair}{Air Temperature.}

\item{SWDOWN}{upper developmental threshold.}

\item{GLW}{Downward flux of near-infrared radiation}

\item{shade}{type of method being used}
}
\description{
[1] T. Fei, A. K. Skidmore, V. Venus, T. Wang, M. Schlerf, B. Toxopeus, 
S. van Overjijk, M. Bian, and Y. Liu, “A body temperature model for 
lizards as estimated from the thermal environment,” J. Therm. 
Biol., vol. 37, no. 1, pp. 56–64, 2012.
TAH, Tsurface, Tair, SWDOWN, GLW, time_step, shade
}
\details{
Calculate Operative Temperature of Lizard.
}
\examples{
\dontrun{
calculate_To(254, 265, 267, 15, 307,  0)
}

}
\author{
ofir
}
\keyword{Fei}
\keyword{Operative}
\keyword{Temperature}
