% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{angle conversions}
\alias{angle conversions}
\alias{radians_to_degrees}
\alias{degrees_to_radians}
\title{Convert Angles Between Radians and Degrees}
\usage{
radians_to_degrees(rad)

degrees_to_radians(deg)
}
\arguments{
\item{rad}{\code{numeric} angle (radians).}

\item{deg}{\code{numeric} angle (degrees).}
}
\value{
\code{numeric} angle (degrees or radians).
}
\description{
The function converts angles in radians to degrees or degrees to radians.
}
\examples{
  radians_to_degrees(0.831)
  degrees_to_radians(47.608)

}
