% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YearMean.R
\name{AverageIndexBins}
\alias{AverageIndexBins}
\title{Means from index binning.}
\usage{
AverageIndexBins(x, ind, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector for which to compute bin means according to the
indices given in \code{ind}.}

\item{ind}{numeric vector of indices specifying the binning windows to
average over; i.e. first binning window is from index number 1 to 2, second
binning window from index number 2 to 3, etc.}

\item{na.rm}{a logical value indicating whether ‘NA’ values should be
stripped before the computation proceeds. Defaults to \code{TRUE}.}
}
\value{
a numeric vector of length \code{length(ind) - 1} with the mean
values of each bin with end points defined by \code{ind}.
}
\description{
Calculate bin means defined by vector index positions.
}
\examples{
x <- rnorm(20)
x.bin <- AverageIndexBins(x, ind = c(1, 7, 13, 17, 20))
}
\author{
Thomas Münch
}
