% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{boundary_layer_resistance}
\alias{boundary_layer_resistance}
\title{Estimate the boundary layer resistance}
\usage{
boundary_layer_resistance(T_a, e_s, e_a, elev, D, u = NA)
}
\arguments{
\item{T_a}{air temperature (K)}

\item{e_s}{saturation vapor pressure (kPa)}

\item{e_a}{actual vapor pressure (kPa)}

\item{elev}{elevation (m)}

\item{D}{characteristic dimension (e.g., body diameter) (m)}

\item{u}{is wind speed in m/s, if not provided assume free convection; if provided, use forced convection if appropriate}
}
\value{
boundary layer resistance (s cm^-1)
}
\description{
Estimate the boundary layer resistance
}
\details{
This function allows you to estimate boundary layer resistance under free convection. Based on the function in Riddell et al. 2017. Source: Riddell EA. 2017. Physical calculations of resistance to water loss improve predictions of species range models. Ecological Monographs 87: 21-23.
}
\examples{
\dontrun{
boundary_layer_resistance(T_a=293, e_s=2.5, e_a=2.0, elev=500, D=0.007, u=2)
}

}
\seealso{
Other biophysical models: \code{\link{Free_or_forced_convection}},
  \code{\link{Grashof_number_Gates}},
  \code{\link{Grashof_number}}, \code{\link{Nu_from_Gr}},
  \code{\link{Nu_from_Re}}, \code{\link{Nusselt_number}},
  \code{\link{Prandtl_number}},
  \code{\link{Qconduction_animal}},
  \code{\link{Qconduction_substrate}},
  \code{\link{Qconvection}},
  \code{\link{Qemitted_thermal_radiation}},
  \code{\link{Qevaporation}},
  \code{\link{Qmetabolism_from_mass_temp}},
  \code{\link{Qmetabolism_from_mass}},
  \code{\link{Qnet_Gates}},
  \code{\link{Qradiation_absorbed}},
  \code{\link{Qthermal_radiation_absorbed}},
  \code{\link{Reynolds_number}},
  \code{\link{Tb_CampbellNorman}}, \code{\link{Tb_Fei}},
  \code{\link{Tb_Gates}}, \code{\link{Tb_butterfly}},
  \code{\link{Tb_grasshopper}}, \code{\link{Tb_limpetBH}},
  \code{\link{Tb_limpet}}, \code{\link{Tb_lizard}},
  \code{\link{Tb_mussel}},
  \code{\link{Tb_salamander_humid}},
  \code{\link{Tb_snail}}, \code{\link{Tbed_mussel}},
  \code{\link{Tsoil}}, \code{\link{actual_vapor_pressure}},
  \code{\link{external_resistance_to_water_vapor_transfer}},
  \code{\link{heat_transfer_coefficient_approximation}},
  \code{\link{heat_transfer_coefficient_simple}},
  \code{\link{heat_transfer_coefficient}},
  \code{\link{saturation_vapor_pressure}},
  \code{\link{saturation_water_vapor_pressure}}
}
\author{
Eric Riddell
}
\concept{biophysical models}
\keyword{boundary}
\keyword{layer}
\keyword{resistance}
