% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{calculate_actual_vapor_pressure}
\alias{calculate_actual_vapor_pressure}
\title{Calculate actual vapor pressure from dewpoint temperature}
\usage{
calculate_actual_vapor_pressure(Tdewpoint)
}
\arguments{
\item{Tdewpoint}{dewpoint temperature (C)}
}
\value{
actual vapor pressure (kPa)
}
\description{
Calculate actual vapor pressure from dewpoint temperature
}
\details{
Calculate actual vapor pressure from dewpoint temperature based on Stull 2000
}
\examples{
\dontrun{
calculate_actual_vapor_pressure(Tdewpoint=20)
}

}
\author{
Eric Riddell
}
\keyword{actual}
\keyword{pressure}
\keyword{vapor}
