% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{calculate_boundary_layer_resistance}
\alias{calculate_boundary_layer_resistance}
\title{Estimate the boundary layer resistance}
\usage{
calculate_boundary_layer_resistance(Ta, e_s, e_a, elev, D, u = NA)
}
\arguments{
\item{Ta}{air temperature (K)}

\item{e_s}{saturation vapor pressure (kPa)}

\item{e_a}{actual vapor pressure (kPa)}

\item{elev}{elevation (m)}

\item{D}{characteristic dimension (e.g., body diameter) (m), ##diameter = 0.0016*log(mass) + 0.0061 for mass(g) #empirical formula for salamander diameter, Riddell et al. 2017}

\item{u}{is wind speed in m/s, if not provided assume free convection; if provided, use forced convection if appropriate}
}
\value{
boundary layer resistance (s cm^-1)
}
\description{
Estimate the boundary layer resistance
}
\details{
This function allows you to estimate boundary layer resistance under free convection. Based on the function in Riddell et al. 2017
}
\examples{
\dontrun{
calculate_boundary_layer_resistance(Ta=293, e_s=2.4, e_a=2.5, elev=500, D=0.007, u=2)
}

}
\author{
Eric Riddell
}
\keyword{boundary}
\keyword{layer}
\keyword{resistance}
