% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllometricFunctions.R
\name{calculate_mass}
\alias{calculate_mass}
\title{Calculate mass from length}
\usage{
calculate_mass(length, taxa)
}
\arguments{
\item{length}{length in m, length is snout-vent length for amphibians and reptiles (excepting turtles where length is carapace length).}

\item{taxa}{Which class of organism, current choices: insect, lizard, salamander, frog, snake, turtle}
}
\value{
mass (g)
}
\description{
Calculate mass from length
}
\details{
This function allows you to estimate mass (g) from length (m) for a variety of taxa
}
\examples{
 \dontrun{
 calculate_mass(0.04,"insect")
}

}
\keyword{length}
\keyword{mass}
