% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{calculate_sa}
\alias{calculate_sa}
\title{Calculate surface area from mass/density (Based on Mitchell 1976)}
\usage{
calculate_sa(mass, density, taxa = "lizard")
}
\arguments{
\item{mass}{Mass in grams.}

\item{density}{Density in grams/cm^3.}

\item{taxa}{Which class of organism, current choice: lizard.}
}
\value{
sa (m^2)
}
\description{
Calculate surface area from mass/density (Based on Mitchell 1976)
}
\details{
This function allows you to calculate surface area (m^2) from mass (g) and density(g/cm^3) for a variety of taxa
}
\examples{
 \dontrun{
 calculate_sa(100,0.45,"lizard")
}

}
\keyword{area}
\keyword{surface}
