% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{calculate_sa_length}
\alias{calculate_sa_length}
\title{Calculate surface area from length (Based on Mitchell 1976)}
\usage{
calculate_sa_length(length, taxa = "lizard")
}
\arguments{
\item{length}{Characteristic dimension (trunk diameter or SVL) in meter.}

\item{taxa}{Which class of organism, current choice: lizard,frog or general 'sphere'.}
}
\value{
sa (m^2)
}
\description{
Calculate surface area from length (Based on Mitchell 1976)
}
\details{
This function allows you to calculate surface area (m^2) from length(m) for a variety of taxa
}
\examples{
 \dontrun{
 calculate_sa_length(.05,"lizard")
}

}
\keyword{area}
\keyword{length}
\keyword{surface}
