% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllometricFunctions.R
\name{calculate_surface_area}
\alias{calculate_surface_area}
\title{Calculate surface area from mass}
\usage{
calculate_surface_area(mass, taxa = "lizard")
}
\arguments{
\item{mass}{Mass in grams.}

\item{taxa}{Which class of organism, current choices: lizard, frog, insect}
}
\value{
sa (m^2)
}
\description{
Calculate surface area from mass
}
\details{
This function allows you to calculate surface area (m^2) from mass (g) for a variety of taxa
}
\examples{
 \dontrun{
 calculate_surface_area(2,"insect")
}

}
\keyword{area}
\keyword{surface}
