% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllometricFunctions.R
\name{calculate_surface_area_length_ellipsoid}
\alias{calculate_surface_area_length_ellipsoid}
\title{Calculate surface area from length approximating the animal's body as a rotational ellipsoid}
\usage{
calculate_surface_area_length_ellipsoid(length)
}
\arguments{
\item{length}{length in m.}
}
\value{
sa (m^2)
}
\description{
Calculate surface area from length approximating the animal's body as a rotational ellipsoid
}
\details{
Calculate surface area (m^2) from length (m) approximating the animal's body as a rotational ellipsoid with half the body length as the semi-major axis q.
}
\examples{
 \dontrun{
 calculate_surface_area_length_ellipsoid(length=0.04)
}

}
\keyword{area}
\keyword{surface}
