% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllometricFunctions.R
\name{calculate_volume_length}
\alias{calculate_volume_length}
\title{Calculate volume from length (Based on Mitchell 1976)}
\usage{
calculate_volume_length(length, taxa)
}
\arguments{
\item{length}{in meters.}

\item{taxa}{Which class of organism, current choices: lizard,frog, or sphere}
}
\value{
sa (m^2)
}
\description{
Calculate volume from length (Based on Mitchell 1976)
}
\details{
This function allows you to calculate volume (m^3) from length(m) for a variety of taxa. Approximates volume for use in a model of convection, which approximates animal shape as a sphere.
}
\examples{
 \dontrun{
 calculate_volume_length(.05,"lizard")
}

}
\keyword{area}
\keyword{length}
\keyword{surface}
