% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompressRecord.R
\name{CompressRecord}
\alias{CompressRecord}
\title{Compress a record}
\usage{
CompressRecord(depth, rec, stretch)
}
\arguments{
\item{depth}{numeric vector woth the original depth scale of \code{rec}.}

\item{rec}{numeric vector of proxy values; must be of same length as
\code{depth}.}

\item{stretch}{numeric value of the amount of compression of the original
depth scale in the same units as \code{depth}.}
}
\value{
a list with two components:
\itemize{
  \item depth: numeric vector of length \code{depth} with the compressed
    depth scale.
  \item rec: numeric vector of length \code{depth} with the proxy values on
    the compressed depth scale.
}
}
\description{
Linearly interpolate given data onto a compressed depth scale.
}
\details{
The compressed depth scale is calculated by reducing the maximum depth of
\code{depth} by the amount of \code{stretch} and dividing the new depth
scale into \code{n} equal bins, where \code{n} is the length of
\code{depth}. The proxy values on the compressed depth scale are found by
linear interpolation of the original values using \code{\link{approx}}.
}
\seealso{
\code{\link{approx}}
}
\author{
Thomas Münch
}
