% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{conductance_animal}
\alias{conductance_animal}
\title{Calculate conductance assuming animal thermal conductivity is rate limiting}
\usage{
conductance_animal(Ts, Tb, lambda, K = 0.5, sa, proportion)
}
\arguments{
\item{Ts}{Surface Temperature in Kelvin.}

\item{Tb}{Body Temperature in Kelvin.}

\item{lambda}{Mean thickness of the animal skin (surface) in meters, assumes well mixed interior}

\item{K}{Thermal conductivity (W K^-1 m^-1 ), K=0.5 W K^-1 m^-1 for naked skin, K=0.15 for insect cuticle ( Galushko et al 2005); conductivity of ground is generally greater than that of animal tissues, so animal thermal conductivity is generally rate limiting step.}

\item{sa}{Surface area  in m^2}

\item{proportion}{In contact to the surface}
}
\value{
conductance (W)
}
\description{
Calculate conductance assuming animal thermal conductivity is rate limiting
}
\details{
This function allows you to calculate conductance (W) of an ectothermic animal to its substate. Method assumes the major resistance to conduction is within surface layers of the animal and that the interior of the animal is equal in temperature to its surface (thermally well mixed). Reference: Spotila et al. 1992. Biophysicas of Heat and Mass Transfer.
}
\examples{
\dontrun{
conductance_animal(Ts= 293,Tb=303,lambda=10^-6,K=0.5,sa=10^-3, proportion=0.2)
}

}
\keyword{conductance}
