% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{day_of_year}
\alias{day_of_year}
\title{Julian Day from Date}
\usage{
day_of_year(day, format = "\%Y-\%m-\%d")
}
\arguments{
\item{day}{R object that is suitable to be converted to a date.}

\item{format}{\code{character} date format following \code{\link[base]{POSIXlt}}" conventions}
}
\value{
\code{numeric} Julian day number, 1-366, for eg. 1 for January 1st
}
\description{
Convert a date to a Julian Day.
}
\examples{
  day_of_year("2017-04-22", format= "\%Y-\%m-\%d")

}
