% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{dec_angle}
\alias{dec_angle}
\title{Calculate solar declination in radians}
\usage{
dec_angle(doy)
}
\arguments{
\item{doy}{day of year}
}
\value{
declination angle in radians
}
\description{
This function allows you to calculate solar declination, which is the angular distance of the sun north or south of the earth’s equator, based on the day of year
}
\details{
Calculate solar declination in radians
}
\examples{
\dontrun{
dec_angle(112)
}
}
\keyword{Declination}
\keyword{angle}
