% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDFunctions.R
\name{degree.days.mat}
\alias{degree.days.mat}
\title{Calculate degree days(using single sine wave approximation)}
\usage{
degree.days.mat(Tdat, LDT)
}
\arguments{
\item{Tdat}{- Vector containing Tmin and Tmax.}

\item{LDT}{lower developmental threshold.}
}
\description{
Single sine wave approximation from Baskerville & Emin 1969
Double Sine wave approximation of degree days from Allen 1976 
(see http://www.ipm.ucdavis.edu/WEATHER/ddss_tbl.html)
}
\details{
This function allows you to calculate degree days using single sine wave approximation.
}
\examples{
degree.days.mat()

}
\keyword{alternate}
\keyword{days}
\keyword{degree}
