% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDFunctions.R
\name{degree.days}
\alias{degree.days}
\title{Calculate degree days(using single sine wave approximation)}
\usage{
degree.days(Tmin, Tmax, LDT)
}
\arguments{
\item{Tmin}{Minimum temperature of the day.}

\item{Tmax}{Maximum temperature of the day.}

\item{LDT}{lower developmental threshold.}
}
\description{
Single sine wave approximation from Baskerville & Emin 1969
Double Sine wave approximation of degree days from Allen 1976 
(see http://www.ipm.ucdavis.edu/WEATHER/ddss_tbl.html)
}
\details{
This function allows you to calculate degree days using single sine wave approximation.
}
\examples{
degree.days()

}
\keyword{days}
\keyword{degree}
