% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDFunctions.R
\name{degree.days.thermoreg}
\alias{degree.days.thermoreg}
\title{Calculate degree days using single sine wave approximation with thermoregulation
 
This function allows you to calculate degree days using single sine wave approximation assuming thermoregulation to Topt.}
\usage{
degree.days.thermoreg(Tmin, Tmax, LDT, Topt, Trep)
}
\arguments{
\item{Tmin}{Minimum temperature of the day.}

\item{Tmax}{Maximum temperature of the day.}

\item{LDT}{lower developmental threshold.}

\item{Topt}{.}

\item{Trep}{.}
}
\description{
Calculate degree days using single sine wave approximation with thermoregulation
 
This function allows you to calculate degree days using single sine wave approximation assuming thermoregulation to Topt.
}
\examples{
degree.days.thermoreg()

}
\keyword{days}
\keyword{degree}
\keyword{thermoregulation}
