% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDFunctions.R
\name{degree_days}
\alias{degree_days}
\title{Calculate degree days}
\usage{
degree_days(T_min, T_max, LDT = NA, UDT = NA, method = "single.sine")
}
\arguments{
\item{T_min}{Minimum temperature of the day (C)}

\item{T_max}{Maximum temperature of the day (C)}

\item{LDT}{lower developmental threshold.}

\item{UDT}{upper developmental threshold.}

\item{method}{type of method being used. Current choices: "single.sine","double.sine", "single.triangulation" or "double.triangulation".}
}
\description{
Single sine wave approximation from  
Double Sine wave approximation of degree days from Allen 1976 
(see http://www.ipm.ucdavis.edu/WEATHER/ddss_tbl.html)
}
\details{
This function allows you to calculate degree days using single or double sine wave and single or double triangulation approximation.
}
\examples{
\dontrun{
degree_days(T_min=7, T_max=14, LDT=12, UDT=33, method="single.sine")
degree_days(T_min=7, T_max=14, LDT=12, UDT=33, method="single.triangulation")
}

}
\keyword{days}
\keyword{degree}
