% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{degree_to_radian}
\alias{degree_to_radian}
\title{@title Converts angle in degrees to radians}
\usage{
degree_to_radian(deg)
}
\arguments{
\item{deg}{\code{numeric} angle in degrees.}
}
\value{
\code{numeric} angle in radians.
}
\description{
Convert angle in degrees to radians.
}
\examples{
  degree_to_radian(47.608)

}
\seealso{
Other utility functions: 
\code{\link{airpressure_from_elev}()},
\code{\link{azimuth_angle}()},
\code{\link{daylength}()},
\code{\link{dec_angle}()},
\code{\link{fahrenheit_to_celsius}()},
\code{\link{fahrenheit_to_kelvin}()},
\code{\link{kelvin_to_celsius}()},
\code{\link{solar_noon}()},
\code{\link{zenith_angle}()}
}
\concept{utility functions}
