% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{degrees_to_radians}
\alias{degrees_to_radians}
\title{Convert Angle in Degrees to Radians}
\usage{
degrees_to_radians(deg)
}
\arguments{
\item{deg}{\code{numeric} angle (degrees).}
}
\value{
\code{numeric} angle (radians).
}
\description{
Convert angle in degrees to radians.
}
\examples{
  degrees_to_radians(47.608)

}
\seealso{
Other utility functions: 
\code{\link{airpressure_from_elev}()},
\code{\link{azimuth_angle}()},
\code{\link{day_of_year}()},
\code{\link{daylength}()},
\code{\link{dec_angle}()},
\code{\link{solar_noon}()},
\code{\link{temperature conversions}},
\code{\link{zenith_angle}()}
}
\concept{utility functions}
