% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensificationRate.R
\name{DensificationRate}
\alias{DensificationRate}
\title{Firn densification rate.}
\usage{
DensificationRate(dat = NULL, lim)
}
\arguments{
\item{dat}{a list providing a firn density profile with two
elements: \code{depth}, a vector with the depth scale of the density
profile in metre; and \code{density}, a vector of firn densities. If
\code{NULL} (the default), the stacked B41/B42 density profile, provided
with the package, is used for the calculation (see Münch et al., 2017).}

\item{lim}{the depth (in m) over which the densification rate shall be
calculated.}
}
\value{
the relative linear densification rate (in \% per metre) from zero
depth to the depth \code{lim}.
}
\description{
Function to return the relative densification rate in \% per metre over a
specified depth, given a record of firn densities. The densification rate is
found from linear regression of the firn density profile against depth.
}
\references{
Münch, T., et al., Constraints on post-depositional isotope modifications
in East Antarctic firn from analysing temporal changes of isotope profiles,
The Cryosphere, doi:10.5194/tc-11-2175-2017, 2017.
}
\author{
Thomas Münch
}
