% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialDiffusion.R
\name{DifferentialDiffusion}
\alias{DifferentialDiffusion}
\title{Differential firn diffusion.}
\source{
Eq. (2) in Münch et al. (2017)
}
\usage{
DifferentialDiffusion(
  dat = NULL,
  z00,
  z01,
  z10,
  z11,
  P = 650,
  T = 228.5,
  bdot = 64
)
}
\arguments{
\item{dat}{a list providing a firn density profile, in order to calculate
diffusion lengths, with two elements: \code{depth}, a vector with the depth
scale of the density profile in [m]; and \code{density}, a vector of firn
densities in [kg/m^3]. If \code{NULL} (the default), the fitted B41/B42
density profile, provided with the package, is used for the calculation (see
Münch et al., 2017).}

\item{z00}{upper boundary of the first depth interval.}

\item{z01}{lower boundary of the first depth interval.}

\item{z10}{upper boundary of the second depth interval.}

\item{z11}{lower boundary of the second depth interval.}

\item{P}{atmospheric surface pressure in [mbar] to calculate diffusion
lengths; default is 650 mbar for Kohnen Station.}

\item{T}{annual mean surface air temperature in [K] to calculate diffusion
lengths; default is 228.5 K for Kohnen Station.}

\item{bdot}{annual mean accumulation rate in [kg/m^2/yr] to calculate
diffusion lengths; default is 64 kg/m^2/yr for Kohnen Station.}
}
\value{
a named vector with two elements: \code{oxy}, the differential
diffusion length value for oxygen isotopes; \code{dtr}, the differential
diffusion length value for deuterium.
}
\description{
Calculate the differential diffusion, as defined in Münch et al. (2007),
over a specified depth interval given a diffusion length profile.
}
\details{
Mean diffusion lengths are calculated for the depth intervals \code{[z00,
z01]} and \code{[z10, z11]}, and the differential diffusion length is
obtained between these two intervals according to Eq. (2) in Münch et
al. (2017).
}
\references{
Münch, T., et al., Constraints on post-depositional isotope modifications
in East Antarctic firn from analysing temporal changes of isotope profiles,
The Cryosphere, 11(5), 2175-2188, doi:10.5194/tc-11-2175-2017, 2017.
}
\author{
Thomas Münch
}
