% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationFunctions.R
\name{direct_solar_radiation}
\alias{direct_solar_radiation}
\title{Estimate direct solar radiation}
\usage{
direct_solar_radiation(lat, doy, elev, t, t0, method = "Campbell 1977")
}
\arguments{
\item{lat}{latitude (degrees)}

\item{doy}{doy of year (1-366)}

\item{elev}{elevation in m}

\item{t}{local time (decimal hours)}

\item{t0}{time of local noon (decimal hours), can be estimated using solar_noon()}

\item{method}{method for estimating direct solar radiation, options: "Campbell 1977","Gates 1962"}
}
\value{
direct solar radiation (W/m^2)
}
\description{
Estimate direct solar radiation (kW/m^2) based on latitude, day of year, elevation, and time. Uses 2 methods compiled in Tracy et al. (1983) Estimating clear-day solar radiation: An evaluation of three models. Journal of Thermal Biology, 8(3), 247-251.
}
\examples{
\dontrun{
direct_solar_radiation(lat=47.61,doy=112,elev=1500,t=9,t0=12, method="Campbell 1977")
}
}
\seealso{
Other microclimate functions: 
\code{\link{air_temp_profile_neutral}()},
\code{\link{air_temp_profile_segment}()},
\code{\link{air_temp_profile}()},
\code{\link{degree_days}()},
\code{\link{diurnal_radiation_variation}()},
\code{\link{diurnal_temp_variation_sineexp}()},
\code{\link{diurnal_temp_variation_sinesqrt}()},
\code{\link{diurnal_temp_variation_sine}()},
\code{\link{estimate_radiation}()},
\code{\link{monthly_solar_radiation}()},
\code{\link{partition_solar_radiation}()},
\code{\link{proportion_diffuse_solar_radiation}()},
\code{\link{surface_roughness}()},
\code{\link{wind_speed_profile_neutral}()},
\code{\link{wind_speed_profile_segment}()}
}
\concept{microclimate functions}
\keyword{radiation}
\keyword{solar}
