% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationFunctions.R
\name{diurnal_radiation_variation}
\alias{diurnal_radiation_variation}
\title{Estimate Diurnal Radiation}
\usage{
diurnal_radiation_variation(doy, solrad, hour, lon, lat)
}
\arguments{
\item{doy}{\code{numeric} is the day of year}

\item{solrad}{\code{numeric} is solar radiation (W/m^2) per day}

\item{hour}{\code{numeric} is hour (0-24)}

\item{lon}{\code{numeric} longitude (degrees)}

\item{lat}{\code{numeric} latitude (degrees)}
}
\value{
\code{numeric} hourly solar radiation (W/m^2)
}
\description{
Estimate hourly solar radiation (W m^-2 per hour) as a function of daily global solar radiation (in W m^-2 per day). Based on \insertCite{Tham2010}{TrenchR} and \insertCite{AlRawahi2011}{TrenchR}.
}
\examples{
diurnal_radiation_variation(doy    = 112, 
                            solrad = 8000, 
                            hour   = 12, 
                            lon    = -122.33, 
                            lat    = 47.61)

}
\references{
\insertAllCited{}
}
\seealso{
Other microclimate functions: 
\code{\link{air_temp_profile_neutral}()},
\code{\link{air_temp_profile_segment}()},
\code{\link{air_temp_profile}()},
\code{\link{degree_days}()},
\code{\link{direct_solar_radiation}()},
\code{\link{diurnal_temp_variation_sineexp}()},
\code{\link{diurnal_temp_variation_sinesqrt}()},
\code{\link{diurnal_temp_variation_sine}()},
\code{\link{estimate_radiation}()},
\code{\link{monthly_solar_radiation}()},
\code{\link{partition_solar_radiation}()},
\code{\link{proportion_diffuse_solar_radiation}()},
\code{\link{surface_roughness}()},
\code{\link{wind_speed_profile_neutral}()},
\code{\link{wind_speed_profile_segment}()}
}
\concept{microclimate functions}
\keyword{diurnal}
\keyword{radiation}
