% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biophysmodel_Sceloporus.R
\name{estimate_lizard_Tb}
\alias{estimate_lizard_Tb}
\title{Predicts body temperatures (operative environmental temperature) of a lizard in degrees C.}
\usage{
estimate_lizard_Tb(Ta, Tg, u, svl, mass, psi, rho_S, elevation, doy,
  sun = TRUE, surface = TRUE, alpha_S = 0.9, alpha_L = 0.965,
  epsilon_s = 0.965, F_d = 0.8, F_r = 0.5, F_a = 0.5, F_g = 0.5)
}
\arguments{
\item{Ta}{is air temperature in C}

\item{Tg}{is surface temperature in C}

\item{u}{is wind speed in m/s}

\item{svl}{is lizard snout vent length in mm}

\item{mass}{is lizard mass in g, note can be estimated as massfromsvl=function(svl) 3.55*10^-5*(svl)^3.00 #Tinkle and Ballinger 1972}

\item{psi}{is solar zenith angle in degrees}

\item{rho_S}{is surface albedo (proportion)}

\item{elevation}{is elevation in m}

\item{doy}{is doy of year (1-366)}

\item{sun}{indicates whether lizard is in sun (TRUE) or shade (FALSE)}

\item{surface}{indicates whether lizard is on ground surface (TRUE) or above surface (FALSE, e.g. in a tree)}

\item{alpha_S}{is lizard solar absorptivity, alpha_S=0.9 (Gates 1980, Table 11.4)}

\item{alpha_L}{is lizard thermal absoptivity, alpha_L=0.965 (Bartlett & Gates 1967)}

\item{epsilon_s}{is surface emisivity of lizards, epsilon_s=0.965 (Bartlett & Gates 1967)}

\item{F_d}{is the view factor between the surface of the lizard and diffuse solar radiation (Bartlett & Gates 1967)}

\item{F_r}{is the view factor between the surface of the lizard and reflected solar radiation}

\item{F_a}{is the view factor between the surface of the lizard and atmospheric radiation}

\item{F_g}{is the view factor between the surface of the lizard and ground thermal radation}
}
\description{
Predicts body temperature (operative environmental temperature) of a lizard in degrees C. Based on Campbell and Norman (1998, An introduction to environmental biophysics). Designed for Sceloporus lizards and described in Buckley (2008, Linking traits to energetics and population dynamics to predict lizard ranges in changing environments. American Naturalist 171: E1-E19).
}
\details{
Predicts body temperatures (operative environmental temperature) of a lizard in degrees C.
}
\keyword{biophysical}
\keyword{body}
\keyword{model}
\keyword{temperature,}
