% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperativeTemperateFunctions.R
\name{estimate_Te_CampbellNorman}
\alias{estimate_Te_CampbellNorman}
\title{Predicts body temperatures (operative environmental temperature) of an ectotherm in K.}
\usage{
estimate_Te_CampbellNorman(Ta, Rabs, emissivity = 0.96, c_p = 29.3, d, u)
}
\arguments{
\item{Ta}{is air temperature in K.}

\item{Rabs}{solar and thermal radiation absorbed (W m^-2)}

\item{emissivity}{longwave infrared emissivity of skin (proportion), 0.95 to 1 for most animals (Gates 1980)}

\item{c_p}{specific heat of air (J mol^-1 K^-1)}

\item{d}{characteristic dimension of the animal (m)}

\item{u}{is wind speed in m/s}
}
\value{
operative environmental temperature (K)
}
\description{
Predicts body temperatures (operative environmental temperature) of an ectotherm in K.
}
\details{
Predicts body temperatures (operative environmental temperature) of an ectotherm in K. Uses approximation in Campbell and Norman (1998).
}
\examples{
\dontrun{
estimate_Te_CampbellNorman(Ta=303, Rabs=823, emissivity=0.96, c_p=29.3, d=0.17, u=1)
}

}
\keyword{environmental}
\keyword{operative}
\keyword{temperature}
