% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperativeTemperateFunctions.R
\name{estimate_Te_Gates}
\alias{estimate_Te_Gates}
\title{Predicts body temperatures (operative environmental temperature) of an ectotherm in K.}
\usage{
estimate_Te_Gates(As, d, psa_dir, psa_ref, psa_air, psa_g, Ts, Ta, Rabs,
  emissivity, h_L, ef = 1.3, K)
}
\arguments{
\item{As}{surface area  in m^2}

\item{d}{characteristic dimension for conduction (m)}

\item{psa_dir}{proportion surface area exposed to sky (or enclosure)}

\item{psa_ref}{proportion surface area exposed to ground}

\item{psa_air}{of surface area exposed to air}

\item{psa_g}{of surface in contact with substrate}

\item{Ts}{ground surface temperatue in K}

\item{Ta}{ambient air temperature in K}

\item{Rabs}{Solar and thermal radiation absorbed (W)}

\item{emissivity}{longwave infrared emissivity of skin (proportion), 0.95 to 1 for most animals (Gates 1980)}

\item{h_L}{Convective heat transfer coefficient (W m^-2 K^-1)}

\item{ef}{is the enhancement factor, used to adjuct h_L to field condictions (using h_L approximation from Mitchell 1976).  Approximated as 1.3 by default, but see Mitchell 1976 for relationship.}

\item{K}{Thermal conductivity (W K^-1 m^-1 ), K=0.5 W K^-1 m^-1 for naked skin, K=0.15 for insect cuticle ( Galushko et al 2005); conductivity of ground is generally greater than that of animal tissues, so animal thermal conductivity is generally rate limiting step.}
}
\value{
operative environmental temperature (K)
}
\description{
Predicts body temperatures (operative environmental temperature) of an ectotherm in K.
}
\details{
Predicts body temperatures (operative environmental temperature) of an ectotherm in K. Uses approximation in Gates (1980, Biophysical ecology). Omits evaporative and metabolic heat loss.
}
\examples{
\dontrun{
estimate_Te_Gates(As=1, d=0.001, psa_dir=0.6, psa_ref=0.4, psa_air=0.6, psa_g=0.2, Ts=303, Ta=310, Rabs=800, emissivity=0.95, h_L=10, ef=1.3, K=0.5)
}

}
\keyword{environmental}
\keyword{operative}
\keyword{temperature}
