% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateInterProfileCorrelation.R
\name{estimateInterProfileCorrelation}
\alias{estimateInterProfileCorrelation}
\title{Correlation between profiles depending on distance}
\usage{
estimateInterProfileCorrelation(
  data,
  distances,
  profilePosition = NULL,
  .var = "d18O",
  rangeTol = 0.05,
  lambda = 0
)
}
\arguments{
\item{data}{either a generic trench data set, or a trench data set in form
of a numeric matrix or data frame, where columns represent the horizontal
trench profile positions and rows represent the vertical scale of the
data.}

\item{distances}{a vector of inter-profile distances for which profile pairs
shall be seeked and analysed for their correlation; must be measured in the
same units as the profile positions of the trench data set.}

\item{profilePosition}{optional vector of the horizontal trench profile
positions; only needed when \code{data} is passed as a simple matrix or
data frame.}

\item{.var}{character string with the name of the trench variable for which to
compute the correlations; see also \code{\link{make2D}}. Only needed when
the \code{data} is passed as a generic trench data set.}

\item{rangeTol}{relative tolerance allowed for profile spacings; i.e. for a
given spacing d, all profile pairs separated by d * (1 +/- rangeTol) are
analysed.}

\item{lambda}{horizontal decorrelation length of the trench data to obtain an
effective number of profiles used for the error estimates (see details);
assumes an AR1 process for the horizontal data variations and must be
measured in the same units as the inter-profile \code{distances}.}
}
\value{
a tibble of five variables: the analysed profile spacings, the found
  number of pairs, the mean correlation between the pairs, and the standard
  deviation and standard error of the mean correlation (see details).
}
\description{
Estimate the pairwise correlation between trench profiles as a function of
the distance between the profiles. To allow for more flexibility, the trench
data can be passed in the generic package form, or as a numeric simple matrix
or data frame where columns represent the horizontal trench profile positions
and rows represent the vertical scale of the data.
}
\details{
For a given distance, the function searches the trench data set for all pairs
which are separated horizontally by this distance, allowing for a given
tolerance in separation distance. It is ensured that profile pairs are
counted only once. For every found pair, the correlation between the two
profiles is calculated using pairwise complete observations, and the
correlations are averaged across all pairs. Uncertainty of the mean
correlation is given by the standard deviation across the correlations from
all pairs, and by the standard error. The latter is obtained from dividing
the standard deviation by the square root of an effective number of
observations, which is calculated from the number of found pairs and the
average horizontal distance between the trench profiles for the given
horizontal decorrelation length (see \code{\link{getEffectiveTrenchDOF}}).
This is still an optimistic error estimate, since e.g. the correlation values
from two profile pairs located at positions [1, 2] and [2, 3] are clearly
not independent of each other, even when accounting for autocorrelation.
}
\examples{

IPC <- t13.trench2 \%>\%
  estimateInterProfileCorrelation(distances = c(10, 30))
# you can pass data also in 2D form,
# but then the profile positions need to be provided seperately
IPC2 <- make2D(t13.trench2) \%>\%
  estimateInterProfileCorrelation(
    profilePosition = getX(t13.trench2),
    distances = c(10, 30))
all.equal(IPC, IPC2) # by definition

}
\seealso{
\code{\link{estimateTrenchDecorrelation}},
  \code{\link{getEffectiveTrenchDOF}}
}
\author{
Thomas Münch
}
