% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateInterTrenchCorrelation.R
\name{estimateInterTrenchCorrelation}
\alias{estimateInterTrenchCorrelation}
\title{Pairwise correlation between trenches}
\usage{
estimateInterTrenchCorrelation(
  t1,
  t2,
  var = "d18O",
  optimize = FALSE,
  lag = NULL
)
}
\arguments{
\item{t1}{a generic trench data set.}

\item{t2}{a second generic trench data set.}

\item{var}{character string with the name of the trench variable for which to
compute the correlations; see also \code{\link{make2D}}.}

\item{optimize}{logical; set to \code{TRUE} to allow for relative shifts
of the records, specified by \code{lag}, to maximize their correlations.}

\item{lag}{a vector of lags, i.e. positive and negative index shifts, used
when \code{optimize = TRUE} to shift the records of one trench relative to
the records of the other trench in order to find their maximum
correlations.}
}
\value{
For \code{optimize = FALSE}, a matrix of all pairwise correlations
  between the records in \code{t1} and in \code{t2}, where rows tabulate the
  profile positions of \code{t1} and columns those of \code{t2}. For
  \code{optimize = TRUE}, a list is returned of two matrices in the same
  format, where component \code{cor} contains the optimized correlations
  and component \code{i} contains the index shifts corresponding to the
  maximum correlations.
}
\description{
Calculate the pairwise correlations of all records in a trench with all
records in a second trench; optionally allowing for relative shifts of the
records with respect to each other in order to maximize their correlations.
}
\examples{

raw <- estimateInterTrenchCorrelation(t13.trench1, t13.trench2)
opt <- estimateInterTrenchCorrelation(t13.trench1, t13.trench2,
                                      optimize = TRUE, lag = -4 : 4)

mean(raw)
mean(opt$cor)

}
\author{
Thomas Münch
}
