% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSNR.R
\name{estimateSNR}
\alias{estimateSNR}
\title{Trench signal-to-noise ratio}
\usage{
estimateSNR(
  data,
  distances,
  profilePosition = NULL,
  .var = "d18O",
  lambda = 0,
  rangeTol = 0.05
)
}
\arguments{
\item{data}{either a generic trench data set, or a trench data set in form
of a numeric matrix or data frame, where columns represent the horizontal
trench profile positions and rows represent the vertical scale of the
data.}

\item{distances}{a vector of inter-profile distances at which average
pairwise profile correlations are to be estimated in order to calculate the
SNR.}

\item{profilePosition}{optional vector of the horizontal trench profile
positions; only needed when \code{data} is passed as a simple matrix or
data frame.}

\item{.var}{character string with the name of the trench variable for which
to estimate the SNR; see also \code{\link{make2D}}. Only needed when the
\code{data} is passed as a generic trench data set.}

\item{lambda}{horizontal decorrelation length of the trench data to obtain an
effective number of profiles used for the error estimates (assuming an AR1
process, see the details in \code{\link{estimateInterProfileCorrelation}});
must be measured in the same units as the inter-profile \code{distances}.}

\item{rangeTol}{relative tolerance allowed for inter-profile distances (see
\code{\link{estimateInterProfileCorrelation}}).}
}
\value{
a tibble of two variables: estimated signal-to-noise ratio
  (\code{snr}) and its standard error (\code{se}).
}
\description{
Estimate the signal-to-noise ratio (SNR) for a trench variable from the
average pairwise correlations evaluated at specified profile distances. To
allow for more flexibility, the trench data can be passed in the generic
package form, or as a simple numeric matrix or data frame where columns
represent the horizontal trench profile positions and rows represent the
vertical scale of the data.
}
\examples{

distances <- c(10, 20, 30)
estimateSNR(t15.trench1, distances, lambda = 1.6)
estimateSNR(t15.trench1, distances, .var = "dxs", lambda = 1.6)

}
\seealso{
\code{\link{estimateInterProfileCorrelation}}
}
\author{
Thomas Münch
}
