% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateStackCorrelation.R
\name{estimateStackCorrelation}
\alias{estimateStackCorrelation}
\title{Profile stack correlation statistics}
\usage{
estimateStackCorrelation(
  data,
  reference,
  distances,
  nprofiles,
  profilePosition = NULL,
  .var = "d18O",
  rangeTol = 0.05,
  na.rm = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{either a generic trench data set, or a trench data set in form
of a numeric matrix or data frame, where columns represent the horizontal
trench profile positions and rows represent the vertical scale of the
data.}

\item{reference}{numeric vector with a reference signal to which the profile
sets are correlated; its length must match the length of the profiles in
\code{data}.}

\item{distances}{a vector of inter-profile distances for which profile sets
shall be seeked and analysed for their correlation with the
\code{reference} profile; must be measured in the same units as the profile
positions of the trench data set.}

\item{nprofiles}{vector of desired number of profiles in each set.}

\item{profilePosition}{optional vector of the horizontal trench profile
positions; only needed when \code{data} is passed as a simple matrix or
data frame.}

\item{.var}{character string with the name of the trench variable for which to
compute the correlations; see also \code{\link{make2D}}. Only needed when
the \code{data} is passed as a generic trench data set.}

\item{rangeTol}{relative tolerance allowed for profile spacings; i.e. for a
given spacing d, all profile pairs separated by d * (1 +/- rangeTol) are
analysed. You can specify one global tolerance that is applied to every
distance in \code{distances}, or one tolerance for each distance.}

\item{na.rm}{a logical evaluating to \code{TRUE} or \code{FALSE} indicating
whether \code{NA} values should be stripped before the computation of the
mean stack profiles proceeds.}

\item{verbose}{logical; if \code{TRUE} print progress messages.}
}
\value{
a data frame with six columns: the requested inter-profile distance,
  the number (`N`) of individual profiles in a stack, the average correlation
  across all found stacks (consisting of `N` profiles at the given
  inter-profile distance) to the reference, the number of found stacks, and
  the actual minimum and maximum inter-profile distances across the found
  stacks (a sanity check, since this range should lie within the specified
  tolerance according to `distance * (1 +/- rangeTol)`).
}
\description{
This function determines with a Monte Carlo search algorithm all possible
unique stacks in a trench which consist of a specified number of individual
trench profiles at a required spacing. It then calculates the mean profile
for every stack and the correlation of this mean profile with a given
reference signal, and returns the average correlation for each combination
of distance and number of records.
}
\details{
This function can be used to empirically evaluate Eqs. (A14) and (A15) in
Münch et al. (2016).
}
\examples{
# How well do trench subsets represent the overall trench mean profile?
reference <- makeMean(t13.trench1, df = FALSE)
representativity <- estimateStackCorrelation(t13.trench1, reference,
                                             distances = c(2, 10),
                                             nprofiles = 5, rangeTol = 0.2,
                                             verbose = FALSE)
representativity
}
\references{
Münch, T., et al., Regional climate signal vs. local noise: a two-dimensional
view of water isotopes in Antarctic firn at Kohnen Station, Dronning Maud
Land, Clim. Past, 12(7), 1565-1581, doi:10.5194/cp-12-1565-2016, 2016.
}
\author{
Thomas Münch
}
