% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTrenchDecorrelation.R
\name{estimateTrenchDecorrelation}
\alias{estimateTrenchDecorrelation}
\title{Horizontal or vertical decorrelation length of trench dataset}
\usage{
estimateTrenchDecorrelation(
  data,
  direction = c("horizontal", "vertical"),
  .var = "d18O",
  vscale = "depth"
)
}
\arguments{
\item{data}{a trench data set following the default structure used in the
package.}

\item{direction}{character string signalling for which trench direction to
compute the decorrelation length: "horizontal" (default) or "vertical".}

\item{.var}{character string with the name of the trench variable for which to
compute the decorrelation length; see also \code{\link{make2D}}.}

\item{vscale}{the name of the vertical scale as a character string; defaults
to \code{"depth"}.}
}
\value{
a named length-1 vector with the estimated decorrelation length in
  the requested direction, measured in the same units as the respective
  sampling position.
}
\description{
Calculate the decorrelation length in horizontal or vertical direction of a
trench dataset. For this, the lag-1 autocorrelation across each vertical or
horizontal trench sampling position is estimated, averaged, and the
decorrelation length ("lambda") is calculated assuming an AR1 autoregressive
process:\cr
\cr
lambda = -1 * delta / log(a1),\cr
\cr
where a1 is the estimated average autocorrelation and delta the sampling
resolution.
}
\details{
The underlying autocorrelation estimation method works both for equidistantly
and non-equidistantly sampled data.

In the equidistant case, the method uses the base R \code{\link[stats]{acf}}
function, which ignores leading or trailing NA values but cannot handle
internal NAs. For that reason, any rows, or columns, of the trench dataset
(as viewed in 2D matrix form) that contain internal NA values trigger a
warning (see examples) and are removed from the data before estimation.

For non-equidistantly sampled data, the method utilizes the Gaussian kernel
estimation technique of Rehfeld et al. (2011) to account for the irregular
sampling. Here, NA values are generally allowed, but if more than 1/3 of the
data points in a row, or column, are NA, the corresponding autocorrelation is
set to NA with a warning (see examples). These NA values from problematic
rows or columns do not influence the decorrelation length, since any NA
values are removed from the individual autocorrelation estimates upon
averaging, unless the entire trench dataset contains too many NAs.
}
\examples{

# data sampled irregulary in horizontal, regularly in vertical direction
estimateTrenchDecorrelation(t13.trench1)
estimateTrenchDecorrelation(t13.trench1, direction = "vertical")
  # <- the warnings originate from removing trench rows and columns which
  # contain any NA, or too many NA values, depending on estimation method

# data sampled regularly in both directions
estimateTrenchDecorrelation(t15.trench2, "horizontal")
estimateTrenchDecorrelation(t15.trench2, "vertical")
estimateTrenchDecorrelation(t15.trench2, .var = "dxs")

}
\references{
Rehfeld, K., Marwan, N., Heitzig, J., and Kurths, J.: Comparison of
  correlation analysis techniques for irregularly sampled time series,
  Nonlinear Proc. Geoph., 18(3), 389–404, doi:
  https://doi.org/10.5194/npg18-389-2011, 2011
}
\seealso{
\code{\link[stats]{acf}}, \code{\link{make2D}}
}
\author{
Thomas Münch
}
