% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTrenchVariance.R
\name{estimateTrenchVariance}
\alias{estimateTrenchVariance}
\title{Horizontal and vertical variance of trench dataset}
\usage{
estimateTrenchVariance(trench, dof.h = NULL, dof.v = NULL, alpha = 1/6)
}
\arguments{
\item{trench}{a trench data set in form of a matrix or a data frame, where
columns represent the horizontal trench profile positions and rows
represent the vertical scale of the data; see also \code{\link{make2D}}.}

\item{dof.h}{number of effective degrees of freedom in horizontal direction.}

\item{dof.v}{number of effective degrees of freedom in vertical direction.}

\item{alpha}{probability for the confidence interval calculation;
i.e. confidence intervals are calculated for confidence levels from
\code{alpha} to \code{1 - alpha}.}
}
\value{
a tibble of four variables: the \code{direction} of the variance
  calculation as a character string ("horizontal" and "vertical"), the
  estimated variance \code{var}, and the \code{lower} and \code{upper} end of
  the confidence interval.
}
\description{
Calculate the average variance in horizontal and vertical direction of a
trench dataset. For this, the variances for each row and column,
respectively, of the trench dataset are estimated and then averaged. Missing
observations are silently removed from the data before calculation.
}
\examples{

make2D(t15.trench1) \%>\%
  estimateTrenchVariance()
# some example DOFs to see the effect (no realistic values!):
make2D(t15.trench1) \%>\%
  estimateTrenchVariance(dof.h = 5, dof.v = 30)

}
\author{
Thomas Münch
}
