% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{external_resistance_to_water_vapor_transfer}
\alias{external_resistance_to_water_vapor_transfer}
\title{Calculate external resistance to water vapor transfer}
\usage{
external_resistance_to_water_vapor_transfer(H, rhocp = 12000)
}
\arguments{
\item{H}{heat transfer (convection) coefficient (W m^-2 C^-1)}

\item{rhocp}{aggregate parameter (J m^-3 C^-1) that is the product of the density of air (kg m^-3)and the specific heat of air at constant pressure (J kg^-1 C^-1). Default of 12000 J m^-3 C^-1 is commonly assumed.}
}
\value{
external resistance to water vapor transfer (s m^-1)
}
\description{
Calculate external resistance to water vapor transfer
}
\details{
This function allows you to estimate external resistance to water vapor transfer using the Lewis rule relating heat and mass transport (Spotila et al. 1992).
}
\examples{
\dontrun{
external_resistance_to_water_vapor_transfer(H=20)
}

}
\keyword{external}
\keyword{resistance}
\keyword{to}
\keyword{transfer}
\keyword{vapor}
\keyword{water}
