% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{Converts Fahrenheit to Celsius}
\usage{
fahrenheit_to_celsius(T)
}
\arguments{
\item{T}{Temperature in Fahrenheit.}
}
\value{
Temperature in Celsius
}
\description{
This function allows you to convert temperature from Fahrenheit to Celsius. Source: https://swcarpentry.github.io.
}
\details{
Converts Fahrenheit to Celsius.
}
\examples{
\dontrun{
fahrenheit_to_celsius(85)
}
}
\keyword{Celsius}
\keyword{Fahrenheit}
