% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemperatureUtil.R
\name{fahrenheit_to_kelvin}
\alias{fahrenheit_to_kelvin}
\title{Converts Fahrenheit to Kelvin}
\usage{
fahrenheit_to_kelvin(temp)
}
\arguments{
\item{temp}{Temperature in Fahrenheit.}
}
\description{
(credit  https://swcarpentry.github.io)
}
\details{
This function allows you to convert temperature from Fahrenheit to Kelvin.
}
\examples{
fahrenheit_to_kelvin(85)

}
\keyword{Fahrenheit}
\keyword{Kelvin}
