% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{Free_or_forced_convection}
\alias{Free_or_forced_convection}
\title{Commpare Grashof and Reyolds numbers to determine whether convection is free or forced (Gates 1980)}
\usage{
Free_or_forced_convection(Gr, Re)
}
\arguments{
\item{Gr}{is the Grashof Number (dimensionless)}

\item{Re}{is the Reynolds Number (dimensionless)}
}
\value{
"free" or "forced"
}
\description{
Commpare Grashof and Reyolds numbers to determine whether convection is free or forced (Gates 1980)
}
\details{
This function allows you to commpare the Grashof and Reyolds numbers to determine whether convection is free or forced (Gates 1980).
}
\examples{
\dontrun{
Free_or_forced_convection(Gr=100, Re=5)
}

}
\keyword{convection}
\keyword{forced}
\keyword{free}
\keyword{or}
