% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeanProfileCorrelation.R
\name{getMeanProfileCorrelation}
\alias{getMeanProfileCorrelation}
\title{Correlation between mean trench profiles}
\usage{
getMeanProfileCorrelation(
  t1,
  t2,
  .var = "d18O",
  vscale = "depth",
  na.rm = FALSE,
  res = NULL,
  lag = NULL
)
}
\arguments{
\item{t1}{a generic trench dataset.}

\item{t2}{a second generic trench dataset.}

\item{.var}{character string with the name of the trench variable for which
to compute the mean profile correlation; see also \code{\link{make2D}}.}

\item{vscale}{the name of the vertical scale as a character string; must be
the same for both datasets and both scales must be equal; defaults to
\code{"depth"}.}

\item{na.rm}{a logical evaluating to \code{TRUE} or \code{FALSE} indicating
whether \code{NA} values should be stripped before the computation of the
mean profiles proceeds.}

\item{res}{resolution of the vertical trench scales (\code{vscale}); here, a
different (higher) than the inherent resolution can be specified onto which
the mean profiles are interpolated before computing their correlation.
Default \code{NULL} means to use the inherent resolution of the trench
datasets.}

\item{lag}{a vector of lags, i.e. positive and negative indices by which to
shift the mean profiles relative to each other in order to find their
maximum correlation. Default \code{NULL} means to apply no shifts.}
}
\value{
a tibble with one row, including per default the correlation and the
  root-mean-square deviation (`rmsd`) between the mean profiles; if a vector
  of \code{lags} has been specified, it includes the correlation and `rmsd`
  for the optimal shift (maximum correlation) and additionally the bin index
  and respective value of the vertical scale for which this optimal
  correlation is observed.
}
\description{
Calculate the mean profiles for two trench datasets and obtain the
correlation between them, optionally allowing for relative shifts between the
profiles and/or for interpolation onto a higher resolution.
}
\examples{

getMeanProfileCorrelation(t15.trench1, t15.trench2)
getMeanProfileCorrelation(t15.trench1, t15.trench2, .var = "dD")

getMeanProfileCorrelation(t15.trench1, t15.trench2,
                          lag = -4 : 4)
getMeanProfileCorrelation(t15.trench1, t15.trench2,
                          res = 0.5, lag = -24 : 24)

}
\author{
Thomas Münch
}
