% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-handling.R
\name{getSurfaceProfile}
\alias{getSurfaceProfile}
\title{Trench surface profile}
\usage{
getSurfaceProfile(data)
}
\arguments{
\item{data}{a trench data set following the default structure used in the
package.}
}
\value{
a tibble of two variables: the horizontal profile position in the
  first column and the surface heights at each position in the second column.
}
\description{
Easily retrieve the profile positions and the corresponding surface heights
for a supplied trench data set.
}
\examples{
getSurfaceProfile(t13.trench2)
}
\author{
Thomas Münch
}
