% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-handling.R
\name{getX}
\alias{getX}
\title{Horizontal trench profile positions}
\usage{
getX(data)
}
\arguments{
\item{data}{a tibble with a trench data set with horizontal profile positions
listed in the column \code{profilePosition}.}
}
\value{
numeric vector of the profile positions.
}
\description{
Easily retrieve the horizontal profile positions of a trench data set.
}
\examples{
getX(t13.trench1)
}
\author{
Thomas Münch
}
