% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-handling.R
\name{getZ}
\alias{getZ}
\title{Common vertical trench scale}
\usage{
getZ(data, var = "depth")
}
\arguments{
\item{data}{a trench data set following the default structure used in the
package.}

\item{var}{the name of the vertical scale as a character string; defaults to
\code{"depth"}.}
}
\value{
numeric vector of the vertical scale.
}
\description{
Easily retrieve the absolute vertical scale that is common to all profiles of
the supplied trench data set. Most commonly this is a depth scale.
}
\examples{
getZ(t13.trench1)
}
\author{
Thomas Münch
}
