% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{Grashof_number_Gates}
\alias{Grashof_number_Gates}
\title{Calculate Grashof Number in Gates}
\usage{
Grashof_number_Gates(Ta, Tg, beta, D, nu)
}
\arguments{
\item{Ta}{Air temperature (C).}

\item{Tg}{Ground (surface) temperature (C).}

\item{beta}{coefficient of volumetric thermal expansion, beta= 3.67 x 10^-3 C^-1  in air and 41.9 x 10^-4 C^-1 in water.}

\item{D}{is characteristic dimension (e.g., body diameter) (m)}

\item{nu}{is the kinematic viscosity, ratio of dynamic viscosity to density of the fluid (m2 s-1), can calculate from DRYAIR or WETAIR}
}
\value{
Grashof number
}
\description{
Calculate Grashof Number in Gates
}
\details{
This function allows you to estimate the Grashof Number, which describes the abilty of a parcel of fluid warmer or colder than the surrounding fluid to rise against or fall with the attractive force of gravity (Gates 1980 Biophysical Ecology). Ratio of a buoyant force times an inertial force to the square of a viscous force.
}
\examples{
\dontrun{
Grashof_number_Gates(Ta=30, Tg=35, beta=0.00367, D=0.001, nu=1.2)
}

}
\keyword{Grashof}
\keyword{number}
