% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{heat_transfer_coefficient_lizard}
\alias{heat_transfer_coefficient_lizard}
\title{Calculate heat transfer coefficient for lizard}
\usage{
heat_transfer_coefficient_lizard(V, orientation = "parallel")
}
\arguments{
\item{V}{Air velocity m/s.}

\item{orientation}{parallel or transverse}
}
\value{
heat transfer coefficient, H (W m^-2 K^-1)
}
\description{
Calculate heat transfer coefficient for lizard
}
\details{
This function allows you estimate the heat transfer coefficient for a lizard (Based on Porter et al. 1973)
}
\examples{
\dontrun{
heat_transfer_coefficient_lizard(V=3, orientation="parallel")
}

}
\keyword{coefficient}
\keyword{heat}
\keyword{transfer}
