% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemperatureUtil.R
\name{kelvin_to_celsius}
\alias{kelvin_to_celsius}
\title{Converts Kelvin to Celsius}
\usage{
kelvin_to_celsius(temp)
}
\arguments{
\item{temp}{Temperature in Fahrenheit.}
}
\description{
(credit  https://swcarpentry.github.io)
}
\details{
This function allows you to convert temperature from Kelvin to Celsius.
}
\examples{
kelvin_to_celsius(270)

}
\keyword{Celsius}
\keyword{Kelvin}
