% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoopParamSpace.R
\name{LoopParamSpace}
\alias{LoopParamSpace}
\title{Loop over parameter space to find minimum RMSD.}
\usage{
LoopParamSpace(TR = prepareTrenchData()$oxy, advSpace = seq(40, 60,
  0.5)/TR$HiRes, sigmaSpace = seq(0, 8, 0.1), densfSpace = seq(0, 10,
  0.1))
}
\arguments{
\item{TR}{tbd}

\item{advSpace}{tbd}

\item{sigmaSpace}{tbd}

\item{densfSpace}{tbd}
}
\value{
A list with nine components:
\describe{
  \item{adv:}{a copy of \code{advSpace};}
  \item{sigma:}{a copy of \code{sigmaSpace};}
  \item{densf:}{a copy of \code{densfSpace};}
  \item{adv.opt:}{the optimal advection value;}
  \item{sigma.opt:}{the optimal diffusion length;}
  \item{densf.opt:}{the optimal compression value;}
  \item{adv.opt.arr:}{an array of dimension \code{length(sigma)} x
    \code{length(densf)} which contains the optimal advection values at
    which, for fixed compression and diffusion, the RMSD between the
    reference and the modified input record is minimal;}
  \item{RMSD:}{an array of dimension \code{length(adv)} x
    \code{length(sigma)} x \code{length(densf)} which contains the RMSD value
    between the reference and the modified input record for every
    combination of downward advection, diffusion and compression;}
  \item{RMSD.opt:}{the array \code{RMSD} projected onto the optimal
    downward-advection value.}
}
}
\description{
Loop over a given parameter space of advection, diffusion and densification
values to modify a given record accordingly, and return the overall minimum
root mean square deviation (RMSD) to a reference record.
}
\details{
Note that for computational efficiency, the implemented order of
modifications is (1) downward advection, (2) diffusion, and (3) compression
from densification. This is slightly unphysical since the diffusional
smoothing acts over the uncompressed depth scale. However, it affects the
results only slightly in the domain of high diffusion lengths and high
compression values, and does not affect the main results discussed in Münch
et al. (2017).
}
\references{
Münch, T., et al., Constraints on post-depositional isotope modifications
in East Antarctic firn from analysing temporal changes of isotope profiles,
The Cryosphere, doi:10.5194/tc-11-2175-2017, 2017.
}
\seealso{
\code{\link{DiffuseRecord}}; \code{\link{CompressRecord}};
\code{?ParamSpace}
}
\author{
Thomas Münch
}
