% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-handling.R
\name{make2D}
\alias{make2D}
\title{2D trench record}
\usage{
make2D(data, .var = "d18O", simplify = FALSE)
}
\arguments{
\item{data}{a trench data set following the default structure used in the
package.}

\item{.var}{character string with the name of the requested variable; must
match one of the data column variables in \code{data}, or be one of the
following:
\describe{
\item{\code{"dxs"}:}{this will return d-excess data as calculated from
  \code{"dD - 8 * d18O"} given in the trench data, but only if a column
\code{"dxs"} is not already present.}
\item{\code{"nssSulfate"}:}{this will return non-sea-salt sulfate data as
  calculated from \code{"Sulfate - 0.252 * Sodium"} given in the trench
  data, but only if a column \code{"nssSulfate"} is not already present.}
}}

\item{simplify}{logical; for \code{FALSE} (the default) the result is
returned as a tibble, else it is simplified to a matrix.}
}
\value{
a tibble or matrix of the requested trench data, with rows indexing
  the vertical scale (most commonly depth) and columns the horizontal profile
  position.
}
\description{
Obtain a two-dimensional (vertical x horizontal) representation of a certain
variable for a given trench data set.
}
\examples{

# Per default, oxygen isotope data is returned:
make2D(t13.trench2)

# Request another variable:
make2D(t13.trench2, .var = "dD")
# Obtain the same result as a simple matrix:
make2D(t13.trench2, .var = "dD", simplify = TRUE)

}
\author{
Thomas Münch
}
