% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-handling.R
\name{makeMean}
\alias{makeMean}
\title{Trench mean profile}
\usage{
makeMean(data, var = "d18O", vscale = "depth", na.rm = FALSE, df = TRUE)
}
\arguments{
\item{data}{a trench data set following the default structure used in the
package.}

\item{var}{character string with the name of the requested variable; must
match one of the data column variables in \code{data}, or be one of the
following:
\describe{
\item{\code{"dxs"}:}{this will return d-excess data as calculated from
  \code{"dD - 8 * d18O"} given in the trench data, but only if a column
\code{"dxs"} is not already present.}
\item{\code{"nssSulfate"}:}{this will return non-sea-salt sulfate data as
  calculated from \code{"Sulfate - 0.252 * Sodium"} given in the trench
  data, but only if a column \code{"nssSulfate"} is not already present.}
}}

\item{vscale}{a character string giving the name of the vertical scale
variable used in the trench data; defaults to \code{"depth"}; see also
\code{\link{getZ}}.}

\item{na.rm}{a logical evaluating to \code{TRUE} or \code{FALSE} indicating
whether \code{NA} values should be stripped before the computation of the
mean profile proceeds.}

\item{df}{a logical; for \code{TRUE} (the default) the mean profile is
returned as a tibble where the first column includes the vertical scale
of the trench data given in the variable \code{vscale}, else the mean
profile is returned as a simple numeric vector.}
}
\value{
a tibble or numeric vector.
}
\description{
Calculate the mean profile for a certain trench variable by averaging across
the horizontal dimension of the trench data.
}
\examples{

# incomplete surface region is stripped
makeMean(t13.trench2)
# mean values are also calculated for incomplete surface region
makeMean(t13.trench2, na.rm = TRUE)

# get dD data as a simple vector
makeMean(t13.trench2, var = "dD", df = FALSE)

}
\author{
Thomas Münch
}
