% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTC17Figures.R
\name{MakeTC17Figures}
\alias{MakeTC17Figures}
\title{Produce TC17 Figures.}
\usage{
MakeTC17Figures(path = file.path(getwd(), "plots"), device = "quartz",
  save.plot = FALSE)
}
\arguments{
\item{path}{}

\item{device}{}

\item{save.plot}{}
}
\description{
Details
}
\author{
Thomas Münch
}
