% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllometricFunctions.R
\name{mass_from_length}
\alias{mass_from_length}
\title{Calculate mass from length}
\usage{
mass_from_length(l, taxa)
}
\arguments{
\item{l}{\code{numeric} length in meters (m). 
\cr \cr
Snout-vent length for amphibians and reptiles (excepting turtles where length is carapace length).}

\item{taxa}{\code{character} taxon of organism, current choices: \code{"insect"}, \code{"lizard"}, \code{"salamander"}, \code{"frog"}, \code{"snake"}, \code{"turtle"}.}
}
\value{
\code{numeric} mass in grams (g)
}
\description{
Estimate mass (g) from length (m) for a variety of taxa.
}
\details{
Relationships come from 
 \itemize{
   \item insect: \insertCite{Sample1993}{TrenchR}.
   \item lizard: \insertCite{Meiri2010}{TrenchR}.
   \item salamander: \insertCite{Pough1980}{TrenchR}.
   \item frog: \insertCite{Pough1980}{TrenchR}.
   \item snake: \insertCite{Pough1980}{TrenchR}.
   \item turtle: \insertCite{Pough1980}{TrenchR}.
 }
}
\examples{
  mass_from_length(l = 0.04, taxa = "insect")
  mass_from_length(l = 0.04, taxa = "lizard")
  mass_from_length(l = 0.04, taxa = "salamander")
  mass_from_length(l = 0.04, taxa = "frog")
  mass_from_length(l = 0.04, taxa = "snake")
  mass_from_length(l = 0.04, taxa = "turtle")

}
\references{
\insertAllCited{}
}
\seealso{
Other allometric functions: 
\code{\link{prop_silhouette_area_shapes}()},
\code{\link{prop_silhouette_area}()},
\code{\link{sa_from_length}()},
\code{\link{sa_from_mass}()},
\code{\link{sa_from_volume}()},
\code{\link{volume_from_length}()}
}
\concept{allometric functions}
\keyword{length}
\keyword{mass}
