% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{metabolic}
\alias{metabolic}
\title{Calculate metabolic expenditure}
\usage{
metabolic(Ta, mass, species = "lizard")
}
\arguments{
\item{Ta}{Air Temperature.}

\item{mass}{Mass in grams.}

\item{species}{Which species.}
}
\value{
metabolic expense
}
\description{
Calculate metabolic expenditure
}
\details{
This function allows you to calculate metabolic expense of an ectotherm(Reptiles)
}
\examples{
\dontrun{
metabolic(24,10.5,"lizard")
}

}
\keyword{metabolism}
