% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{metabolic_rate}
\alias{metabolic_rate}
\title{Calculate metabolic expenditure}
\usage{
metabolic_rate(Tb = NA, mass, taxa = "lizard")
}
\arguments{
\item{Tb}{body temperature in degrees C, currently only used for lizard.}

\item{mass}{Mass in grams.}

\item{taxa}{Which taxa. options: lizard, reptile, bird, mammal}
}
\value{
metabolic expense
}
\description{
Caters to heat generated because of metabolism.
}
\details{
This function allows you to calculate basal or field metabolic rate (j/s) of various taxa (BMR or FMR depends on taxa)
}
\examples{
\dontrun{
metabolic_rate(24,10.5,"lizard")
}

}
\keyword{metabolism}
