% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{mr_from_mass}
\alias{mr_from_mass}
\title{Calculate metabolism as a function of mass}
\usage{
mr_from_mass(mass, taxa = "reptile")
}
\arguments{
\item{mass}{Mass in grams.}

\item{taxa}{Which taxa. Current options: reptile, bird, mammal}
}
\value{
metabolim (W)
}
\description{
Calculate metabolism as a function of mass
}
\details{
This function allows you to estimate field metabolic rate (W) of various taxa as a function of mass(g). Does not account for temperature. Uses empirical relationships from Nagy et al. 1999.
}
\examples{
\dontrun{
mr_from_mass(10.5,"reptile")
}

}
\keyword{metabolism}
