% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{mr_mass_temperature}
\alias{mr_mass_temperature}
\title{Calculate basal (or resting) metabolism as a function of mass and body temperature.}
\usage{
mr_mass_temperature(mass, Tb, taxa)
}
\arguments{
\item{mass}{Mass in grams.}

\item{Tb}{body temperature in K}

\item{taxa}{Which taxa. options: bird, mammal, reptile, amphibian, invertebrate.}
}
\value{
metabolim (W)
}
\description{
Calculate basal (or resting) metabolism as a function of mass and body temperature.
}
\details{
This function allows you to estimate basal (or resting) metabolic rate (W) as a function of mass (g) and temperature (K). Based on empirical data and the metabolic theory of ecology (3/4 scaling exponent). From Gilooly et al. 2001.
}
\examples{
\dontrun{
mr_mass_temperature(mass=100, Tb=303,"reptile")
}

}
\keyword{metabolism}
