% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docu.R
\name{Muench2016}
\alias{Muench2016}
\title{Citation for Münch et al. (2016)}
\description{
Citation for Münch et al. (2016)
}
\references{
Münch, T., et al., Regional climate signal vs. local noise: a two-dimensional
view of water isotopes in Antarctic firn at Kohnen Station, Dronning Maud
Land, Clim. Past, 12(7), 1565-1581, doi:10.5194/cp-12-1565-2016, 2016.
}
