% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docu.R
\name{Muench2017}
\alias{Muench2017}
\title{Citation for Münch et al. (2017)}
\description{
Citation for Münch et al. (2017)
}
\references{
Münch, T., et al., Constraints on post-depositional isotope modifications
in East Antarctic firn from analysing temporal changes of isotope profiles,
The Cryosphere, 11(5), 2175-2188, doi:10.5194/tc-11-2175-2017, 2017.
}
