% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{Nu_from_Gr}
\alias{Nu_from_Gr}
\title{Estimate the Nusselt number from the Grashof number (based on Gates 1980)}
\usage{
Nu_from_Gr(Gr)
}
\arguments{
\item{Gr}{is the Grashof Number (dimensionless)}
}
\value{
Nusselt number (dimensionless)
}
\description{
Estimate the Nusselt number from the Grashof number (based on Gates 1980)
}
\details{
This function allows you to estimate the Nusselt number from the Grashof Number.
}
\examples{
\dontrun{
Nu_from_Gr(Gr=5)
}

}
\keyword{Nusselt}
\keyword{number}
