% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{Nu_from_Re}
\alias{Nu_from_Re}
\title{Estimate the Nusselt number from the Reynolds number (based on Mitchell 1976)}
\usage{
Nu_from_Re(Re, taxa = "cylinder")
}
\arguments{
\item{Re}{is the Reynolds Number (dimensionless)}

\item{taxa}{Which class of organism, current choices: sphere, cylinder, frog, lizard_traverse_to_air_flow, lizard_parallel_to_air_flow, lizard_surface, lizard_elevated, flyinginsect, spider}
}
\value{
Nusselt number (dimensionless)
}
\description{
Estimate the Nusselt number from the Reynolds number (based on Mitchell 1976)
}
\details{
This function allows you to estimate the Nusselt number from the Reynolds number for various taxa.  Source: Mitchell JW. 1976. Heat transfer from spheres and other animal forms. Biophysical Journal 16: 561-569. See Table 1. Convective Heat Trasfer Relations for Animal Shapes.
}
\examples{
\dontrun{
Nu_from_Re(Re=5, taxa="cylinder")
}

}
\keyword{Nusselt}
\keyword{number}
