% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{Nusselt_number}
\alias{Nusselt_number}
\title{Calculate Nusselt Number}
\usage{
Nusselt_number(H_L, D, K)
}
\arguments{
\item{H_L}{Convective heat transfer coefficient (W m^-2 K^-1)}

\item{D}{is characteristic dimension (e.g., body diameter) (m)}

\item{K}{Thermal conductivity (W K^-1 m^-1 )}
}
\value{
Nusselt number
}
\description{
Calculate Nusselt Number
}
\details{
This function allows you to estimate the Nusselt Number, which describes dimensionless conductance (Gates 1980 Biophysical Ecology).
}
\examples{
\dontrun{
Nusselt_number(H_L=20, D=0.01, K=0.5)
}

}
\keyword{Nusselt}
\keyword{number}
