% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docu.R
\docType{data}
\name{ParamSpace}
\alias{ParamSpace}
\title{Data for TC2017 Figure 5.}
\format{A list with nine components:
\describe{
  \item{adv:}{A vector with the range of advection values for which the RMSD
    is computed.}
  \item{sigma:}{A vector with the range of differential diffusion lengths
    for which the RMSD is computed.}
  \item{densf:}{A vector with the range of compression values for which the
    RMSD is computed.}
  \item{adv.opt:}{The optimal advection value.}
  \item{sigma.opt:}{The optimal differential diffusion length.}
  \item{densf.opt:}{The optimal compression value.}
  \item{adv.opt.arr:}{An array of dimension \code{length(sigma)} x
    \code{length(densf)} which contains the optimal advection values at
    which, for fixed compression and differential diffusion, the RMSD between
    the mean T15 and the modified mean T13 profile is minimal.}
  \item{RMSD:}{An array of dimension \code{length(adv)} x
    \code{length(sigma)} x \code{length(densf)} which contains the RMSD value
    between T15 and the modified T13 mean profile for every combination of
    downward-advection, differential diffusion and compression.}
  \item{RMSD.opt:}{The array \code{RMSD} projected onto the optimal
    downward-advection value.}
}}
\source{
A call to \code{LoopParamSpace}.
}
\usage{
ParamSpace
}
\description{
This file contains the data displayed in Figure 5 of the TC17 paper and is
the result of a call to \code{\link{LoopParamSpace}} for the parameter ranges
of downward-advection, differential diffusion and compression from
densification as described in the paper.
}
\references{
Muench, T., et al., Constraints on post-depositional isotope modifications in
    East Antarctic firn from analysing temporal changes of isotope profiles,
    The Cryosphere Discussions, doi:10.5194/tc-2017-35, 2017.
}
\seealso{
\code{\link{LoopParamSpace}}
}
\keyword{datasets}
