% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationPartitioningFunctions.R
\name{partition_solar_radiation}
\alias{partition_solar_radiation}
\title{Partition solar radiation (W m^-2) into direct and diffuse components
IMPLEMENT OTHER MODELS IN Wong and Chow (2001, Applied Energy 69:1991-224)?}
\usage{
partition_solar_radiation(rad, kt)
}
\arguments{
\item{rad}{is solar radiation in W m^-2}

\item{kt}{is the clearnex index (dimensionless), which is the ratio of the global solar radiation measured at the surface to the total solar radiation at the top of the atmosphere.}
}
\description{
Partition solar radiation (W m^-2) into direct and diffuse components
IMPLEMENT OTHER MODELS IN Wong and Chow (2001, Applied Energy 69:1991-224)?
}
\details{
This function allows you to partition solar radiation (W m^-2) into direct and diffuse components using the Erbs et al. model presented in Wong and Chow (2001, Applied Energy 69:1991-224). Correction for CO from Olyphant 1984. Returns the direct and diffuse components of radiation in W m^-2.
}
\examples{
\dontrun{
partition_solar_radiation(500, 0.5)
}
}
\keyword{radiation}
\keyword{solar}
