% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polyplot.R
\name{Polyplot}
\alias{Polyplot}
\title{Wrapper for polygon plot.}
\usage{
Polyplot(x, y2, col = "grey", ...)
}
\arguments{
\item{x}{x}

\item{y2}{y2}

\item{col}{colour.}

\item{...}{Further graphical parameters passed to the \code{polygon}
function.}
}
\description{
Description.
}
\author{
Thomas Münch
}
