% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{Prandtl_number}
\alias{Prandtl_number}
\title{Calculate Prandtl Number}
\usage{
Prandtl_number(c_p, mu, K)
}
\arguments{
\item{c_p}{is specific heat at constant pressure (J mol^{-1} K^{-1})}

\item{mu}{is dynamic viscosity (mol s^{-1}m^{-1})}

\item{K}{Thermal conductivity (W K^-1 m^-1 )}
}
\value{
Prandtl number
}
\description{
Calculate Prandtl Number
}
\details{
This function allows you to estimate the Prandtl Number, which describes the ratio of kinematic viscosity to thermal diffusivity (Gates 1980 Biophysical Ecology).
}
\examples{
\dontrun{
Prandtl_number(c_p=29.3, mu=0.00001, K=0.5)
}

}
\keyword{Prandtl}
\keyword{number}
