% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareTrenchData.R
\name{prepareTrenchData}
\alias{prepareTrenchData}
\title{Process T13 and T15 Trench Data.}
\usage{
prepareTrenchData(index.range = 1:59, LoRes = 3, HiRes = 0.5, k13 = 3,
  k15 = -0.5, na.treat = FALSE)
}
\arguments{
\item{index.range}{Vector of indices to subset the T15 trench dataset;
defaults to the range used in the TC paper.}

\item{LoRes}{Depth resolution of trench dataset in cm; defaults to the
original T13 and T15 sampling resolution (3 cm).}

\item{HiRes}{Higher depth resolution to interpolate trench data onto;
defaults to 0.5 cm as used in the TC paper.}

\item{k13}{Vertical shift [cm] of the mean T13-2 relative to the mean T13-1
profile; defaults to the optimal shift that maximizes the profile
correlation.}

\item{k15}{Vertical shift [cm] of the mean T15-2 relative to the mean T15-1
profile; defaults to the optimal shift that maximizes the profile
correlation.}

\item{na.treat}{if \code{TRUE} include the incomplete surface region of the
trenches when calculating the mean profiles; defaults to \code{FALSE}.}
}
\value{
A list with three elements corresponding to the three measured
    isotopologue data (index i = 1: d18O, i = 2: d2H, i = 3: d-excess). Each
    element again is a list with the following components:
    \itemize{
    \item trench13.1: Data of isotope species \code{i} for trench T13-1.
    \item trench13.2: Data of isotope species \code{i} for trench T13-2.
    \item trench15.1: Data of isotope species \code{i} for trench T15-1
          for the  depth range specified by \code{index.range}.
    \item trench15.2: Data of isotope species \code{i} for trench T15-2
          for the depth range specified by \code{index.range}.
    \item mean13.1: Mean isotope profile \code{i} of trench T13-1.
    \item mean13.2: Mean isotope profile \code{i} of trench T13-2.
    \item mean15.1: Mean isotope profile \code{i} of trench T15-1.
    \item mean15.2: Mean isotope profile \code{i} of trench T15-2.
    \item trench13.1_HiRes: \code{trench13.1} linearly interpolated onto the
          higher resolution specified by \code{HiRes}.
    \item trench13.2_HiRes: Same for \code{trench13.2}.
    \item trench15.1_HiRes: Same for \code{trench15.1}.
    \item trench15.2_HiRes: Same for \code{trench15.2}.
    \item mean13.1_HiRes: Same for \code{mean13.1}.
    \item mean13.2_HiRes: Same for \code{mean13.2}.
    \item mean15.1_HiRes: Same for \code{mean15.1}.
    \item mean15.2_HiRes: Same for \code{mean15.2}.
    \item mean13: Total mean isotope profile \code{i} of trench T13,
          accounting for the vertical shift specified by \code{k13}.
    \item mean13_HiRes: \code{mean13} linearly interpolated onto the
          higher resolution specified by \code{HiRes}.
    \item mean15: Total mean isotope profile \code{i} of trench T15.
    \item mean15_HiRes: \code{mean15} linearly interpolated onto the
          higher resolution specified by \code{HiRes} and accounting for the
          vertical shift specified by \code{k15}.
    \item k13: Copy of input \code{k13}.
    \item k15: Copy of input \code{k15}.
    \item LoRes: Copy of input \code{LoRes}.
    \item HiRes: Copy of input \code{HiRes}.
    \item depth: Numeric vector of sample depths on their original
          resolution (specified by \code{LoRes}) for the T15 index range
          \code{index.range}. 
    \item depth_HiRes: Numeric vector of sample depths corresponding to the
          linearly interpolated data on the higher resolution \code{HiRes}.
    \item XPOS: Horizontal profile positions in [m] of the T15-1 and T15-2
          profiles relative to the trench starting point (= profile #1).
    \item XPOS.t13.1: Horizontal profile positions in [m] of the T15-1
          profiles relative to the trench starting point.
    \item XPOS.t13.2: Horizontal profile positions in [m] of the T15-2
          profiles relative to the trench starting point.
    \item SPRF.t1: List: Horizontal position (component \code{x}) and
          corresponding surface height in [cm] (component \code{y})
          relative to the highest observed surface height for trench
          T15-1 (positive downwards). This set includes the additional
          profile at position 26.5 m for which no isotope profile is
          present.
    \item SPRF.t2: List: Horizontal position (component \code{x}) and
          corresponding surface height in [cm] (component \code{y})
          relative to the highest observed surface height for trench
          T15-2 (positive downwards).
    \item SRF.b: A list with six elements of the following quantities (in
          that order): starting depths of horizontally complete datasets for
          T13-1; T13-2; T15-1; T15-2; total T13 (accounting for vertical
          shift of T13-2); and total T15.
}
}
\description{
This function processes the T13 and T15 trench isotopologue data in a way
that is suited for analysis and plotting used in the TC paper.
}
\author{
Thomas Münch
}
