% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper-tc17-plots.R
\name{produceTC17Figures}
\alias{produceTC17Figures}
\title{M端nch et al. (2017) figures}
\usage{
produceTC17Figures(
  which.figure = c("f1", "f2", "f3a", "f3b", "f3c", "f4", "f5", "f6", "f7")
)
}
\arguments{
\item{which.figure}{character string to name the figure which shall be
reproduced; must match one of "f1", "f2", "f3a", "f3b", "f3c", "f4", "f5",
"f6", "f7".}
}
\description{
Reproduce a desired figure of the M端nch et al. (2017) publication (TC17).
}
\details{
Appropriate figure aspect ratios (height x width (h x w)) in inch for the
graphical output device are as follows:

\describe{
  \item{Fig. 1:}{h = 6.75, w = 16}
  \item{Fig. 2:}{h = 8, w = 8}
  \item{Fig. 3a, b:}{h = 6, w = 9.75}
  \item{Fig. 3c:}{h = 6, w = 8}
  \item{Fig. 4:}{h = 6, w = 8}
  \item{Fig. 5:}{h = 6, w = 8}
  \item{Fig. 6:}{h = 6, w = 16}
  \item{Fig. 7:}{h = 6, w = 12}
}
}
\references{
M端nch, T., et al., Constraints on post-depositional isotope modifications
in East Antarctic firn from analysing temporal changes of isotope profiles,
The Cryosphere, 11(5), 2175-2188, doi:10.5194/tc-11-2175-2017, 2017.
}
\author{
Thomas M端nch
}
