% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationPartitioningFunctions.R
\name{proportion_diffuse_solar_radiation}
\alias{proportion_diffuse_solar_radiation}
\title{Estimate the ratio of diffuse to direct solar radiation}
\usage{
proportion_diffuse_solar_radiation(psi, p_a, A)
}
\arguments{
\item{psi}{Zenith angle of the sun (degrees)}

\item{p_a}{Atmospheric pressure (kPa)}

\item{A}{albedo of the substrate (fraction of 1)}
}
\value{
diffuse fraction
}
\description{
This function estimates the ratio of diffuse to direct solar radiation.  Based on the approximation of the McCullough and Porter (1971) SOLRAD model described in Tracy et al. (1983) Estimating clear-day solar radiation: An evaluation of three models. Journal of Thermal Biology, 8(3), 247-251.
}
\details{
Estimate the ratio of diffuse to direct solar radiation
}
\examples{
\dontrun{
proportion_diffuse_solar_radiation(psi=60, p_a=86.1, A=0.25)
}
}
\seealso{
Other microclimate functions: 
\code{\link{air_temp_profile_neutral}()},
\code{\link{air_temp_profile_segment}()},
\code{\link{air_temp_profile}()},
\code{\link{degree_days}()},
\code{\link{direct_solar_radiation}()},
\code{\link{diurnal_radiation_variation}()},
\code{\link{diurnal_temp_variation_sineexp}()},
\code{\link{diurnal_temp_variation_sinesqrt}()},
\code{\link{diurnal_temp_variation_sine}()},
\code{\link{estimate_radiation}()},
\code{\link{monthly_solar_radiation}()},
\code{\link{partition_solar_radiation}()},
\code{\link{surface_roughness}()},
\code{\link{wind_speed_profile_neutral}()},
\code{\link{wind_speed_profile_segment}()}
}
\concept{microclimate functions}
\keyword{radiation}
\keyword{solar}
