% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{Qconvection}
\alias{Qconvection}
\title{Calculate convection}
\usage{
Qconvection(T_a, T_b, H = 10.45, A, proportion, ef = 1.3)
}
\arguments{
\item{T_a}{Air Temperature in Kelvin.}

\item{T_b}{Initial Body Temperature in Kelvin.}

\item{H}{Convective heat transfer coefficient (W m^-2 K^-1)}

\item{A}{Surface area  in m^2}

\item{proportion}{of surface area exposed to air}

\item{ef}{is the enhancement factor, used to adjust H to field condictions (using H approximation from Mitchell 1976).  Approximated as 1.3 by default, but see Mitchell 1976 for more relationship.}
}
\value{
convection (W)
}
\description{
Calculate convection
}
\details{
This function allows you to calculate convection as in Mitchell (1976). Includes an enhancement factor associated with outdoor environments.
}
\examples{
\dontrun{
Qconvection(T_a= 293,T_b= 303,H=10.45,A=0.0025, proportion=0.85)
}

}
\keyword{convection}
