% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{Qmetabolism_from_mass}
\alias{Qmetabolism_from_mass}
\title{Calculate metabolism as a function of mass}
\usage{
Qmetabolism_from_mass(m, taxa = "reptile")
}
\arguments{
\item{m}{Mass in grams.}

\item{taxa}{Which taxa. Current options: reptile, bird, mammal}
}
\value{
metabolism (W)
}
\description{
Calculate metabolism as a function of mass
}
\details{
This function allows you to estimate field metabolic rate (W) of various taxa as a function of mass(g). Does not account for temperature. Uses empirical relationships from Nagy KA. 2005. Field metabolic rate and body size. Journal of Experimental Biology 208: 1621-1625.
}
\examples{
\dontrun{
Qmetabolism_from_mass(m=12,"reptile")
}

}
\keyword{metabolism}
