% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{Qmetabolism_from_mass_temp}
\alias{Qmetabolism_from_mass_temp}
\title{Calculate basal (or resting) metabolism as a function of mass and body temperature.}
\usage{
Qmetabolism_from_mass_temp(m, T_b, taxa)
}
\arguments{
\item{m}{Mass in grams.}

\item{T_b}{body temperature in K}

\item{taxa}{Which taxa. options: bird, mammal, reptile, amphibian, invertebrate.}
}
\value{
metabolim (W)
}
\description{
Calculate basal (or resting) metabolism as a function of mass and body temperature.
}
\details{
This function allows you to estimate basal (or resting) metabolic rate (W) as a function of mass (g) and temperature (K). Based on empirical data and the metabolic theory of ecology (3/4 scaling exponent). From Gilooly et al. 2001.
}
\examples{
\dontrun{
Qmetabolism_from_mass_temp(m=100, T_b=303,"reptile")
}

}
\keyword{metabolism}
