% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperativeTemperatureFunctions.R
\name{Qnet_Gates}
\alias{Qnet_Gates}
\title{Estimates net energy exchange between an animal and the environment in W.}
\usage{
Qnet_Gates(Qabs, Qconv, Qcond, Qmet, Qevap)
}
\arguments{
\item{Qabs}{Solar and thermal radiation absorbed (W)}

\item{Qconv}{Energy exchange between an animal and the air (W)}

\item{Qcond}{Energy exchange between an animal and the surface (W)}

\item{Qmet}{Energy emitted due to metabolism (W)}

\item{Qevap}{Energy emitted due to evaporative water loss (W)}

\item{Qemit}{Thermal radiation emitted (W)}
}
\value{
net energy exchange (W)
}
\description{
Estimates net energy exchange between an animal and the environment in W.
}
\details{
Estimates net energy exchange between an animal and the environment in W. Follows Gates (1980, Biophysical ecology) and others.
}
\examples{
\dontrun{
Qnet_Gates(Qabs=500, Qconv=100, Qcond=100, Qmet=10, Qevap=5)
}

}
\keyword{environmental}
\keyword{operative}
\keyword{temperature}
