% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{Qradiation_absorbed}
\alias{Qradiation_absorbed}
\title{Calculate absorbed solar and thermal radiation}
\usage{
Qradiation_absorbed(a = 0.9, A, psa_dir, psa_ref, S_dir, S_dif,
  S_ref = NA, a_s = NA)
}
\arguments{
\item{a}{solar absorptivity of animal surface (proportion), default value is for reptiles}

\item{A}{surface area  in m^2}

\item{psa_dir}{proportion surface area exposed to solar radiation}

\item{psa_ref}{proportion surface area exposed to reflected solar radiation}

\item{S_dir}{direct solar radiation (W/m^2)}

\item{S_dif}{diffuse solar radiation (W/m^2)}

\item{S_ref}{reflected solar radiation (W/m^2), either provided or estimated if surface albedo is provided instead}

\item{a_s}{is surface albedo (proportion), optional (not used) if reflected radiation is provided, Values available in Gates Biophysical ecology Table 8.2.}
}
\value{
solar radiation absorbed (W)
}
\description{
Calculate absorbed solar and thermal radiation
}
\details{
This function allows you to estimate solar and thermal radiation (W) absorbed by the surface of an animal. 
         Follows Gates' Biophysical Ecology and Spotila et al. 1992. Reference: Spotila et al. 1992. Biophysics of Heat and Mass Transfer. In Feder and Burggren. Environmental Physiology of the Amphibians.
}
\examples{
\dontrun{
Qradiation_absorbed(a=0.9, A=1, psa_dir=0.4, psa_ref=0.4, S_dir=1000, S_dif=200, a_s=0.5)
}

}
\keyword{Solar}
\keyword{absorbed}
\keyword{radiation}
