% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{radians_to_degrees}
\alias{radians_to_degrees}
\title{Convert Angle in Radians to Degrees}
\usage{
radians_to_degrees(rad)
}
\arguments{
\item{rad}{\code{numeric} angle (radians).}
}
\value{
\code{numeric} angle (degrees).
}
\description{
Convert angle in radians to degrees.
}
\examples{
  radians_to_degrees(0.831)

}
