% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MISCfunctions.R
\name{radiation.erbs}
\alias{radiation.erbs}
\title{Radiation with cloudiness}
\usage{
radiation.erbs(rad, tau)
}
\arguments{
\item{rad}{angle in radians}

\item{tau}{clearnex index #!specify}
}
\description{
USE ERBS TO REPARTITION RADIATION (Olyphant 1984)
Separate Total radiation into components
kt is clearness index
Models presented in Wong and Chow. 2001. Applied Energy 69(2001):1991-224
Use Erbs et al model
}
\details{
This function allows you to calculate direct and diffused radtation.
}
\examples{
\dontrun{
radiation.erbs()
}
}
\keyword{radiation}
