% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationFunctions.R
\name{radiation}
\alias{radiation}
\title{Partition radiation into direct, diffuse, and reflected components using models from Campbell & Norman 1988}
\usage{
radiation(J, psi, tau, elev, rho = 0.7)
}
\arguments{
\item{J}{Julian Day}

\item{psi}{zenith angle in radians}

\item{tau}{transmissivity (%)}

\item{elev}{Elevation (m)}

\item{rho}{albedo (%)}
}
\description{
This function allows you to Partition radiation into direct, diffuse, and reflected components
}
\examples{
radiation()
}
\keyword{radiation}
