% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{radiation_absorbed}
\alias{radiation_absorbed}
\title{Calculate absorbed solar and thermal radiation}
\usage{
radiation_absorbed(abs = 0.9, As, psa_dir, psa_ref, R_dir, R_dif,
  R_ref = NA, rho_s = NA)
}
\arguments{
\item{abs}{solar absorptivity of animal surface (proportion), default value is for reptiles}

\item{As}{surface area  in m^2}

\item{psa_dir}{proportion surface area exposed to solar radiation}

\item{psa_ref}{proportion surface area exposed to reflected solar radiation}

\item{R_dir}{direct solar radiation (W/m^2)}

\item{R_dif}{diffuse solar radiation (W/m^2)}

\item{R_ref}{reflected solar radiation (W/m^2), either provided or estimated if surface albedo is provided instead}

\item{rho_S}{is surface albedo (proportion), optional (not used) if reflected radiation is provided}
}
\value{
solar radiation absorbed (W)
}
\description{
Calculate absorbed solar and thermal radiation
}
\details{
This function allows you to calculate solar and thermal radiation (W) absorbed by the surface of an animal. 
         Follows Gates Biophysical Ecology and Spotila et al. 1992.
}
\examples{
\dontrun{
radiation_absorbed(abs=0.9, As=1, psa_dir=0.4, psa_ref=0.4, R_dir=1000, R_dif=200, rho_s=0.5)
}

}
\keyword{Solar}
\keyword{absorbed}
\keyword{radiation}
