% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordCompression.R
\name{RecordCompression}
\alias{RecordCompression}
\title{Calculate compression of firn-core record.}
\usage{
RecordCompression(z.start = 0, adv, length.in, rate, res = 1)
}
\arguments{
\item{z.start}{the depth of the upper end of the firn-core record in [cm].}

\item{adv}{the amount of downward advection in [cm].}

\item{length.in}{the initial length of the firn-core record in [cm].}

\item{rate}{constant rate of firn densification relative to the surface
density; in [\% per metre].}

\item{res}{resolution (in [cm]) of the compression calculation, i.e. the
size of the individual depth bins of the record ("snow parcels") that are
advected downwards.}
}
\value{
the length change of the firn-core record in [cm].
}
\description{
Calculate the compression due to densification with a constant rate during
downward-advection of a firn-core record.
}
\examples{
# Length change of trench record for mean estimated densification rate
# (see Münch et al., 2017):
RecordCompression(adv = 50, length.in = 100, rate = 4.5)
}
\references{
Münch, T., et al., Constraints on post-depositional isotope modifications
in East Antarctic firn from analysing temporal changes of isotope profiles,
The Cryosphere, doi:10.5194/tc-11-2175-2017, 2017.
}
\seealso{
\code{\link{Compression}}
}
\author{
Thomas Münch
}
