% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-handling.R
\name{removeSurfaceRegion}
\alias{removeSurfaceRegion}
\title{Remove incomplete trench "surface" region}
\usage{
removeSurfaceRegion(data, .var = "d18O", vscale = "depth")
}
\arguments{
\item{data}{a trench data set following the default structure used in the
package.}

\item{.var}{character string with the name of a trench data variable for
which to check for missing values at the "surface".}

\item{vscale}{character string giving the name of the vertical scale
variable used in the trench data; defaults to \code{"depth"}.}
}
\value{
the input trench data set with the "surface region" removed.
}
\description{
When tabulating a trench data set on an absolute scale in vertical dimension
(e.g. along "depth"), sampling positions have no data, i.e. are NA, where the
vertical scale variable is below the maximum observed vertical scale across
the trench. This function removes this incomplete "surface" region to get a
trench data set without missing values.
}
\examples{
t13.trench2 \%>\% make2D()
removeSurfaceRegion(t13.trench2) \%>\% make2D()
}
\seealso{
getFirstCompleteDepthBin
}
\author{
Thomas Münch
}
