% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{Reynolds_number}
\alias{Reynolds_number}
\title{Calculate Reynolds Number}
\usage{
Reynolds_number(V, D, nu)
}
\arguments{
\item{V}{is wind speed in m/s}

\item{D}{is characteristic dimension (e.g., body diameter) (m)}

\item{nu}{is the kinematic viscosity, ratio of dynamic viscosity to density of the fluid (m^2 s^(-1)), can calculate from DRYAIR or WETAIR}
}
\value{
Reynolds number
}
\description{
Calculate Reynolds Number
}
\details{
This function allows you to estimate the Reynolds Number, which describes the dynamic properties of the fluid surrounding the animal as the ratio of internal viscous forces (Gates 1980 Biophysical Ecology).
}
\examples{
\dontrun{
Reynolds_number(V=1, D=0.001, nu=1.2)
}

}
\keyword{Reynolds}
\keyword{number}
