% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsd.R
\name{rmsd}
\alias{rmsd}
\title{Root-mean-square deviation.}
\usage{
rmsd(v1, v2, na.rm = FALSE)
}
\arguments{
\item{v1}{Numeric vector for which to compute the rmsd with \code{v2}.}

\item{v2}{Numeric vector for which to compute the rmsd with \code{v1}. Must
have the same length as \code{v1}.}

\item{na.rm}{a logical value indicating whether ‘NA’ values should be
stripped before the computation proceeds. Defaults to \code{FALSE}.}
}
\value{
The root-mean square deviation of \code{v1} and \code{v2}, or
    \code{NA} if any of the elements of the vectors is \code{NA} and if
    \code{na.rm = FALSE}.
}
\description{
This function calculates the root-mean square deviation of two numeric
vectors.
}
\author{
Thomas Münch
}
