% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllometricFunctions.R
\name{sa_from_length}
\alias{sa_from_length}
\title{Calculate surface area from length by approximating the animal's body as a rotational ellipsoid}
\usage{
sa_from_length(l)
}
\arguments{
\item{l}{Length in m}
}
\value{
sa (m^2)
}
\description{
Calculate surface area from length by approximating the animal's body as a rotational ellipsoid
}
\details{
Estimate surface area (m^2) from length (m) by approximating the animal's body as a rotational ellipsoid with half the body length as the semi-major axis.
}
\examples{
 \dontrun{
  sa_from_length(l=0.04)
}

}
\seealso{
Other allometric functions: 
\code{\link{mass_from_length}()},
\code{\link{prop_silhouette_area_shapes}()},
\code{\link{prop_silhouette_area}()},
\code{\link{sa_from_mass}()},
\code{\link{sa_from_volume}()},
\code{\link{volume_from_length}()}
}
\concept{allometric functions}
\keyword{area}
\keyword{surface}
