% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllometricFunctions.R
\name{sa_from_mass}
\alias{sa_from_mass}
\title{Calculate organism surface area from mass}
\usage{
sa_from_mass(m, taxa)
}
\arguments{
\item{m}{Mass in grams (g). \insertCite{Simonis2021}{TrenchR}}

\item{taxa}{Class of organism, current choices: \code{"lizard"}, \code{"salamander"}, \code{"frog"}, \code{"insect"}.}
}
\value{
Surface area in meters squared (m^2).
}
\description{
Estimate surface area (m^2) from mass (g) for a variety of taxa.
}
\details{
Relationships come from 
 \itemize{
   \item{lizard: \insertCite{Porter1979;textual}{TrenchR} and \insertCite{Porter1979;textual}{TrenchR} in \insertCite{Roughgarden1981;textual}{TrenchR}}
   \item{salamander: whit}
   \item{frog: \insertCite{McClanahan1969;textual}{TrenchR}}
   \item{insect: \insertCite{Lactin1997;textual}{TrenchR}}
 }
}
\examples{
 sa_from_mass(m = 2, taxa = "lizard")
 sa_from_mass(m = 2, taxa = "salamander")
 sa_from_mass(m = 2, taxa = "frog")
 sa_from_mass(m = 2, taxa = "insect")


}
\references{
\insertAllCited{}
}
\seealso{
Other allometric functions: 
\code{\link{mass_from_length}()},
\code{\link{prop_silhouette_area_shapes}()},
\code{\link{prop_silhouette_area}()},
\code{\link{sa_from_length}()},
\code{\link{sa_from_volume}()},
\code{\link{volume_from_length}()}
}
\concept{allometric functions}
\keyword{area}
\keyword{surface}
