% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllometricFunctions.R
\name{sa_from_volume}
\alias{sa_from_volume}
\title{Calculate organism surface area from volume}
\usage{
sa_from_volume(V, taxa)
}
\arguments{
\item{V}{\code{numeric} volume in meters cubed (\ifelse{html}{\out{m<sup>3</sup>}}{\eqn{m^3}{ASCII}}).}

\item{taxa}{\code{character} taxon of organism, current choices: \code{"lizard"}, \code{"frog"}, \code{"sphere"}.}
}
\value{
\code{numeric} surface area in meters squared (\ifelse{html}{\out{m<sup>2</sup>}}{\eqn{m^2}{ASCII}}).
}
\description{
Estimate surface area (\ifelse{html}{\out{m<sup>2</sup>}}{\eqn{m^2}{ASCII}}) from volume (\ifelse{html}{\out{m<sup>3</sup>}}{\eqn{m^3}{ASCII}}) for a variety of taxa following \insertCite{Mitchell1976}{TrenchR}
}
\details{
Relationships come from \itemize{
  \item Lizards: \insertCite{Norris1965;textual}{TrenchR}.
  \item Frogs: \insertCite{Tracy1972}{TrenchR}.
  \item Sphere: \insertCite{Mitchell1976}{TrenchR}.
 }
}
\examples{
  sa_from_volume(V = 0.001, taxa = "lizard")
  sa_from_volume(V = 0.001, taxa = "frog")
  sa_from_volume(V = 0.001, taxa = "sphere")

}
\references{
\insertAllCited{}
}
\seealso{
Other allometric functions: 
\code{\link{mass_from_length}()},
\code{\link{prop_silhouette_area_shapes}()},
\code{\link{prop_silhouette_area}()},
\code{\link{sa_from_length}()},
\code{\link{sa_from_mass}()},
\code{\link{volume_from_length}()}
}
\concept{allometric functions}
\keyword{area}
\keyword{surface}
