% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{saturation_vapor_pressure}
\alias{saturation_vapor_pressure}
\title{Calculate saturation vapor pressure}
\usage{
saturation_vapor_pressure(T_a)
}
\arguments{
\item{T_a}{air temperature (K)}
}
\value{
saturation vapor pressure, e_s (kPa)
}
\description{
Calculate saturation vapor pressure
}
\details{
Calculate saturation vapor pressure (kPa) based on the Clausius-Clapeyron equation (Stull 2000). Source: Riddell EA. 2017. Physical calculations of resistance to water loss improve predictions of species range models. Ecological Monographs 87: 21-23.
}
\examples{
\dontrun{
saturation_vapor_pressure(T_a=293)
}

}
\seealso{
Other biophysical models: \code{\link{Free_or_forced_convection}},
  \code{\link{Grashof_number_Gates}},
  \code{\link{Grashof_number}}, \code{\link{Nu_from_Gr}},
  \code{\link{Nu_from_Re}}, \code{\link{Nusselt_number}},
  \code{\link{Prandtl_number}},
  \code{\link{Qconduction_animal}},
  \code{\link{Qconduction_substrate}},
  \code{\link{Qconvection}},
  \code{\link{Qemitted_thermal_radiation}},
  \code{\link{Qevaporation}},
  \code{\link{Qmetabolism_from_mass_temp}},
  \code{\link{Qmetabolism_from_mass}},
  \code{\link{Qnet_Gates}},
  \code{\link{Qradiation_absorbed}},
  \code{\link{Qthermal_radiation_absorbed}},
  \code{\link{Reynolds_number}},
  \code{\link{Tb_CampbellNorman}}, \code{\link{Tb_Fei}},
  \code{\link{Tb_Gates}}, \code{\link{Tb_butterfly}},
  \code{\link{Tb_grasshopper}}, \code{\link{Tb_limpetBH}},
  \code{\link{Tb_limpet}}, \code{\link{Tb_lizard}},
  \code{\link{Tb_mussel}},
  \code{\link{Tb_salamander_humid}},
  \code{\link{Tb_snail}}, \code{\link{Tbed_mussel}},
  \code{\link{Tsoil}}, \code{\link{actual_vapor_pressure}},
  \code{\link{boundary_layer_resistance}},
  \code{\link{external_resistance_to_water_vapor_transfer}},
  \code{\link{heat_transfer_coefficient_approximation}},
  \code{\link{heat_transfer_coefficient_simple}},
  \code{\link{heat_transfer_coefficient}},
  \code{\link{saturation_water_vapor_pressure}}
}
\author{
Eric Riddell
}
\concept{biophysical models}
\keyword{pressure}
\keyword{saturation}
\keyword{vapor}
