% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{saturation_water_vapor_pressure}
\alias{saturation_water_vapor_pressure}
\title{Approximate saturation water vapor pressure}
\usage{
saturation_water_vapor_pressure(Ta)
}
\arguments{
\item{Ta}{air temperature (C)}
}
\value{
Saturation water vapor pressure, e_s (Pa)
}
\description{
Approximate saturation water vapor pressure
}
\details{
Approximate saturation water vapor pressure as a function of ambient temperature for temperatures from 0 to 40C (Rosenberg 1974 in Spotila et al. 1992, see also NichMapR WETAIR and DRYAIR functions from Tracy et al. 1980).
}
\examples{
\dontrun{
saturation_water_vapor_pressure(Ta=20)
}

}
\keyword{Saturation}
\keyword{pressure}
\keyword{vapor}
\keyword{water}
