% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TsoilFunctions.R
\name{soil_specific_heat}
\alias{soil_specific_heat}
\title{Estimate soil specific heat}
\usage{
soil_specific_heat(x_o, x_m, x_w, rho_so)
}
\arguments{
\item{x_o}{is volume fraction of organic material}

\item{x_m}{is volume fraction of minerals}

\item{x_w}{is volume fraction of water}

\item{rho_so}{is particle density of soil in kg/m3 (bulk density)}
}
\description{
This function allows you to estimate soil specific heat in J kg^-1 K-1 using the methods of de Vries (1963, The Physics of Plant Environments, Ch2 in Environmental Control of Plant Growth).  Uses the volume fraction of organic material, minerals, and water in soil.  CHECK Campbell and Norman (2000) section 8.2.
}
\details{
Estimate soil specific heat in J kg^-1 K-1.
}
\examples{
\dontrun{
soil_specific_heat(0.01, 0.6, 0.2, 1620)
}
}
\keyword{soil}
\keyword{temperature}
