% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TsoilFunctions.R
\name{soil_temp_integrand}
\alias{soil_temp_integrand}
\title{Solve equation for soil temperature}
\usage{
soil_temp_integrand(x, L, z0)
}
\arguments{
\item{x}{is a vector of volume fractions of soil constituents (e.g., clay, quartz, minerals other than quartz, organic matter, water, air).  The volume fractions should sum to 1. Note that x and lambda values in the example correspond to these soil constituents.}

\item{L}{is ???????????????????????}

\item{z0}{is surface rougness in m}
}
\description{
Function called by soil_temp_noint to solve equation for soil temperature from Beckman et al. (1973, Thermal Model for Prediction of a Desert Iguana's Daily and Seasonal Behavior).
}
\details{
Function called by soil_temp_noint to solve equation for soil temperature.
}
\examples{
\dontrun{
soil_temp_integrand(x=c(0.10,0.40,0.11,0.01,0.2, 0.18), L=-10,z0=0.2)
}
}
