% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TsoilFunctions.R
\name{soil_temperature_equation}
\alias{soil_temperature_equation}
\title{Function called by soil_temperature_function() to solve equation for soil temperature.}
\usage{
soil_temperature_equation(L, rho_a, c_a, k, V_inst, z_r, z0, T_inst, T_s)
}
\arguments{
\item{L}{is the Monin-Obukhov length, a measure of the instability of heat flow (see Beckman et al. 1973)}

\item{rho_a}{is density of air in kg m^-3}

\item{c_a}{is specific heat of air (J/(kg*K))}

\item{k}{is von Karman's constant}

\item{V_inst}{is instantaneous wind speed in m/s}

\item{z_r}{is reference height in m}

\item{z0}{is surface roughness in m}

\item{T_inst}{instantaneous air temperature in K}

\item{T_s}{initial soil suface temperature in °C}
}
\value{
soil temperature function
}
\description{
Function called by soil_temp_noint to solve equation for soil temperature from Beckman et al. (1973, Thermal Model for Prediction of a Desert Iguana's Daily and Seasonal Behavior).
}
\details{
Function called by soil_temperature_function() to solve equation for soil temperature.
}
\examples{
\dontrun{
soil_temperature_equation(L=-10, rho_a=1.177, c_a=1006, k=0.41, V_inst=0.3, z_r=1.5, z0=0.02, T_inst=265, T_s=20)
}
}
\seealso{
Other soil temperature functions: \code{\link{soil_conductivity}},
  \code{\link{soil_specific_heat}},
  \code{\link{soil_temperature_function}},
  \code{\link{soil_temperature_integrand}},
  \code{\link{soil_temperature}}
}
\author{
Joseph Grigg
}
\concept{soil temperature functions}
\keyword{soil}
\keyword{temperature}
