% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TsoilFunctions.R
\name{soil_temperature_noint}
\alias{soil_temperature_noint}
\title{Function to calculate soil temperature in C using ODEs.}
\usage{
soil_temperature_noint(z.intervals = 12, z, Tair, u_z, Tsoil0, z0, SSA,
  TimeIn, solrad, water_content = 0.2, air_pressure, rho_so = 1620,
  shade = FALSE)
}
\arguments{
\item{z.intervals}{is the number of intervals in the soil profile to calculate}

\item{z}{is reference height in m}

\item{Tair}{is a vector of air temperature in degrees C, Note: missing values will be linearly interpolated}

\item{u_z}{is a vector of wind speed (m/s)}

\item{Tsoil0}{is the initial soil temperature in degrees C}

\item{z0}{is surface roughness in m}

\item{SSA}{is the solar absorbtivity of soil surface as a fraction}

\item{TimeIn}{is a vector of time periods for the model}

\item{solrad}{is solar radiation in W m^-2}

\item{water_content}{is percent water content (%)}

\item{air_pressure}{is air pressure in kPa}

\item{shade}{is whether or not soil temperature should be calculated in the shade, TRUE or FALSE}

\item{rho_so=}{1620 particle density of soil}
}
\description{
Function called to calculate soil temperature in C from Beckman et al. (1973, Thermal Model for Prediction of a Desert Iguana's Daily and Seasonal Behavior). Wrapper for soil_temperature function that uses ODE to calculate soil profile.
}
\details{
Function to calculate soil temperature in C using ODEs.
}
\examples{
\dontrun{
temp_vector= runif(96, min=-10, max=10)
wind_speed_vector= runif(96, min=0, max=0.4)
time_vector= rep(1:24,4)
solrad_vector= rep(c(rep(0,6),seq(10,700,length.out=6), seq(700,10,length.out=6),rep(0,6)),4)

soil_temperature_noint(z.intervals=12,z=1.5, Tair=temp_vector, u_z=wind_speed_vector, Tsoil0= 20, z0=0.02, SSA=0.7, TimeIn=time_vector, solrad= solrad_vector, water_content=0.2, air_pressure=85, rho_so=1620, shade=FALSE)
}
}
