% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{solar_noon}
\alias{solar_noon}
\title{Calculate time of solar noon}
\usage{
solar_noon(lon, doy)
}
\arguments{
\item{lon}{longitude in degrees}

\item{doy}{day of year}
}
\value{
time at solar noon
}
\description{
Calculate time of solar noon
}
\details{
This function allows you to calculate the time of solar noon in hours as a function of the day of year and longitude
}
\examples{
\dontrun{
solar_noon(-122.335, 112)
}
}
\keyword{Solar}
\keyword{noon}
\keyword{time}
