% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{solar_radiation_absorbed}
\alias{solar_radiation_absorbed}
\title{Calculate absorbed solar radiation}
\usage{
solar_radiation_absorbed(sa, solar, thermal_abs = 0.965)
}
\arguments{
\item{sa}{surface area in m^2}

\item{solar}{Downward solar radiation(W/m^2).}

\item{thermal_abs}{thermal absorptivity as a percent, default value is for lizards, Bartlett & Gates 1967.}
}
\value{
solar radiation absorbed
}
\description{
Calculate absorbed solar radiation
}
\details{
This function allows you to calculate absorbed solar radiation of an ectotherm(Reptiles)
}
\examples{
\dontrun{
solar_radiation_absorbed(10.5,1000,"lizard")
}

}
\keyword{Solar}
\keyword{absorbed}
\keyword{radiation}
