% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBin.R
\name{summarizeBin}
\alias{summarizeBin}
\title{Bin summary values}
\usage{
summarizeBin(
  x,
  bins,
  summary = "mean",
  right = TRUE,
  label.name = "",
  na.rm = FALSE
)
}
\arguments{
\item{x}{a data frame with two columns supplying a trench mean profile
(e.g. obtained from \code{\link{makeMean}}): the first column provides the
vertical scale and the second column the proxy values.}

\item{bins}{a data frame with two columns supplying the break point positions
defining the bins: the first column provides the break points in units of
the vertical scale (e.g. \code{"depth"}) and the second column provides
corresponding label values for the bins (e.g. the years for a depth-age
relationship).}

\item{summary}{character string for the summary operation to perform on each
bin: default \code{"mean"} is to calculate the mean value across each bin;
other options are \code{"min"} and \code{"max"} for the minimum and maximum
values, respectively.}

\item{right}{logical; indicate whether the bins should be closed on the right
and open on the left (\code{TRUE}, the default), or vice versa
(\code{FALSE}). For \code{n} break point positions (i.e. = the number of
rows of \code{bins}), there are \code{n-1} bin intervals with corresponding
summary values. If the bins are closed on the right, the label values 2 : n
in \code{bins} are used to label the output; if the bins are closed on the
left, the first \code{n-1} label values are used instead.}

\item{label.name}{optional name for the label value column in the output data
frame; if none is specified (the default) "label" is used.}

\item{na.rm}{logical; indicate whether \code{NA} values should be stripped
before the computation of the bin summary proceeds.}
}
\value{
a tibble with two columns and \code{n-1} rows, where the first column
  provides the bin label values and the second column the proxy summary
  values for those bins.
}
\description{
Calculate summary values of a trench mean profile partitioned into
differently sized bins. Possible summary operations are the mean value,
e.g. to calculate annual mean values based on a given depth-age relationship,
or minimum and maximum values.
}
\examples{

# bin average for T13-1, but with completely random bins

x <- makeMean(t13.trench1)
bins <- data.frame(depth = c(1.5, 22.5, 46.5, 88.5), bin_no = 1 : 4)

summarizeBin(x, bins)
summarizeBin(x, bins, na.rm = TRUE)

summarizeBin(x, bins, na.rm = TRUE,
             right = FALSE, label.name = "year")

# winter d-excess minima from actual dating

bins <- data.frame(depth = c(1.5, 28.5, 49.5, 67.5, 94.5),
                   year = 2013 : 2009)

library(magrittr)
makeMean(t13.trench1, .var = "dxs", na.rm = TRUE) \%>\%
  summarizeBin(bins, summary = "min")

}
\author{
Thomas Münch
}
