% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeTrench.R
\name{summarizeTrench}
\alias{summarizeTrench}
\title{Retrieve summary information}
\usage{
summarizeTrench(data, .var = NULL)
}
\arguments{
\item{data}{a generic trench dataset.}

\item{.var}{optional character string with the name of a trench variable for
which summary information shall be retrieved; default setting only returns
information on the sampled trench profiles, see the examples.}
}
\value{
a tibble with one row.
}
\description{
Summarize a trench dataset by retrieving information on the number and
distances of sampled profiles and on statistical data of a requested trench
variable: number of samples and their range, mean, and standard deviation.
}
\examples{

summarizeTrench(t13.trench1)
summarizeTrench(t13.trench1, .var = "d18O")
summarizeTrench(t13.trench1, .var = "dD")

}
\author{
Thomas Münch
}
