% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docu.R
\docType{data}
\name{t13.trench2}
\alias{t13.trench2}
\title{Trench T13-2 stable water isotopologue data and meta information}
\format{
A list with three components:
\describe{
  \item{data:}{A 3 x 38 x 4 array with the isotopologue data: the first
    index refers to the isotopic species (1 = d18O, 2 = d2H, 3 =
    d-excess), the second index to the sample depth, the third index to the
    profile position.}
  \item{depth:}{Numeric vector of length 38 providing the sample depths in
    [cm] on the absolute depth scale, thus 0 cm referring to the maximum
    surface height of the trench.}
  \item{meta:}{A data frame with 4 observations of 3 variables:
  \describe{
    \item{profileName:}{Character vector with the profile names.}
    \item{profilePos:}{Numeric vector of the horizontal profile positions in
      [cm] along the trench.}
    \item{profileSurfaceHeight:}{Numeric vector of the surface heights in
    [cm] at the profile positions (positive downwards) relative to the
    maximum surface height of the trench.}}}
}
}
\source{
\url{https://doi.pangaea.de/10.1594/PANGAEA.861675}
}
\usage{
t13.trench2
}
\description{
A dataset providing the stable water isotopologue data measured on Kohnen
Snow Trench T13-2 (oxygen isotopes, hydrogen isotopes, and deuterium excess)
together with meta information.
}
\details{
To create this dataset, the data available from PANGAEA has to be converted
to the format described here.
}
\keyword{datasets}
