% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T13AnnualMeans.R
\name{T13AnnualMeans}
\alias{T13AnnualMeans}
\title{T13 annual mean isotope time series.}
\usage{
T13AnnualMeans(t1, t2, depth, i.max = c(5, 11, 19, 24, 33, 38), i.min = c(7,
  14, 21, 26, 37), start.year = 2013)
}
\arguments{
\item{t1}{}

\item{t2}{}

\item{depth}{}
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Thomas Muench
}
