% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docu.R
\docType{data}
\name{t15.trench2}
\alias{t15.trench2}
\title{Trench T15-2 Stable Water Isotopologue Data and Meta Information.}
\format{
A list with three components:
\describe{
  \item{data:}{A 3 x 125 x 11 array with the isotopologue data: the first
    index refers to the isotopic species (1 = d18O, 2 = d2H, 3 =
    d-excess), the second index to the sample depth, the third index to the
    profile position.}
  \item{depth:}{Numeric vector of length 125 providing the sample depths in
    [cm] on the absolute depth scale, thus 0 cm referring to the maximum
    surface height of the trench.}
  \item{meta:}{A data frame with 11 observations of 3 variables:
  \describe{
    \item{profileName:}{Character vector with the profile names.}
    \item{profilePos:}{Numeric vector of the horizontal profile positions in
      [m] along the trench.}
    \item{profileSurfaceHeight:}{Numeric vector of the surface heights in
    [cm] at the profile positions (positive downwards) relative to the
    maximum surface height of the trench.}}}
}
}
\source{
\url{https://doi.pangaea.de/10.1594/PANGAEA.876639}
}
\usage{
t15.trench2
}
\description{
A dataset providing the stable water isotopologue data measured on Kohnen
Snow Trench T15-2 (oxygen isotopes, hydrogen isotopes, and deuterium excess)
together with meta information.
}
\details{
To create this dataset, the data available from PANGAEA has to be converted
to the format described here.
}
\keyword{datasets}
