% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{T_sky}
\alias{T_sky}
\title{Effective radiant temperature of sky (K)}
\usage{
T_sky(T_a, formula)
}
\arguments{
\item{T_a}{\code{numeric} ambient air temperature (K)}

\item{formula}{\code{character} Formula to use, current choices are: \code{"Brunt"}, \code{"Swinbank"}.}
}
\value{
\code{numeric} T_sky (K).
}
\description{
This function estimates the effective radiant temperature of the sky (K) using either the Brunt or Swinbank formulations \insertCite{Gates1980}{TrenchR}.
}
\examples{
  T_sky(T_a=293, formula="Swinbank")

}
\references{
\insertAllCited{}
}
\concept{biolphysical models}
