% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperativeTemperatureFunctions.R
\name{Tb_CampbellNorman}
\alias{Tb_CampbellNorman}
\title{Predicts body temperatures (operative environmental temperature) of an ectotherm in K.}
\usage{
Tb_CampbellNorman(T_a, S, epsilon = 0.96, c_p = 29.3, D, V)
}
\arguments{
\item{T_a}{is air temperature in K.}

\item{S}{solar and thermal radiation absorbed (W m^-2)}

\item{epsilon}{longwave infrared emissivity of skin (proportion), 0.95 to 1 for most animals (Gates 1980)}

\item{c_p}{specific heat of air (J mol^-1 K^-1)}

\item{D}{characteristic dimension of the animal (m)}

\item{V}{is wind speed in m/s}
}
\value{
operative environmental temperature, T_e (K)
}
\description{
Predicts body temperatures (operative environmental temperature) of an ectotherm in K.
}
\details{
Predicts body temperatures (operative environmental temperature) of an ectotherm in K. Uses approximation in Campbell and Norman (1998).
}
\examples{
\dontrun{
Tb_CampbellNorman(T_a=303, S=823, epsilon=0.96, c_p=29.3, D=0.17, V=1)
}

}
\seealso{
Other biophysical models: \code{\link{Free_or_forced_convection}},
  \code{\link{Grashof_number_Gates}},
  \code{\link{Grashof_number}}, \code{\link{Nu_from_Gr}},
  \code{\link{Nu_from_Re}}, \code{\link{Nusselt_number}},
  \code{\link{Prandtl_number}},
  \code{\link{Qconduction_animal}},
  \code{\link{Qconduction_substrate}},
  \code{\link{Qconvection}},
  \code{\link{Qemitted_thermal_radiation}},
  \code{\link{Qevaporation}},
  \code{\link{Qmetabolism_from_mass_temp}},
  \code{\link{Qmetabolism_from_mass}},
  \code{\link{Qnet_Gates}},
  \code{\link{Qradiation_absorbed}},
  \code{\link{Qthermal_radiation_absorbed}},
  \code{\link{Reynolds_number}}, \code{\link{Tb_Fei}},
  \code{\link{Tb_Gates}}, \code{\link{Tb_butterfly}},
  \code{\link{Tb_grasshopper}}, \code{\link{Tb_lizard}},
  \code{\link{Tb_mussel}},
  \code{\link{Tb_salamander_humid}}, \code{\link{Tsoil}},
  \code{\link{actual_vapor_pressure}},
  \code{\link{boundary_layer_resistance}},
  \code{\link{external_resistance_to_water_vapor_transfer}},
  \code{\link{heat_transfer_coefficient_approximation}},
  \code{\link{heat_transfer_coefficient_simple}},
  \code{\link{heat_transfer_coefficient}},
  \code{\link{saturation_vapor_pressure}},
  \code{\link{saturation_water_vapor_pressure}}
}
\concept{biophysical models}
\keyword{environmental}
\keyword{operative}
\keyword{temperature}
