% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_To.R
\name{Tb_Fei}
\alias{Tb_Fei}
\title{Calculate body temperature(Operative Temperature) of Lizard using model by Fei et al.}
\usage{
Tb_Fei(TAH, T_s, T_a, UDT, GLW, shade)
}
\arguments{
\item{TAH}{Temperature in canopy (K).}

\item{T_s}{Surface Temperature (K).}

\item{T_a}{Air Temperature (K).}

\item{UDT}{upper developmental threshold.}

\item{GLW}{Downward flux of near-infrared radiation}

\item{shade}{type of method being used}
}
\description{
[1] T. Fei, A. K. Skidmore, V. Venus, T. Wang, M. Schlerf, B. Toxopeus, 
S. van Overjijk, M. Bian, and Y. Liu, “A body temperature model for 
lizards as estimated from the thermal environment,” J. Therm. 
Biol., vol. 37, no. 1, pp. 56–64, 2012.
TAH, T_s, T_a, UDT, GLW, time_step, shade
}
\details{
Predicts body temperature (operative environmental temperature) of a lizard in K.
}
\examples{
\dontrun{
Tb_Fei(TAH=293, T_s=300, T_a=295, UDT=600, GLW=30, shade=0.5)
}

}
\author{
Ofir Levy
}
\keyword{Fei}
\keyword{Operative}
\keyword{Temperature}
