% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biophysmodel_Grasshopper.R
\name{Tb_grasshopper}
\alias{Tb_grasshopper}
\title{Predicts body temperatures (operative environmental temperature) of a grasshopper in degrees C.}
\usage{
Tb_grasshopper(Ta, Tg, u, rad, kt, psi, L, Acondfact = 0.25, z = 0.001,
  abs = 0.7, r_g = 0.3)
}
\arguments{
\item{Ta}{is air temperature in C}

\item{Tg}{is surface temperature in C, Kingsolver (1983) assumes Tg-Ta=8.4}

\item{u}{is wind speed in m/s}

\item{rad}{is total (direct + diffuse) solar radiation flux in W/m^2}

\item{kt}{is the clearnex index (dimensionless), which is the ratio of the global solar radiation measured at the surface to the total solar radiation at the top of the atmosphere.}

\item{psi}{is solar zenith angle in degrees}

\item{L}{in grasshopper length in m}

\item{Acondfact}{is the proportion of the grasshopper surface area that is in contact with the ground}

\item{z}{is grasshopper's distance from the ground in m}

\item{abs}{is absorptivity of grasshopper to solar radiation (proportion), See Anderson et al. (1979).}

\item{r_g}{is substrate solar reflectivity (proportion), see Kingsolver (1983)}
}
\description{
Predicts body temperature (operative environmental temperature) of a grasshopper in degrees C. Described in Buckleyet al. (2014, Phenotypic clines, energy balances, and ecological responses to climate change. Journal of Animal Ecology 83:41-50.) See also a related model by Anderson et al. (1979, Habitat selection in two species of short-horned grasshoppers. Oecologia 38:359–74.)
}
\details{
Predicts body temperatures (operative environmental temperature) of a grasshopper in degrees C.
}
\examples{
\dontrun{
Tb_grasshopper(Ta=25, Tg=25, u=0.4, rad=400, kt=0.7, psi=30, L=0.02, Acondfact=0.25, z=0.001, abs=0.7, r_g=0.3)
}
}
\keyword{biophysical}
\keyword{body}
\keyword{model}
\keyword{temperature,}
