% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biophysmodel_Limpet.R
\name{Tb_limpet}
\alias{Tb_limpet}
\title{Predicts body temperature (operative environmental temperature) of a limpet in °C.}
\usage{
Tb_limpet(T_a, T_r, L, H, I, u, psi, c, position = "anterior")
}
\arguments{
\item{T_a}{air temperature (°C)}

\item{T_r}{rock temperature (°C)}

\item{L}{limpet length (anterior/posterior axis) (m)}

\item{H}{limpet height (dorsal/ventral axis) (m)}

\item{I}{solar irradiance (W m^-2)}

\item{u}{wind speed (m/s)}

\item{psi}{solar zenith angle (degrees): can be calculated from zenith_angle function}

\item{c}{fraction of the sky covered by cloud}

\item{position}{the direction of the limpet that is facing upwind. Options are "anterior", "posterior" and "broadside".}
}
\value{
predicted body temperature (°C)
}
\description{
Predicts body temperature of a limpet in °C. Based on Denny and Harley 2006. Hot limpets: predicting body temperature in a conductance-mediated thermal system
}
\details{
Predicts body temperature of a limpet in °C.
}
\examples{
\dontrun{
Tb_limpet(
  T_a = 25, 
  T_r = 30, 
  L = 0.0176, 
  H = 0.0122, 
  I = 1300, 
  u = 1, 
  psi = 30, 
  c = 1, 
  position = "anterior")
}
}
\seealso{
Other biophysical models: 
\code{\link{Free_or_forced_convection}()},
\code{\link{Grashof_number_Gates}()},
\code{\link{Grashof_number}()},
\code{\link{Nu_from_Gr}()},
\code{\link{Nu_from_Re}()},
\code{\link{Nusselt_number}()},
\code{\link{Prandtl_number}()},
\code{\link{Qconduction_animal}()},
\code{\link{Qconduction_substrate}()},
\code{\link{Qconvection}()},
\code{\link{Qemitted_thermal_radiation}()},
\code{\link{Qevaporation}()},
\code{\link{Qmetabolism_from_mass_temp}()},
\code{\link{Qmetabolism_from_mass}()},
\code{\link{Qnet_Gates}()},
\code{\link{Qradiation_absorbed}()},
\code{\link{Qthermal_radiation_absorbed}()},
\code{\link{Reynolds_number}()},
\code{\link{Tb_CampbellNorman}()},
\code{\link{Tb_Fei}()},
\code{\link{Tb_Gates}()},
\code{\link{Tb_butterfly}()},
\code{\link{Tb_grasshopper}()},
\code{\link{Tb_limpetBH}()},
\code{\link{Tb_lizard}()},
\code{\link{Tb_mussel}()},
\code{\link{Tb_salamander_humid}()},
\code{\link{Tb_snail}()},
\code{\link{Tbed_mussel}()},
\code{\link{Tsoil}()},
\code{\link{actual_vapor_pressure}()},
\code{\link{boundary_layer_resistance}()},
\code{\link{external_resistance_to_water_vapor_transfer}()},
\code{\link{heat_transfer_coefficient_approximation}()},
\code{\link{heat_transfer_coefficient_simple}()},
\code{\link{heat_transfer_coefficient}()},
\code{\link{saturation_vapor_pressure}()},
\code{\link{saturation_water_vapor_pressure}()}
}
\concept{biophysical models}
