% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biophysmodel_Sceloporus.R
\name{Tb_lizard}
\alias{Tb_lizard}
\title{Predicts body temperatures (operative environmental temperature) of a lizard in °C.}
\usage{
Tb_lizard(
  T_a,
  T_g,
  u,
  svl,
  m,
  psi,
  rho_S,
  elev,
  doy,
  sun = TRUE,
  surface = TRUE,
  alpha_S = 0.9,
  alpha_L = 0.965,
  epsilon_s = 0.965,
  F_d = 0.8,
  F_r = 0.5,
  F_a = 0.5,
  F_g = 0.5
)
}
\arguments{
\item{T_a}{is air temperature in °C}

\item{T_g}{is surface temperature in °C}

\item{u}{is wind speed in m/s}

\item{svl}{is lizard snout vent length in mm}

\item{m}{is lizard mass in g, note that it can be estimated as massfromsvl=function(svl) 3.55*10^-5*(svl)^3.00 (Tinkle and Ballinger 1972)}

\item{psi}{is solar zenith angle in degrees}

\item{rho_S}{is surface albedo (proportion). ~0.25 for grass, ~0.1 for dark soil, >0.75 for fresh snow (Campbell & Norman 1998)}

\item{elev}{is elevation in m}

\item{doy}{is day of year (1-366)}

\item{sun}{indicates whether lizard is in sun (TRUE) or shade (FALSE)}

\item{surface}{indicates whether lizard is on ground surface (TRUE) or above surface (FALSE, e.g. in a tree)}

\item{alpha_S}{is lizard solar absorptivity, alpha_S=0.9 (Gates 1980, Table 11.4)}

\item{alpha_L}{is lizard thermal absoptivity, alpha_L=0.965 (Bartlett & Gates 1967)}

\item{epsilon_s}{is surface emisivity of lizards, epsilon_s=0.965 (Bartlett & Gates 1967)}

\item{F_d}{is the view factor between the surface of the lizard and diffuse solar radiation (Bartlett & Gates 1967)}

\item{F_r}{is the view factor between the surface of the lizard and reflected solar radiation}

\item{F_a}{is the view factor between the surface of the lizard and atmospheric radiation}

\item{F_g}{is the view factor between the surface of the lizard and ground thermal radation}
}
\value{
T_e Operative temperature (°C)
}
\description{
Predicts body temperature (operative environmental temperature) of a lizard in °C. 
Based on Campbell and Norman (1998, An introduction to environmental biophysics). 
Designed for Sceloporus lizards and described in Buckley (2008, Linking traits to energetics and population dynamics to predict lizard ranges in changing environments. American Naturalist 171: E1-E19).
}
\details{
Predicts body temperatures (operative environmental temperature) of a lizard in °C.
}
\examples{
\dontrun{
Tb_lizard(
  T_a=25, 
  T_g=30, 
  u=0.1, 
  svl=60, 
  m=10, 
  psi=34, 
  rho_S=0.24, 
  elev=500, 
  doy=200, 
  sun=TRUE, 
  surface=TRUE, 
  alpha_S=0.9, 
  alpha_L=0.965, 
  epsilon_s=0.965, 
  F_d=0.8, 
  F_r=0.5, 
  F_a=0.5, 
  F_g=0.5)
}

}
\seealso{
Other biophysical models: 
\code{\link{Free_or_forced_convection}()},
\code{\link{Grashof_number_Gates}()},
\code{\link{Grashof_number}()},
\code{\link{Nu_from_Gr}()},
\code{\link{Nu_from_Re}()},
\code{\link{Nusselt_number}()},
\code{\link{Prandtl_number}()},
\code{\link{Qconduction_animal}()},
\code{\link{Qconduction_substrate}()},
\code{\link{Qconvection}()},
\code{\link{Qemitted_thermal_radiation}()},
\code{\link{Qevaporation}()},
\code{\link{Qmetabolism_from_mass_temp}()},
\code{\link{Qmetabolism_from_mass}()},
\code{\link{Qnet_Gates}()},
\code{\link{Qradiation_absorbed}()},
\code{\link{Qthermal_radiation_absorbed}()},
\code{\link{Reynolds_number}()},
\code{\link{Tb_CampbellNorman}()},
\code{\link{Tb_Fei}()},
\code{\link{Tb_Gates}()},
\code{\link{Tb_butterfly}()},
\code{\link{Tb_grasshopper}()},
\code{\link{Tb_limpetBH}()},
\code{\link{Tb_limpet}()},
\code{\link{Tb_mussel}()},
\code{\link{Tb_salamander_humid}()},
\code{\link{Tb_snail}()},
\code{\link{Tbed_mussel}()},
\code{\link{Tsoil}()},
\code{\link{actual_vapor_pressure}()},
\code{\link{boundary_layer_resistance}()},
\code{\link{external_resistance_to_water_vapor_transfer}()},
\code{\link{heat_transfer_coefficient_approximation}()},
\code{\link{heat_transfer_coefficient_simple}()},
\code{\link{heat_transfer_coefficient}()},
\code{\link{saturation_vapor_pressure}()},
\code{\link{saturation_water_vapor_pressure}()}
}
\concept{biophysical models}
