% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biophysmodel_Mussel.R
\name{Tb_mussel}
\alias{Tb_mussel}
\title{Predicts body temperature (operative environmental temperature) of a mussel in 째C.}
\usage{
Tb_mussel(L, H, T_a, T_g, u, psi, S, c, group = TRUE)
}
\arguments{
\item{L}{mussel length (anterior/posterior axis) (m)}

\item{H}{mussel height (dorsal/ventral axis) (m)}

\item{u}{wind speed (m/s)}

\item{psi}{solar zenith angle (degrees): can be calculated from zenith_angle function}

\item{S}{direct solar flux density acting upon that area (W m^-2)}

\item{c}{fraction of the sky covered by cloud}

\item{group}{TRUE if aggregated, FALSE if solitary}
}
\value{
predicted body temperature (째C)
}
\description{
Predicts body temperature of a mussel in 째C. Based on Helmuth 1998, INTERTIDAL MUSSEL MICROCLIMATES: PREDICTING THE BODY TEMPERATURE OF A SESSILE INVERTEBRATE
}
\details{
Predicts body temperature of a mussel in 째C.
}
\examples{
\dontrun{
Tb_mussel(L = 0.1, H = 0.05, T_a = 25, T_g = 30, u = 2, psi = 30, S = 500, c = 0, group = FALSE)
}
}
\seealso{
Other biophysical models: \code{\link{Free_or_forced_convection}},
  \code{\link{Grashof_number_Gates}},
  \code{\link{Grashof_number}}, \code{\link{Nu_from_Gr}},
  \code{\link{Nu_from_Re}}, \code{\link{Nusselt_number}},
  \code{\link{Prandtl_number}},
  \code{\link{Qconduction_animal}},
  \code{\link{Qconduction_substrate}},
  \code{\link{Qconvection}},
  \code{\link{Qemitted_thermal_radiation}},
  \code{\link{Qevaporation}},
  \code{\link{Qmetabolism_from_mass_temp}},
  \code{\link{Qmetabolism_from_mass}},
  \code{\link{Qnet_Gates}},
  \code{\link{Qradiation_absorbed}},
  \code{\link{Qthermal_radiation_absorbed}},
  \code{\link{Reynolds_number}},
  \code{\link{Tb_CampbellNorman}}, \code{\link{Tb_Fei}},
  \code{\link{Tb_Gates}}, \code{\link{Tb_butterfly}},
  \code{\link{Tb_grasshopper}}, \code{\link{Tb_lizard}},
  \code{\link{Tb_salamander_humid}}, \code{\link{Tsoil}},
  \code{\link{actual_vapor_pressure}},
  \code{\link{boundary_layer_resistance}},
  \code{\link{external_resistance_to_water_vapor_transfer}},
  \code{\link{heat_transfer_coefficient_approximation}},
  \code{\link{heat_transfer_coefficient_simple}},
  \code{\link{heat_transfer_coefficient}},
  \code{\link{saturation_vapor_pressure}},
  \code{\link{saturation_water_vapor_pressure}}
}
\concept{biophysical models}
\keyword{biophysical}
\keyword{body}
\keyword{model}
\keyword{temperature,}
