% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{Tb_salamander_humid}
\alias{Tb_salamander_humid}
\title{Calculate humid operative temperature}
\usage{
Tb_salamander_humid(r_i, r_b, diameter, Ta, elev, e_a, e_s, Rabs,
  emissivity = 0.96)
}
\arguments{
\item{r_i}{internal (skin) resistance (s cm^-1)}

\item{r_b}{boundary layer resistance (s cm^-1)}

\item{diameter}{body diameter (m), ##diameter = 0.0016*log(mass) + 0.0061 for mass(g) #empirical formula for diameter, Riddell et al. 2017}

\item{Ta}{ambient temperature (C)}

\item{elev}{elevation (m)}

\item{e_a}{actual vapor pressure (kPa)}

\item{e_s}{saturation vapor pressure (kPa)}

\item{Rabs}{Solar and thermal radiation absorbed (W)}

\item{emissivity}{of salamander skin, default emissivity=0.96}

\item{radiative_conductance}{}
}
\value{
humid operative temperature (C)
}
\description{
Calculate humid operative temperature
}
\details{
This function allows you to calculate humid operative temperature (adaptation of Campbell and Norman 1998).
}
\examples{
\dontrun{
Tb_salamander_humid(r_i=4,r_b=1,diameter=0.007,T=20,elev=500,e_a=2.5,e_s=2.3,Rabs=400, emissivity=0.96)
}

}
\author{
Eric Riddell
}
\keyword{humid}
\keyword{operative}
\keyword{temperature}
