% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TC17.Fig02.R
\name{TC17.Fig02}
\alias{TC17.Fig02}
\title{title}
\usage{
TC17.Fig02(path = file.path(getwd(), "plots"), file.name = "tc17_fig_02",
  device = "quartz", save.plot = FALSE)
}
\arguments{
\item{path}{}

\item{file.name}{}

\item{device}{}

\item{save.plot}{}
}
\description{
Details
}
\author{
Thomas Münch
}
