% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper-tc17-plots.R
\name{TC17.Fig03c}
\alias{TC17.Fig03c}
\title{Produce TC17 Figure 03c.}
\usage{
TC17.Fig03c(TR = prepareTrenchData(na.treat = TRUE)$oxy)
}
\arguments{
\item{TR}{the data set; defaults to the processed T15 oxygen isotope data.}
}
\description{
Produce TC17 Figure 03c.
}
\references{
Münch, T., et al., Constraints on post-depositional isotope modifications
in East Antarctic firn from analysing temporal changes of isotope profiles,
The Cryosphere, 11(5), 2175-2188, doi:10.5194/tc-11-2175-2017, 2017.
}
\author{
Thomas Münch
}
