% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TC17.Fig05.R
\name{TC17.Fig05}
\alias{TC17.Fig05}
\title{Produce TC17 Figure 05.}
\usage{
TC17.Fig05(dat = NULL)
}
\arguments{
\item{dat}{input data structure resulting from a call of
\code{\link{LoopParamSpace}}; if \code{NULL} (the default), the original
data presented in M端nch et al. (2017) is used for plotting which is
supplied with this package in the variable \code{\link{ParamSpace}}.}
}
\description{
This function makes all necessary calculations and plots the results for
Figure 05 shown in M端nch et al. (2017).
}
\references{
M端nch, T., et al., Constraints on post-depositional isotope modifications
in East Antarctic firn from analysing temporal changes of isotope profiles,
The Cryosphere, 11(5), 2175-2188, doi:10.5194/tc-11-2175-2017, 2017.
}
\author{
Thomas M端nch
}
