% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TC17.Fig06.R
\name{TC17.Fig06}
\alias{TC17.Fig06}
\title{Title}
\usage{
TC17.Fig06(TR, path = file.path(getwd(), "plots"),
  file.name = "tc17_fig_06", device = "quartz", save.plot = FALSE,
  mod.param)
}
\arguments{
\item{TR}{}

\item{path}{}

\item{file.name}{}

\item{device}{}

\item{mod.param}{}
}
\description{
Description
}
\author{
Thomas Münch
}
