% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TC17.Fig07.R
\name{TC17.Fig07}
\alias{TC17.Fig07}
\title{Title}
\usage{
TC17.Fig07(TR, path = file.path(getwd(), "plots"),
  file.name = "tc17_fig_07", device = "quartz", dev.size = list(h = 6, w =
  8), save.plot = FALSE, mod.param)
}
\arguments{
\item{TR}{}

\item{path}{}

\item{file.name}{}

\item{device}{}

\item{dev.size}{}

\item{mod.param}{}
}
\description{
Description
}
\author{
Thomas Münch
}
