% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{temperature conversions}
\alias{temperature conversions}
\alias{fahrenheit_to_kelvin}
\alias{kelvin_to_celsius}
\alias{celsius_to_kelvin}
\alias{fahrenheit_to_celsius}
\title{Convert Among Temperature Scales}
\usage{
fahrenheit_to_kelvin(T)

kelvin_to_celsius(T)

celsius_to_kelvin(T)

fahrenheit_to_celsius(T)
}
\arguments{
\item{T}{\code{numeric} temperature (Celsius, Fahrenheit, or Kelvin).}
}
\value{
\code{numeric} temperature (Celsius, Fahrenheit, or Kelvin).
}
\description{
Convert temperatures among Celsius, Fahrenheit, and Kelvin \insertCite{Blischak}{TrenchR}.
}
\examples{
  kelvin_to_celsius(T = 270)
  fahrenheit_to_kelvin(T = 85)
  fahrenheit_to_celsius(T = 85)
  celsius_to_kelvin(T = -10)

}
\references{
\insertAllCited{}
}
\seealso{
Other utility functions: 
\code{\link{airpressure_from_elev}()},
\code{\link{azimuth_angle}()},
\code{\link{day_of_year}()},
\code{\link{daylength}()},
\code{\link{dec_angle}()},
\code{\link{degrees_to_radians}()},
\code{\link{solar_noon}()},
\code{\link{zenith_angle}()}
}
\concept{utility functions}
