% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{thermal_radiation_absorbed}
\alias{thermal_radiation_absorbed}
\title{Estimate absorbed longwave (thermal) radiation}
\usage{
thermal_radiation_absorbed(Ta, emissivity_ground = 0.97,
  abs_longwave = 0.965)
}
\arguments{
\item{Ta}{air temperature (C)}

\item{emissivity_ground}{emmisitivity (proportion) for more soil types (Campbell and Norman 1998), default value of 0.97}

\item{abs_longwave}{absorptance (proportion) of organism to longwave radiation (Bartlett and Gates 1967, Buckley 2008), default value of 0.965}
}
\value{
thermal radiation absorbed
}
\description{
Estimate absorbed longwave (thermal) radiation
}
\details{
This function allows you to estimate longwave (thermal) radiation (W) absorbed from the sky and the ground (adaptation of Campbell and Norman 1998).
}
\examples{
\dontrun{
thermal_radiation_absorbed(Ta=20, emissivity_ground=0.97, abs_longwave=0.965)
}

}
\author{
Eric Riddell
}
\keyword{(thermal)}
\keyword{absorbed}
\keyword{longwave}
\keyword{radiation}
