% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{thermal_radiation_emitted}
\alias{thermal_radiation_emitted}
\title{Calculate emitted thermal radiation}
\usage{
thermal_radiation_emitted(emissivity = 0.96, As, psa_dir, psa_ref, Tb, Ts, Ta,
  enclosed = FALSE)
}
\arguments{
\item{emissivity}{longwave infrared emissivity of skin (proportion), 0.95 to 1 for most animals (Gates 1980)}

\item{As}{surface area  in m^2}

\item{psa_dir}{proportion surface area exposed to sky (or enclosure)}

\item{psa_ref}{proportion surface area exposed to ground}

\item{Tb}{body surface temperatue in K}

\item{Ts}{ground surface temperatue in K}

\item{Ta}{ambient air temperature in K, only required if animal is in enclosed environment}

\item{enclosed}{TRUE or FALSE}
}
\value{
emitted thermal radiation (W)
}
\description{
Calculate emitted thermal radiation
}
\details{
This function allows you to estimate thermal radiation (W) emitted by the surface of an animal. Follows Gates Biophysical Ecology and Spotila et al. 1992.
}
\examples{
\dontrun{
thermal_radiation_emitted(emissivity=0.96, As=1, psa_dir=0.4, psa_ref=0.6, Tb=303, Ts=293, Ta=298, enclosed=FALSE)
}

}
\keyword{emitted}
\keyword{radiation}
\keyword{thermal}
