% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtr_sine.R
\name{Thour.sine}
\alias{Thour.sine}
\title{Temperature across hours}
\usage{
Thour.sine(Tmx, Tmn, Hr)
}
\arguments{
\item{Tmx}{maximum daily temperature (C)}

\item{Tmn}{minimum daily temperature (C)}

\item{Hr}{hour for temperature estimate}
}
\description{
This function allows you to calculate temperature across hours.
}
\examples{
Thours.sine()
}
\keyword{Temperature}
