% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTRfunction.R
\name{Thour.sineexp}
\alias{Thour.sineexp}
\title{Estimate temperature across hours using a diurnal temperature variation function incorporating sine and exponential components}
\usage{
Thour.sineexp(Tmx, Tmn, Hr, tr, ts, alpha = 2.59, beta = 1.55,
  gamma = 2.2)
}
\arguments{
\item{Tmx}{maximum daily temperature (C)}

\item{Tmn}{minimum daily temperature (C)}

\item{Hr}{hour for temperature estimate}

\item{Ts}{time of sunrise (hour)}

\item{Tr}{time of sunset (hour)}
}
\description{
This function allows you to calculate temperature across hours using a diurnal temperature variation function incorporating sine and exponential components.
}
\examples{
Thour.sineexp()
}
\keyword{Temperature}
