% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtr_sine.R
\name{Thours.mat}
\alias{Thours.mat}
\title{Temperature across hours}
\usage{
Thours.mat(Tmat, alpha = 1.86, beta = -0.17, gamma = 2.2)
}
\arguments{
\item{Tmat}{Vector containing Tmin,Tmax, Start Time, and End Time}
}
\description{
This function allows you to calculate temperature across hours.
}
\examples{
Thours.mat()
}
\keyword{Temperature}
