% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenchR-package.R
\name{TrenchR}
\alias{TrenchR}
\title{Tools for Microclimate and Biophysical Ecology}
\description{
Tools for translating environmental change into organismal response. Microclimate models to vertically scale weather station data to organismal heights. The biophysical modeling tools include both general models for heat flows and specific models to predict body temperatures for a variety of ectothermic taxa. Additional functions model and temporally partition air and soil temperatures and solar radiation. Utility functions estimate the organismal and environmental parameters needed for biophysical ecology. 'TrenchR' focuses on relatively simple and modular functions so users can create transparent and flexible biophysical models. Many functions are derived from Gates (1980) and Campbell and Norman (1988).
}
\keyword{package}
