% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_velocity.R
\name{V_z}
\alias{V_z}
\title{Estimate ...
Credit from Porter et al 1973
z_0 is surface roughness
z_r is initial reference height
z is height to scale to}
\usage{
V_z(V_r, z_0 = 0.02, z_r = 1.54, z = 0.2)
}
\arguments{
\item{V_r}{is wind velocity at reference height.}
}
\description{
This function allows you to calculate
}
\examples{
V_z()
}
\keyword{Wind}
