% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllometricFunctions.R
\name{volume_from_length}
\alias{volume_from_length}
\title{Calculate organism volume from length}
\usage{
volume_from_length(l, taxa)
}
\arguments{
\item{l}{\code{numeric} length in meters (m). 
\cr \cr
Snout-vent length for amphibians and reptiles (excepting turtles where length is carapace length).}

\item{taxa}{\code{character} taxon of organism, current choices: \code{"lizard"}, \code{"frog"}, \code{"sphere"}.}
}
\value{
\code{numeric} volume in meters cubed (\ifelse{html}{\out{m<sup>3</sup>}}{\eqn{m^3}{ASCII}}).
}
\description{
Estimate volume (\ifelse{html}{\out{m<sup>3</sup>}}{\eqn{m^3}{ASCII}}) from length (m) for a variety of taxa following \insertCite{Mitchell1976}{TrenchR}
}
\details{
Relationships come from \itemize{
  \item Lizards: \insertCite{Norris1965;textual}{TrenchR}.
  \item Frogs: \insertCite{Tracy1972}{TrenchR}.
  \item Sphere: \insertCite{Mitchell1976}{TrenchR}.
 }
}
\examples{
  volume_from_length(l = 0.05, taxa = "lizard")
  volume_from_length(l = 0.05, taxa = "frog")
  volume_from_length(l = 0.05, taxa = "sphere")

}
\references{
\insertAllCited{}
}
\seealso{
Other allometric functions: 
\code{\link{mass_from_length}()},
\code{\link{prop_silhouette_area_shapes}()},
\code{\link{prop_silhouette_area}()},
\code{\link{sa_from_length}()},
\code{\link{sa_from_mass}()},
\code{\link{sa_from_volume}()}
}
\concept{allometric functions}
\keyword{length}
\keyword{volume}
