% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.peaks.R
\name{which.peaks}
\alias{which.peaks}
\title{Find Locations of Maxima or Minima.}
\usage{
which.peaks(x, partial = TRUE, decreasing = FALSE)
}
\arguments{
\item{x}{}

\item{partial}{}

\item{decreasing}{}
}
\value{
The indices of the maxima or minima of x.
}
\description{
Details
}
\author{
Thomas Münch
}
