% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempWindProfileFunctions.R
\name{wind_speed_profile_neutral}
\alias{wind_speed_profile_neutral}
\title{Estimate wind speed at a specific height under neutral conditions.}
\usage{
wind_speed_profile_neutral(u_r, zr, z0, z)
}
\arguments{
\item{u_r}{is wind velocity at reference height in m/s.}

\item{zr}{is initial reference height in m.}

\item{z0}{is surface roughness in m.}

\item{z}{is height to scale to in m.}
}
\value{
windspeed in m/s
}
\description{
This function allows you to calculate wind speed (m/s) at a 
specified height (m) within a boundary layer near the surface.  
The velocity profile is the neutral profile described by Sellers (1965). 
Function in equations (2) and (3) of Porter et al. (1973)
Profiles in neutral conditions.
}
\details{
Calculates wind speed at a specified height under neutral conditions
}
\examples{
\dontrun{
wind_speed_profile_neutral(u_r=0.1, zr=0.1, z0=0.2, z=0.15)
}
}
\keyword{profile}
\keyword{wind}
