% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YearMean.R
\name{YearMean}
\alias{YearMean}
\title{Annual means from binning.}
\usage{
YearMean(x, ind, na.rm = TRUE)
}
\arguments{
\item{x}{The vector for which to compute the annual means according to the
binning specified by \code{ind}.}

\item{ind}{A vector of indices specifying the binning windows for
calculating annual means.}

\item{na.rm}{a logical value indicating whether ‘NA’ values should be
stripped before the computation proceeds. Defaults to \code{TRUE}.}
}
\value{
A vector of annual means.
}
\description{
Description.
}
\author{
Thomas Münch
}
