% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZenithAngle.R
\name{zenith.rad}
\alias{zenith.rad}
\title{Calculate Zenith in radians...}
\usage{
zenith.rad(J, lat, lon, Hr)
}
\arguments{
\item{J}{is Julian day.}

\item{lat}{is latitude.}

\item{lon}{is longitude.}

\item{Hr}{is hour of the day.}
}
\description{
This function allows you to calculate zenith in radians
}
\examples{
\dontrun{
zenith.rad()
}
}
\keyword{radian}
\keyword{zenith}
