% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZenithAngle.R
\name{zenith}
\alias{zenith}
\title{Calculate Zenith angle}
\usage{
zenith(J, lat, lon, Hr)
}
\arguments{
\item{J}{is Julian day.}

\item{lat}{is latitude.}

\item{lon}{is longitude.}

\item{Hr}{is hour of the day.}
}
\description{
Calculate Zenith angle
}
\details{
This function allows you to calculate zenith
}
\examples{
\dontrun{
zenith()
}
}
\keyword{zenith}
