% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{zenith_angle}
\alias{zenith_angle}
\title{Calculate Zenith Angle}
\usage{
zenith_angle(doy, lat, lon, hour, offset = NA)
}
\arguments{
\item{doy}{\code{numeric}  day of year (1-366). (This can be obtained from standard date via the TrenchR::day_of_year() function.)}

\item{lat}{\code{numeric}  latitude (decimal degrees)}

\item{lon}{\code{numeric}  longitude (decimal degrees)}

\item{hour}{\code{numeric}  hour of the day}

\item{offset}{\code{numeric}  the number of hours to add to UTC to get local time (improves accuracy but not always necessary). Optional. Defaults to NA.}
}
\value{
\code{numeric} zenith angle (degrees)
}
\description{
calculate the zenith angle, the location of the sun as an angle (in degrees) measured from vertical \insertCite{Campbell1998}{TrenchR}.
}
\examples{
  zenith_angle(doy  = 112, 
               lat  = 47.61, 
               lon  = -122.33, 
               hour = 12)

}
\references{
\insertAllCited{}
}
\seealso{
Other utility functions: 
\code{\link{airpressure_from_elev}()},
\code{\link{azimuth_angle}()},
\code{\link{day_of_year}()},
\code{\link{daylength}()},
\code{\link{dec_angle}()},
\code{\link{degree_to_radian}()},
\code{\link{fahrenheit_to_celsius}()},
\code{\link{fahrenheit_to_kelvin}()},
\code{\link{kelvin_to_celsius}()},
\code{\link{solar_noon}()}
}
\concept{utility functions}
