% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempWindProfileFunctions.R
\name{air_temp_profile}
\alias{air_temp_profile}
\title{Estimate air temperature profile as in NicheMapR}
\usage{
air_temp_profile(T_r, u_r, zr, z0, z, T_s)
}
\arguments{
\item{T_r}{is temperature at reference height in degrees C.}

\item{u_r}{is windspeed at reference height in m/s.}

\item{zr}{is initial reference height in m.}

\item{z0}{is surface roughness in m.}

\item{z}{is height to scale to in m.}

\item{T_s}{is surface temperatures in degrees C.}
}
\value{
temperature (degrees C)
}
\description{
This function allows you to estimate temperature (C) at a specified height (m).  
Estimates a single, unsegmented temperature profile using the MICRO routine from 
NicheMapR as described in Kearney and Porter 2016.
}
\details{
Estimate temperature at a specified height
}
\examples{
\dontrun{
air_temp_profile(T_r=20, u_r=0.1, zr=0.1, z0=0.2, z=0.15, Ts=25)
}

}
\keyword{profile}
\keyword{temperature}
