% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempWindProfileFunctions.R
\name{air_temp_profile_neutral}
\alias{air_temp_profile_neutral}
\title{Estimate temperature at a specified height under neutral conditions}
\usage{
air_temp_profile_neutral(T_r, zr, z0, z, T_s)
}
\arguments{
\item{T_r}{is temperature at reference height in C.}

\item{zr}{is initial reference height in m.}

\item{z0}{is surface roughness in m.}

\item{z}{is height to scale to in m.}

\item{T_s}{is surface temperatures in C.}
}
\value{
temperature (C)
}
\description{
calculate temperature (C) at a specified height (m) within a 
boundary layer near the surface.  The velocity profile is the neutral profile described by Sellers (1965). 
Function in equations (2) and (3) of Porter et al. (1973). Source: Porter WP et al. 1973. Behavioral implications of mechanistic ecology. Oecologia 13:1-54.
}
\examples{
air_temp_profile_neutral(T_r=20, zr=0.1, z0=0.2, z=0.15, T_s=25)


}
\seealso{
Other microclimate functions: 
\code{\link{air_temp_profile_segment}()},
\code{\link{air_temp_profile}()},
\code{\link{degree_days}()},
\code{\link{direct_solar_radiation}()},
\code{\link{diurnal_radiation_variation}()},
\code{\link{diurnal_temp_variation_sineexp}()},
\code{\link{diurnal_temp_variation_sinesqrt}()},
\code{\link{diurnal_temp_variation_sine}()},
\code{\link{estimate_radiation}()},
\code{\link{monthly_solar_radiation}()},
\code{\link{partition_solar_radiation}()},
\code{\link{proportion_diffuse_solar_radiation}()},
\code{\link{surface_roughness}()},
\code{\link{wind_speed_profile_neutral}()},
\code{\link{wind_speed_profile_segment}()}
}
\concept{microclimate functions}
\keyword{profile}
\keyword{temperature}
