% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_pressure.R
\name{airpressure_elev}
\alias{airpressure_elev}
\title{Estimate air pressure in kPa (Kilo Pascal)
Credit - #http://www.engineeringtoolbox.com/air-altitude-pressure-d_462.html}
\usage{
airpressure_elev(h)
}
\arguments{
\item{h}{height in meters.}
}
\description{
This function allows you to calculate estimated air pressure (kPa) as afunction of elevation
}
\details{
Estimate air pressure (kPa) as a function of elevation
}
\examples{
airpressure_elev()
}
\keyword{air}
\keyword{pressure}
