% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{airpressure_from_elev}
\alias{airpressure_from_elev}
\title{Estimate air pressure in kPa (Kilo Pascal)}
\usage{
airpressure_from_elev(elev)
}
\arguments{
\item{elev}{elevation in meters.}
}
\value{
Air pressure in kPa
}
\description{
This function allows you to calculate estimated air pressure (kPa) as a function of elevation
}
\details{
Credit - #http://www.engineeringtoolbox.com/air-altitude-pressure-d_462.html

Estimate air pressure (kPa) as a function of elevation
}
\examples{
\dontrun{
airpressure_from_elev(1500)
}
}
\keyword{Air}
\keyword{Pressure}
