% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{airpressure_from_elev}
\alias{airpressure_from_elev}
\title{Air Pressure in kPa}
\usage{
airpressure_from_elev(elev)
}
\arguments{
\item{elev}{\code{numeric} elevation (meters).}
}
\value{
\code{numeric} air pressure (kPa).
}
\description{
Calculate estimated air pressure (kPa) as a function of elevation \insertCite{engingeeringtoolbox}{TrenchR}.
}
\examples{
  airpressure_from_elev(elev = 1500)

}
\references{
\insertAllCited{}
}
\seealso{
Other utility functions: 
\code{\link{azimuth_angle}()},
\code{\link{day_of_year}()},
\code{\link{daylength}()},
\code{\link{dec_angle}()},
\code{\link{degree_to_radian}()},
\code{\link{solar_noon}()},
\code{\link{temperature conversions}},
\code{\link{zenith_angle}()}
}
\concept{utility functions}
