% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AverageIndexBins.R
\name{AverageIndexBins}
\alias{AverageIndexBins}
\title{Means from index binning.}
\usage{
AverageIndexBins(x, ind, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector for which to compute bin means according to the
indices given in \code{ind}.}

\item{ind}{numeric vector of indices specifying the binning windows to
average over, see details.}

\item{na.rm}{a logical value indicating whether ‘NA’ values should be
stripped before the computation proceeds. Defaults to \code{TRUE}.}
}
\value{
a numeric vector of length \code{length(ind) - 1} with the mean
values of each bin with end points defined by \code{ind}.
}
\description{
Calculate bin means defined by vector index positions.
}
\details{
Averaging bins are defined consecutively from the first index position
to the position one before the second index, i.e., such that the right end
of the first bin does not overlap with the left end of the second bin, and
so on. The final bin is treated separately: if the last index coincides
with the last data point of \code{x}, this point is included in the last bin
average, else the bin is treated as the other bins.
}
\examples{
x <- rnorm(20)

# final bin includes final data point:
AverageIndexBins(x, ind = c(1, 7, 13, 17, 20))
# final index is not final data point:
AverageIndexBins(x, ind = c(1, 7, 13, 17))
}
\author{
Thomas Münch
}
