% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{conductance}
\alias{conductance}
\title{Calculate conductance}
\usage{
conductance(Ts, Ta, sa, lambda, K = 0.5, position = "flat")
}
\arguments{
\item{Ts}{Surface Temperature.}

\item{Ta}{Air Temperature.}

\item{sa}{surface area in m^2}

\item{lambda}{Mean thickness.}

\item{K}{thermal conductivity}

\item{position}{Whether the organism is lying flat or standing.}

\item{mass}{Mass in grams.}
}
\value{
conductance
}
\description{
Calculate conductance
}
\details{
This function allows you to calculate conductance of an ectotherm(Reptiles)
}
\examples{
\dontrun{
conductance(24,22,10.5,0.02,"flat","lizard")
}

}
\keyword{conductance}
