% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{conductance}
\alias{conductance}
\title{Calculate conductance}
\usage{
conductance(Ts, Ta, Tb = Ta, lambda, K = 0.5, A_contact)
}
\arguments{
\item{Ts}{Surface Temperature in Kelvin.}

\item{Ta}{Air Temperature in Kelvin.}

\item{Tb}{Body Temperature in Kelvin.}

\item{lambda}{Mean thickness(or diameter of the body) in meters.}

\item{K}{Thermal conductivity (W K^-1 m^-1 )}

\item{A_contact}{Area of contact in m^2}
}
\value{
conductance (W m^2)
}
\description{
Calculate conductance
}
\details{
This function allows you to calculate conductance of an ectotherm(Reptiles)
}
\examples{
\dontrun{
conductance(246,227,227,0.05,0.5,0.02)
}

}
\keyword{conductance}
