% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{convection}
\alias{convection}
\title{Calculate convection}
\usage{
convection(Ts, Ta, mass, species = "lizard")
}
\arguments{
\item{Ts}{Surface Temperature.}

\item{Ta}{Air Temperature.}

\item{mass}{Mass in grams.}

\item{species}{Which species.}
}
\value{
convection
}
\description{
Calculate convection
}
\details{
This function allows you to calculate convection of an ectotherm(Reptiles)
}
\examples{
\dontrun{
convection(24,22,10.5,"lizard")
}

}
\keyword{convection}
