% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{convection}
\alias{convection}
\title{Calculate convection}
\usage{
convection(Ta, Tb, h_L = 10.45, sa, proportion, ef = 1.3)
}
\arguments{
\item{Ta}{Air Temperature in Kelvin.}

\item{Tb}{Initial Body Temperature in Kelvin.}

\item{h_L}{Convective heat transfer coefficient (W m^-2 K^-1)}

\item{sa}{Surface area  in m^2}

\item{proportion}{of surface area exposed to air}

\item{ef}{is the enhancement factor, used to adjuct h_L to field condictions (using h_L approximation from Mitchell 1976).  Approximated as 1.3 by default, but see Mitchell 1976 for more relationship.}
}
\value{
convection (W)
}
\description{
Calculate convection
}
\details{
This function allows you to calculate convection. Includes enhncement factor associated for outdoor environments in Mitchell (1976).
}
\examples{
\dontrun{
convection(224,222,10.45,.0025, 0.85)
}

}
\keyword{convection}
