% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{convection}
\alias{convection}
\title{Calculate convection}
\usage{
convection(Ts, Ta, sa, K = 0.5)
}
\arguments{
\item{Ts}{Surface Temperature in Kelvin.}

\item{Ta}{Air Temperature in Kelvin.}

\item{sa}{surface area in m^2}

\item{K}{thermal conductivity (W K-1 m-1)}
}
\value{
convection
}
\description{
Calculate convection
}
\details{
This function allows you to calculate convection of an ectotherm(Reptiles)
}
\examples{
\dontrun{
convection(224,222,10.5,"lizard")
}

}
\keyword{convection}
