% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{convection}
\alias{convection}
\title{Calculate convection}
\usage{
convection(Ta, To, h_L = 10.45, sa, proportion)
}
\arguments{
\item{Ta}{Air Temperature in Kelvin.}

\item{To}{Initial Body Temperature in Kelvin.}

\item{h_L}{Convective heat transfer ceofficient (W m^-2 K^-1)}

\item{sa}{Surface area  in m^2}

\item{proportion}{exposed to air}
}
\value{
convection (W)
}
\description{
Calculate convection
}
\details{
This function allows you to calculate convection of an ectotherm(Reptiles)
}
\examples{
\dontrun{
convection(224,222,10.45,.0025, 0.85)
}

}
\keyword{convection}
