% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{convection}
\alias{convection}
\title{Calculate convection}
\usage{
convection(Ta, To, h_L = 10.45, A_air)
}
\arguments{
\item{Ta}{Air Temperature in Kelvin.}

\item{To}{Initial Body Temperature in Kelvin.}

\item{h_L}{Convective heat transfer ceofficient (W m^-2 K^-1)}

\item{A_air}{Area exposed to air m^2}
}
\value{
convection (W)
}
\description{
Calculate convection
}
\details{
This function allows you to calculate convection of an ectotherm(Reptiles)
}
\examples{
\dontrun{
convection(224,222,10.45,.05)
}

}
\keyword{convection}
