% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{day_of_year}
\alias{day_of_year}
\title{Calculate day of the passed date}
\usage{
day_of_year(day, format = "\%Y-\%m-\%d")
}
\arguments{
\item{day}{day}

\item{format}{date format("%Y-%m-%d") following "POSIXlt" conventions}
}
\value{
day number, for eg. 1 for January 1st
}
\description{
This function allows you to calculate day of year from text specifying a date
}
\details{
Calculate day of the year
}
\examples{
\dontrun{
day_of_year("2017-04-22", format= "\%Y-\%m-\%d")
}
}
\keyword{day}
