% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{daylength}
\alias{daylength}
\title{@title Calculate day length}
\usage{
daylength(lat, doy)
}
\arguments{
\item{lat}{\code{numeric} latitude in decimal degrees.}

\item{doy}{\code{numeric} day of year (1-366).}
}
\value{
\code{numeric} day length in hours.
}
\description{
Calculate daylength in hours as a function of latitude and day of year. Uses the CMB model \insertCite{Campbell1998}{TrenchR}.
}
\examples{
  daylength(lat=47.61, doy=112)

}
\references{
\insertAllCited{}
}
\seealso{
Other utility functions: 
\code{\link{airpressure_from_elev}()},
\code{\link{azimuth_angle}()},
\code{\link{dec_angle}()},
\code{\link{degree_to_radian}()},
\code{\link{fahrenheit_to_celsius}()},
\code{\link{fahrenheit_to_kelvin}()},
\code{\link{kelvin_to_celsius}()},
\code{\link{solar_noon}()},
\code{\link{zenith_angle}()}
}
\concept{utility functions}
