% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{daylength}
\alias{daylength}
\title{Calculate day length}
\usage{
daylength(lat, doy)
}
\arguments{
\item{lat}{latitude in degrees}

\item{doy}{day of year}
}
\value{
hours
}
\description{
This function allows you to calculate daylength in hours as a function of latitude and day of year. Uses the CMB model (Forsythe et al. 1995).
}
\details{
Calculate day length
}
\examples{
\dontrun{
daylength(lat=47.61, doy=112)
}
}
\keyword{Day}
\keyword{length}
