% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{daylength}
\alias{daylength}
\title{Calculate day length}
\usage{
daylength(lat, doy)
}
\arguments{
\item{lat}{latitude in degrees}

\item{doy}{day of year}
}
\value{
hours
}
\description{
This function allows you to calculate daylength in hours as a function of latitude and day of year. Uses the CMB model (Forsythe et al. 1995). Source: Campbell and Norman. 1998. An Introduction to Environmental Biophysics.
}
\details{
Calculate day length
}
\examples{
\dontrun{
daylength(lat=47.61, doy=112)
}
}
\seealso{
Other utility functions: \code{\link{airpressure_from_elev}},
  \code{\link{azimuth_angle}}, \code{\link{dec_angle}},
  \code{\link{degree_to_radian}},
  \code{\link{fahrenheit_to_celsius}},
  \code{\link{fahrenheit_to_kelvin}},
  \code{\link{kelvin_to_celsius}},
  \code{\link{solar_noon}}, \code{\link{zenith_angle}}
}
\concept{utility functions}
\keyword{day}
\keyword{length}
