% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDFunctions.R
\name{degree_days}
\alias{degree_days}
\title{Calculate degree days}
\usage{
degree_days(T_min, T_max, LDT = NA, UDT = NA, method = "single.sine")
}
\arguments{
\item{T_min}{Minimum temperature of the day (C)}

\item{T_max}{Maximum temperature of the day (C)}

\item{LDT}{lower developmental threshold (C).}

\item{UDT}{upper developmental threshold (C).}

\item{method}{type of method being used. Current choices: "single.sine","double.sine", "single.triangulation" or "double.triangulation".}
}
\value{
degree days (C)
}
\description{
This function allows you to calculate degree days using single or double sine wave and single or double triangulation approximation. Source: http://ipm.ucanr.edu/WEATHER/ddfigindex.html. Double methods assume symmetry, that is that next day's thermal minima is equal to previous day. Double Sine wave approximation of degree days from Allen JC. 1976.  A Modified Sine Wave Method for Calculating Degree Days. Environmental Entomology 5:388–396.
}
\details{
Calculate degree days
}
\examples{

degree_days(T_min=7, T_max=14, LDT=12, UDT=33, method="single.sine")
degree_days(T_min=7, T_max=14, LDT=12, UDT=33, method="single.triangulation")

}
\seealso{
Other microclimate functions: 
\code{\link{air_temp_profile_neutral}()},
\code{\link{air_temp_profile_segment}()},
\code{\link{air_temp_profile}()},
\code{\link{direct_solar_radiation}()},
\code{\link{diurnal_radiation_variation}()},
\code{\link{diurnal_temp_variation_sineexp}()},
\code{\link{diurnal_temp_variation_sinesqrt}()},
\code{\link{diurnal_temp_variation_sine}()},
\code{\link{estimate_radiation}()},
\code{\link{monthly_solar_radiation}()},
\code{\link{partition_solar_radiation}()},
\code{\link{proportion_diffuse_solar_radiation}()},
\code{\link{surface_roughness}()},
\code{\link{wind_speed_profile_neutral}()},
\code{\link{wind_speed_profile_segment}()}
}
\concept{microclimate functions}
\keyword{days}
\keyword{degree}
