% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{degree_to_radian}
\alias{degree_to_radian}
\title{Converts angle in degrees to radians}
\usage{
degree_to_radian(deg)
}
\arguments{
\item{deg}{angle in degrees}
}
\value{
angle in radians
}
\description{
This function allows you to convert angle in degrees to radians.
}
\details{
Converts angle in degrees to radians.
}
\examples{
\dontrun{
degree_to_radian(47.608)
}
}
\seealso{
Other utility functions: \code{\link{airpressure_from_elev}},
  \code{\link{azimuth_angle}}, \code{\link{daylength}},
  \code{\link{dec_angle}},
  \code{\link{fahrenheit_to_celsius}},
  \code{\link{fahrenheit_to_kelvin}},
  \code{\link{kelvin_to_celsius}},
  \code{\link{solar_noon}}, \code{\link{zenith_angle}}
}
\concept{utility functions}
\keyword{degrees}
\keyword{radians}
\keyword{to}
