% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{degree_to_radian}
\alias{degree_to_radian}
\title{Converts angle in degrees to radians}
\usage{
degree_to_radian(deg)
}
\arguments{
\item{deg}{angle in degrees}
}
\value{
angle in radians
}
\description{
This function allows you to convert angle in degrees to radians
}
\details{
Converts angle in degrees to radians
}
\examples{
\dontrun{
degree_to_radian(47.608)
}
}
\keyword{degrees}
\keyword{radians}
\keyword{to}
