% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{degrees_to_radians}
\alias{degrees_to_radians}
\title{Converts angle in degrees to radians}
\usage{
degrees_to_radians(deg)
}
\arguments{
\item{deg}{angle in degrees}
}
\value{
angle in radians
}
\description{
This function allows you to convert angle in degrees to radians
}
\details{
Converts angle in degrees to radians
}
\examples{
\dontrun{
degrees_to_radians(47.608)
}
}
\keyword{degrees}
\keyword{radians}
\keyword{to}
