% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationFunctions.R
\name{diurnal_radiation_variation}
\alias{diurnal_radiation_variation}
\title{Estimate diurnal radiation}
\usage{
diurnal_radiation_variation(doy, solrad, hour, lon, lat)
}
\arguments{
\item{doy}{is the day of year}

\item{solrad}{is solar radiation in W m^-2 per day}

\item{hour}{is hour (0-24)}

\item{lon}{longitude in degrees}

\item{lat}{latitude in degrees}
}
\value{
hourly solar radiation (W/m^2)
}
\description{
Function to estimate hourly solar radiation (W m^-2 per hour) as a function of daily global solar radiation (in W m^-2 per day).
Based on Tham et al. (2010, Estimation of hourly averaged solar irradiation: evaluation of models. Building Serv. Eng. Res. Technol. 31: 9-25) 
and Al-Rawahi et al. (2011, Prediction of Hourly Solar Radiation on Horizontal and Inclined Surfaces for Muscat/Oman. 
The Journal of Engineering Research 8:19-31).
}
\details{
Estimate hourly solar radiation (W m^-2 per hour) as a function of daily global solar radiation (in W m^-2 per day).
}
\examples{
\dontrun{
diurnal_radiation_variation(doy=112, solrad=8000, hour=12, lon=-122.33, lat=47.61)
}
}
\seealso{
Other microclimate functions: 
\code{\link{air_temp_profile_neutral}()},
\code{\link{air_temp_profile_segment}()},
\code{\link{air_temp_profile}()},
\code{\link{degree_days}()},
\code{\link{direct_solar_radiation}()},
\code{\link{diurnal_temp_variation_sineexp}()},
\code{\link{diurnal_temp_variation_sinesqrt}()},
\code{\link{diurnal_temp_variation_sine}()},
\code{\link{estimate_radiation}()},
\code{\link{monthly_solar_radiation}()},
\code{\link{partition_solar_radiation}()},
\code{\link{proportion_diffuse_solar_radiation}()},
\code{\link{surface_roughness}()},
\code{\link{wind_speed_profile_neutral}()},
\code{\link{wind_speed_profile_segment}()}
}
\concept{microclimate functions}
\keyword{diurnal}
\keyword{radiation}
