% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTRFunctions.R
\name{diurnal_temp_variation_sine}
\alias{diurnal_temp_variation_sine}
\title{Diurnal temperature across hours}
\usage{
diurnal_temp_variation_sine(T_max, T_min, t)
}
\arguments{
\item{T_max}{maximum daily temperature in degree celsius}

\item{T_min}{minimum daily temperature in degree celsius}

\item{t}{hour for temperature estimate}
}
\value{
temperature (C) at a specified hour
}
\description{
This function allows you to estimate temperature for a specified hour using the sine interpolation in Campbell and Norman (1998). Source: Campbell and Norman. 1998. An Introduction to Environmental Biophysics.
}
\details{
Uses a sine interpolation to estimate temperature across hours.
}
\examples{
\dontrun{
diurnal_temp_variation_sine(T_max=30, T_min=10, t=11)
}
}
\keyword{Temperature}
