% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTRFunctions.R
\name{diurnal_temp_variation_sine}
\alias{diurnal_temp_variation_sine}
\title{Diurnal temperature across hours
From Campbell and Norman 1998 - Uses sine interpolation}
\usage{
diurnal_temp_variation_sine(T_max, T_min, t)
}
\arguments{
\item{T_max}{maximum daily temperature in degree celsius}

\item{T_min}{minimum daily temperature in degree celsius}

\item{t}{hour for temperature estimate}
}
\description{
Diurnal temperature across hours
From Campbell and Norman 1998 - Uses sine interpolation
}
\details{
This function allows you to estimate temperature for a specified hour using the sine interpolation in Campbell and Norman (1998).
}
\examples{
\dontrun{
diurnal_temp_variation_sine(T_max=30, T_min=10, t=11)
}
}
\keyword{Temperature}
