% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTRFunctions.R
\name{diurnal_temp_variation_sine}
\alias{diurnal_temp_variation_sine}
\title{Diurnal temperature across hours
From Campbell and Norman 1998 - Uses sine interpolation}
\usage{
diurnal_temp_variation_sine(Tmx, Tmn, Hr)
}
\arguments{
\item{Tmx}{maximum daily temperature in degree celsius}

\item{Tmn}{minimum daily temperature in degree celsius}

\item{Hr}{hour for temperature estimate}
}
\description{
Diurnal temperature across hours
From Campbell and Norman 1998 - Uses sine interpolation
}
\details{
This function allows you to estimate temperature for a specified hour using the sine interpolation in Campbell and Norman (1998).
}
\examples{
\dontrun{
diurnal_temp_variation_sine()
}
}
\keyword{Temperature}
