% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTRFunctions.R
\name{diurnal_temp_variation_sine}
\alias{diurnal_temp_variation_sine}
\title{Diurnal temperature across hours}
\usage{
diurnal_temp_variation_sine(T_max, T_min, t)
}
\arguments{
\item{T_max}{maximum daily temperature (°C)}

\item{T_min}{minimum daily temperature (°C)}

\item{t}{time for temperature estimate (hour)}
}
\value{
temperature (°C) at a specified hour
}
\description{
This function allows you to estimate temperature for a specified hour using the sine interpolation in Campbell and Norman (1998). Source: Campbell and Norman. 1998. An Introduction to Environmental Biophysics.
}
\details{
Uses a sine interpolation to estimate temperature across hours.
}
\examples{
\dontrun{
diurnal_temp_variation_sine(T_max=30, T_min=10, t=11)
}
}
\seealso{
Other microclimate functions: 
\code{\link{air_temp_profile_neutral}()},
\code{\link{air_temp_profile_segment}()},
\code{\link{air_temp_profile}()},
\code{\link{degree_days}()},
\code{\link{diurnal_radiation_variation}()},
\code{\link{diurnal_temp_variation_sineexp}()},
\code{\link{diurnal_temp_variation_sinesqrt}()},
\code{\link{estimate_radiation}()},
\code{\link{monthly_solar_radiation}()},
\code{\link{partition_solar_radiation}()},
\code{\link{surface_roughness}()},
\code{\link{wind_speed_profile_neutral}()},
\code{\link{wind_speed_profile_segment}()}
}
\concept{microclimate functions}
\keyword{Temperature}
