% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTRFunctions.R
\name{diurnal_temp_variation_sineexp}
\alias{diurnal_temp_variation_sineexp}
\title{Estimate temperature across hours using a diurnal temperature variation function incorporating sine and exponential componenT_s}
\usage{
diurnal_temp_variation_sineexp(T_max, T_min, t, t_r, t_s, alpha = 2.59,
  beta = 1.55, gamma = 2.2)
}
\arguments{
\item{T_max}{maximum daily temperature (C)}

\item{T_min}{minimum daily temperature (C)}

\item{t}{hour for temperature estimate}

\item{t_r}{time of sunset (hour)}

\item{t_s}{time of sunrise (hour)}

\item{alpha}{time difference between t_x(time of maximum temperature) and noon}

\item{beta}{time difference between t_x and sunrise}

\item{gamma}{decay parameter for rate of t change from sunset to t_n(time of minimum temp)}
}
\description{
Estimate temperature across hours using a diurnal temperature variation function incorporating sine and exponential componenT_s
}
\details{
This function allows you to estimate temperature across hours using a diurnal temperature variation function incorporating sine and exponential componenT_s as in Parton and Logan (1981).
}
\examples{
\dontrun{
Thour.sineexp()
}
}
\keyword{Temperature}
