% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTRFunctions.R
\name{diurnal_temp_variation_sineexp}
\alias{diurnal_temp_variation_sineexp}
\title{Estimate temperature across hours using a diurnal temperature variation function incorporating sine and exponential components}
\usage{
diurnal_temp_variation_sineexp(Tmx, Tmn, Hr, tr, ts, alpha = 2.59,
  beta = 1.55, gamma = 2.2)
}
\arguments{
\item{Tmx}{maximum daily temperature (C)}

\item{Tmn}{minimum daily temperature (C)}

\item{Hr}{hour for temperature estimate}

\item{tr}{time of sunset (hour)}

\item{ts}{time of sunrise (hour)}

\item{alpha}{time difference between tx(time of maximum temperature) and noon}

\item{beta}{time difference between tx and sunrise}

\item{gamma}{decay parameter for rate of t change from sunset to tn(time of minimum temp)}
}
\description{
Estimate temperature across hours using a diurnal temperature variation function incorporating sine and exponential components
}
\details{
This function allows you to estimate temperature across hours using a diurnal temperature variation function incorporating sine and exponential components as in Parton and Logan (1981). We provide default values for alpha (2.59), beta (1.55), and gamma (2.2) for North Carolina from Wann 1985. See function for other parameterizations.
}
\examples{
\dontrun{
diurnal_temp_variation_sineexp(Tmx=40, Tmn=10, Hr=11, tr=6, ts=18, alpha=2.59, beta= 1.55, gamma=2.2)
}
}
\keyword{Temperature}
