% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTRFunctions.R
\name{diurnal_temp_variation_sineexp}
\alias{diurnal_temp_variation_sineexp}
\title{Estimate temperature across hours using a diurnal temperature variation function incorporating sine and exponential components}
\usage{
diurnal_temp_variation_sineexp(
  T_max,
  T_min,
  t,
  t_r,
  t_s,
  alpha = 2.59,
  beta = 1.55,
  gamma = 2.2
)
}
\arguments{
\item{T_max}{maximum daily temperature (°C)}

\item{T_min}{minimum daily temperature (°C)}

\item{t}{time for temperature estimate (hour)}

\item{t_r}{time of sunset (hour)}

\item{t_s}{time of sunrise (hour)}

\item{alpha}{time difference between t_x (time of maximum temperature) and noon (hour)}

\item{beta}{time difference between t_x and sunrise (hour)}

\item{gamma}{decay parameter for rate of t change from sunset to t_n (time of minimum temp)}
}
\value{
temperature (°C) at a specified hour
}
\description{
This function allows you to estimate temperature across hours using a diurnal temperature variation function incorporating sine and exponential components. Source: Parton WJ and Logan JA. 1981. A model for diurnal variation in soil and air temperature. Agricultural Meteorology 23: 205-216.
}
\details{
Estimate temperature across hours using a diurnal temperature variation function incorporating sine and exponential components.
}
\examples{
\dontrun{
diurnal_temp_variation_sineexp(T_max=30, T_min=10, t=11, t_r=6, t_s=18, alpha=2.59, beta= 1.55, gamma=2.2)
}
}
\seealso{
Other microclimate functions: 
\code{\link{air_temp_profile_neutral}()},
\code{\link{air_temp_profile_segment}()},
\code{\link{air_temp_profile}()},
\code{\link{degree_days}()},
\code{\link{diurnal_radiation_variation}()},
\code{\link{diurnal_temp_variation_sinesqrt}()},
\code{\link{diurnal_temp_variation_sine}()},
\code{\link{estimate_radiation}()},
\code{\link{monthly_solar_radiation}()},
\code{\link{partition_solar_radiation}()},
\code{\link{surface_roughness}()},
\code{\link{wind_speed_profile_neutral}()},
\code{\link{wind_speed_profile_segment}()}
}
\concept{microclimate functions}
\keyword{Temperature}
