% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTRFunctions.R
\name{diurnal_temp_variation_sineexp}
\alias{diurnal_temp_variation_sineexp}
\title{Estimate temperature across hours using a diurnal temperature variation function incorporating sine and exponential components}
\usage{
diurnal_temp_variation_sineexp(T_max, T_min, t, t_r, t_s, alpha = 2.59,
  beta = 1.55, gamma = 2.2)
}
\arguments{
\item{T_max}{maximum daily temperature (C)}

\item{T_min}{minimum daily temperature (C)}

\item{t}{hour for temperature estimate}

\item{t_r}{time of sunset (hour)}

\item{t_s}{time of sunrise (hour)}

\item{alpha}{time difference between t_x(time of maximum temperature) and noon}

\item{beta}{time difference between t_x and sunrise}

\item{gamma}{decay parameter for rate of t change from sunset to t_n(time of minimum temp)}
}
\value{
temperature (C) at a specified hour
}
\description{
This function allows you to estimate temperature across hours using a diurnal temperature variation function incorporating sine and exponential components. Source: Parton WJ and Logan JA. 1981. A model for diurnal variation in soil and air temperature. Agricultural Meteorology 23: 205-216.
}
\details{
Estimate temperature across hours using a diurnal temperature variation function incorporating sine and exponential components.
}
\examples{
\dontrun{
diurnal_temp_variation_sineexp(T_max=30, T_min=10, t=11, t_r=6, t_s=18, alpha=2.59, beta= 1.55, gamma=2.2)
}
}
\keyword{Temperature}
