% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTRFunctions.R
\name{diurnal_temp_variation_sineexp}
\alias{diurnal_temp_variation_sineexp}
\title{Estimate temperature across hours using a diurnal temperature variation function incorporating sine and exponential components}
\usage{
diurnal_temp_variation_sineexp(
  T_max,
  T_min,
  t,
  t_r,
  t_s,
  alpha = 2.59,
  beta = 1.55,
  gamma = 2.2
)
}
\arguments{
\item{T_max}{\code{numeric} maximum daily temperature (C)}

\item{T_min}{\code{numeric} minimum daily temperature (C)}

\item{t}{\code{numeric} time for temperature estimate (hour)}

\item{t_r}{\code{numeric} time of sunset (hour)}

\item{t_s}{\code{numeric} time of sunrise (hour)}

\item{alpha}{\code{numeric} time difference between t_x (time of maximum temperature) and noon (hour). Defaults to 2.59}

\item{beta}{\code{numeric} time difference between t_x and sunrise (hour). Defaults to 1.55}

\item{gamma}{\code{numeric} decay parameter for rate of t change from sunset to t_n (time of minimum temp). Defaults to 2.2}
}
\value{
\code{numeric} temperature (C) at a specified hour
}
\description{
estimate temperature across hours using a diurnal temperature variation function incorporating sine and exponential components. Source: \insertCite{Parton1981}{TrenchR}. Default alpha, beta, gamma values are the average of 5 North Carolina sites  \insertCite{Wann1985}{TrenchR}. Other alpha, beta, gamma parameterizations include values for Denver, Colorado from \insertCite{Parton1981}{TrenchR}: 150cm air temperature: alpha = 1.86, beta = 2.20, gamma = -0.17; 10cm air temperature: alpha = 1.52, beta = 2.00, gamma = -0.18; soil surface temperature: alpha = 0.50, beta = 1.81, gamma = 0.49; 10cm soil temperature: alpha = 0.45, beta = 2.28, gamma = 1.83.
}
\examples{
  diurnal_temp_variation_sineexp(T_max = 30, T_min = 10, t = 11, t_r = 6, t_s = 18, alpha = 2.59, beta = 1.55, gamma = 2.2)

}
\references{
\insertAllCited{}
}
\seealso{
Other microclimate functions: 
\code{\link{air_temp_profile_neutral}()},
\code{\link{air_temp_profile_segment}()},
\code{\link{air_temp_profile}()},
\code{\link{degree_days}()},
\code{\link{direct_solar_radiation}()},
\code{\link{diurnal_radiation_variation}()},
\code{\link{diurnal_temp_variation_sinesqrt}()},
\code{\link{diurnal_temp_variation_sine}()},
\code{\link{estimate_radiation}()},
\code{\link{monthly_solar_radiation}()},
\code{\link{partition_solar_radiation}()},
\code{\link{proportion_diffuse_solar_radiation}()},
\code{\link{surface_roughness}()},
\code{\link{wind_speed_profile_neutral}()},
\code{\link{wind_speed_profile_segment}()}
}
\concept{microclimate functions}
\keyword{Temperature}
