% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTRFunctions.R
\name{diurnal_temp_variation_sinesqrt}
\alias{diurnal_temp_variation_sinesqrt}
\title{Estimate Temperature Across Hours Using Sine and Square Root Functions}
\usage{
diurnal_temp_variation_sinesqrt(t, tr, ts, T_max, T_min, T_minp)
}
\arguments{
\item{t}{\code{numeric} hour or hours for temperature estimate}

\item{tr}{\code{numeric} sunrise hour (0-23)}

\item{ts}{\code{numeric} sunset hour (0-23)}

\item{T_max}{\code{numeric} maximum temperature of current day (C)}

\item{T_min}{\code{numeric} minimum temperature of current day (C)}

\item{T_minp}{\code{numeric} minimum temperature of following day (C)}
}
\value{
\code{numeric} temperature (C) at a specified hour
}
\description{
estimate temperature for a specified hour using sine and square root functions. Source: \insertCite{Cesaraccio2001}{TrenchR}
}
\details{
Estimates temperature across hours using sine and square root functions
}
\examples{
  diurnal_temp_variation_sinesqrt(t      = 8, 
                                  tr     = 6, 
                                  ts     = 18, 
                                  T_max  = 30, 
                                  T_min  = 10, 
                                  T_minp = 12)

}
\references{
\insertAllCited{}
}
\seealso{
Other microclimate functions: 
\code{\link{air_temp_profile_neutral}()},
\code{\link{air_temp_profile_segment}()},
\code{\link{air_temp_profile}()},
\code{\link{degree_days}()},
\code{\link{direct_solar_radiation}()},
\code{\link{diurnal_radiation_variation}()},
\code{\link{diurnal_temp_variation_sineexp}()},
\code{\link{diurnal_temp_variation_sine}()},
\code{\link{estimate_radiation}()},
\code{\link{monthly_solar_radiation}()},
\code{\link{partition_solar_radiation}()},
\code{\link{proportion_diffuse_solar_radiation}()},
\code{\link{surface_roughness}()},
\code{\link{wind_speed_profile_neutral}()},
\code{\link{wind_speed_profile_segment}()}
}
\concept{microclimate functions}
\keyword{Temperature}
