% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTRFunctions.R
\name{diurnal_temp_variation_sinesqrt}
\alias{diurnal_temp_variation_sinesqrt}
\title{Estimates temperature across hours
From Cesaraccio et al 2001}
\usage{
diurnal_temp_variation_sinesqrt(Hr, tr, ts, Tmx, Tmn, Tmn_p)
}
\arguments{
\item{Hr}{hour or hours for temperature estimate}

\item{tr}{sunrise hour (0-23)}

\item{ts}{sunset hour (0-23)}

\item{Tmx}{maximum temperature of current day (C)}

\item{Tmn}{minimum temperature of current day (C)}

\item{Tmn_p}{minimum temperature of following day (C)}
}
\description{
Estimates temperature across hours
From Cesaraccio et al 2001
}
\details{
This function allows you to estimate temperature for a specified hour using sine and square root functions (Cesaraccio et al 2001).
}
\examples{
\dontrun{
diurnal_temp_variation_sinesqrt( Hr=8, tr=6, ts=18, Tmx=30, Tmn=20, Tmn_p=22)
}
}
\keyword{Temperature}
