% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTRFunctions.R
\name{diurnal_temp_variation_sinesqrt}
\alias{diurnal_temp_variation_sinesqrt}
\title{Estimates temperature across hours using sine and square root functions}
\usage{
diurnal_temp_variation_sinesqrt(t, tr, ts, T_max, T_min, T_minp)
}
\arguments{
\item{t}{hour or hours for temperature estimate}

\item{tr}{sunrise hour (0-23)}

\item{ts}{sunset hour (0-23)}

\item{T_max}{maximum temperature of current day (째C)}

\item{T_min}{minimum temperature of current day (째C)}

\item{T_minp}{minimum temperature of following day (째C)}
}
\value{
temperature (째C) at a specified hour
}
\description{
This function allows you to estimate temperature for a specified hour using sine and square root functions. Source: Cesaraccio C et al. 2001. An improved model for determining degree-day values from daily temperature data. International Journal of Biometeorology 45:161-169.
}
\details{
Estimates temperature across hours using sine and square root functions
}
\examples{
\dontrun{
diurnal_temp_variation_sinesqrt(t=8, tr=6, ts=18, T_max=30, T_min=10, T_minp=12)
}
}
\seealso{
Other microclimate functions: \code{\link{air_temp_profile_neutral}},
  \code{\link{air_temp_profile_segment}},
  \code{\link{air_temp_profile}},
  \code{\link{degree_days}},
  \code{\link{diurnal_radiation_variation}},
  \code{\link{diurnal_temp_variation_sineexp}},
  \code{\link{diurnal_temp_variation_sine}},
  \code{\link{estimate_radiation}},
  \code{\link{monthly_solar_radiation}},
  \code{\link{partition_solar_radiation}},
  \code{\link{surface_roughness}},
  \code{\link{wind_speed_profile_neutral}},
  \code{\link{wind_speed_profile_segment}}
}
\concept{microclimate functions}
\keyword{Temperature}
