% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationFunctions.R
\name{estimate_radiation}
\alias{estimate_radiation}
\title{Estimate radiation (three parts - Direct, Diffused and Reflected)}
\usage{
estimate_radiation(doy, psi, tau, elev, rho = 0.7)
}
\arguments{
\item{doy}{is the day of year}

\item{psi}{is the zenith angle in radians}

\item{tau}{atmospheric transmissivity (proportion), which is ratio of global solar radiation at ground level to extra-terrestrial solar radiation}

\item{elev}{is elevation (m)}

\item{rho}{is the albedo as a proportion}
}
\value{
radiation components - direct, diffused and reflected (W/m^2)
}
\description{
Function to estimate direct, diffuse, and reflected components of solar radiation in W m^-2 using the model in Campbell & Norman (1998).
}
\details{
Function to estimate direct, diffuse, and reflected components of solar radiation in (W/m^2).
}
\examples{
\dontrun{
estimate_radiation(doy=112, psi=1, tau=0.6, elev=1500, rho=0.7)
}
}
\seealso{
Other microclimate functions: 
\code{\link{air_temp_profile_neutral}()},
\code{\link{air_temp_profile_segment}()},
\code{\link{air_temp_profile}()},
\code{\link{degree_days}()},
\code{\link{diurnal_radiation_variation}()},
\code{\link{diurnal_temp_variation_sineexp}()},
\code{\link{diurnal_temp_variation_sinesqrt}()},
\code{\link{diurnal_temp_variation_sine}()},
\code{\link{monthly_solar_radiation}()},
\code{\link{partition_solar_radiation}()},
\code{\link{surface_roughness}()},
\code{\link{wind_speed_profile_neutral}()},
\code{\link{wind_speed_profile_segment}()}
}
\concept{microclimate functions}
\keyword{radiation}
