% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationFunctions.R
\name{estimate_radiation}
\alias{estimate_radiation}
\title{Estimate radiation (three parts - Direct, Diffused and Reflected)}
\usage{
estimate_radiation(doy, psi, tau, elev, rho = 0.7)
}
\arguments{
\item{doy}{\code{numeric} the day of year (Can be calculated using TrenchR::day_of_year() function)}

\item{psi}{\code{numeric}  zenith angle (radians)}

\item{tau}{\code{numeric} atmospheric transmissivity (proportion), which is ratio of global solar radiation at ground level to extra-terrestrial solar radiation}

\item{elev}{\code{numeric} elevation (meters)}

\item{rho}{\code{numeric} albedo as a proportion (0-1). Defaults to 0.7}
}
\value{
radiation components - direct, diffused and reflected (W/m^2)
}
\description{
Estimate direct, diffuse, and reflected components of solar radiation (W m^-2) using the model in \insertCite{Campbell1998}{TrenchR}
}
\examples{
estimate_radiation(doy = 112, psi = 1, tau = 0.6, elev = 1500, rho = 0.7)

}
\references{
\insertAllCited{}
}
\seealso{
Other microclimate functions: 
\code{\link{air_temp_profile_neutral}()},
\code{\link{air_temp_profile_segment}()},
\code{\link{air_temp_profile}()},
\code{\link{degree_days}()},
\code{\link{direct_solar_radiation}()},
\code{\link{diurnal_radiation_variation}()},
\code{\link{diurnal_temp_variation_sineexp}()},
\code{\link{diurnal_temp_variation_sinesqrt}()},
\code{\link{diurnal_temp_variation_sine}()},
\code{\link{monthly_solar_radiation}()},
\code{\link{partition_solar_radiation}()},
\code{\link{proportion_diffuse_solar_radiation}()},
\code{\link{surface_roughness}()},
\code{\link{wind_speed_profile_neutral}()},
\code{\link{wind_speed_profile_segment}()}
}
\concept{microclimate functions}
\keyword{radiation}
