% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateInterProfileCorrelation.R
\name{estimateInterProfileCorrelation}
\alias{estimateInterProfileCorrelation}
\title{Correlation between profiles depending on distance}
\usage{
estimateInterProfileCorrelation(
  trench,
  profilePosition,
  distances,
  rangeTol = 0.05,
  a1 = 0
)
}
\arguments{
\item{trench}{a trench data set in form of a matrix or a data frame, where
columns represent the horizontal trench profile positions and rows
represent the vertical scale of the data; see also \code{\link{make2D}}.}

\item{profilePosition}{vector of the horizontal trench profile positions; see
also \code{\link{getSurfaceProfile}}.}

\item{distances}{vector of profile spacings for which profile pairs shall be
seeked and analysed for their correlation; must have the same units as
\code{profilePosition}.}

\item{rangeTol}{relative tolerance allowed for profile spacings; i.e. for a
given spacing d, all profile pairs separated by d * (1 +/- rangeTol) are
analysed.}

\item{a1}{horizontal autocorrelation of the trench data at lag 1 to obtain an
effective number of profiles used for the error estimates (see details).}
}
\value{
a tibble of five variables: the analysed profile spacings, the found
  number of pairs, the mean correlation between the pairs, and the standard
  deviation and standard error of the mean correlation (see details).
}
\description{
Estimate the pairwise correlation between trench profiles as a function of
the distance between the profiles.
}
\details{
For a given distance, the function searches the trench data set for all pairs
which are separated horizontally by this distance, allowing for a given
tolerance in separation distance. It is ensured that profile pairs are
counted only once. For every found pair, the correlation between the two
profiles is calculated using pairwise complete observations, and the
correlations are averaged across all pairs. Uncertainty of the mean
correlation is given by the standard deviation across the correlations from
all pairs, and by the standard error. The latter is calculated using the
number of found pairs, or the effective horizontal number of profiles
(degrees of freedom; as determined from the given horizontal autocorrelation
at lag 1), whichever number is smaller.
}
\examples{

IPC <- make2D(t13.trench2) \%>\%
  estimateInterProfileCorrelation(
    profilePosition = getSurfaceProfile(t13.trench2)$position,
    distances = c(10, 30))

}
\author{
Thomas Münch
}
