% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSNR.R
\name{estimateSNR}
\alias{estimateSNR}
\title{Trench signal-to-noise ratio}
\usage{
estimateSNR(data, distances, a1 = 0, rangeTol = 0.05, ...)
}
\arguments{
\item{data}{a trench data set following the generic structure used in the
package.}

\item{distances}{a vector of inter-profile distances at which average
pairwise profile correlations are to be estimated in order to calculate the
SNR.}

\item{a1}{horizontal autocorrelation of the trench data at lag 1, where lag 1
is measured relative to unit profile distance, used for the error estimate
(see \code{\link{estimateInterProfileCorrelation}}).}

\item{rangeTol}{relative tolerance allowed for inter-profile distances (see
\code{\link{estimateInterProfileCorrelation}}).}

\item{...}{parameters passed on to other functions; currently used for
\code{\link{make2D}} to specify the trench variable (see examples).}
}
\value{
a tibble of two variables: estimated signal-to-noise ratio
  (\code{snr}) and its standard error (\code{se}).
}
\description{
Estimate the signal-to-noise ratio (SNR) for a trench variable from the
average pairwise correlations evaluated at specified profile distances.
}
\examples{

distances <- c(10, 20, 30)
estimateSNR(t15.trench1, distances, a1 = exp(-1 / 1.53))
estimateSNR(t15.trench1, distances, a1 = exp(-1 / 1.53), var = "dxs")

}
\author{
Thomas Münch
}
