% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTrenchVariance.R
\name{estimateTrenchVariance}
\alias{estimateTrenchVariance}
\title{Horizontal and vertical variance of trench dataset}
\usage{
estimateTrenchVariance(
  data,
  .var = NULL,
  dof.h = NULL,
  dof.v = NULL,
  alpha = 1/6
)
}
\arguments{
\item{data}{either a generic trench data set, or a trench data set in form
of a numeric matrix or data frame, where columns represent the horizontal
trench profile positions and rows represent the vertical scale of the
data.}

\item{.var}{character string with the name of the trench variable for which to
compute the variances; see also \code{\link{make2D}}. Only needed when
\code{data} is passed as a generic trench data set.}

\item{dof.h}{number of effective degrees of freedom in horizontal direction.}

\item{dof.v}{number of effective degrees of freedom in vertical direction.}

\item{alpha}{probability for the confidence interval calculation;
i.e. confidence intervals are calculated for confidence levels from
\code{alpha} to \code{1 - alpha}.}
}
\value{
a tibble of four variables: the \code{direction} of the variance
  calculation as a character string ("horizontal" and "vertical"), the
  estimated variance \code{var}, and the \code{lower} and \code{upper} end of
  the confidence interval.
}
\description{
Calculate the average variance in horizontal and vertical direction of a
trench dataset. For this, the variances across each vertical and horizontal
position, respectively, of the trench dataset are estimated and then
averaged. Missing observations are silently removed from the data before
calculation. To allow for a more flexible input, the trench data can be
passed in the generic package form, or as a simple numeric matrix or data
frame where columns represent the horizontal trench profile positions and
rows represent the vertical scale of the data.
}
\examples{

estimateTrenchVariance(t15.trench1, .var = "d18O")

# different input format and some example DOFs to see the effect:
make2D(t15.trench1, var = "d18O") \%>\%
  estimateTrenchVariance(dof.h = 5, dof.v = 30) # (no realistic DOFs!)

}
\author{
Thomas Münch
}
