% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{Converts Fahrenheit to Celsius}
\usage{
fahrenheit_to_celsius(T)
}
\arguments{
\item{T}{\code{numeric} temperature (Fahrenheit)}
}
\value{
\code{numeric} temperature (Celsius)
}
\description{
Convert temperature from Fahrenheit to Celsius. Source: https://swcarpentry.github.io.
}
\examples{
  fahrenheit_to_celsius(T = 85)

}
\references{
\insertAllCited{}
}
\seealso{
Other utility functions: 
\code{\link{airpressure_from_elev}()},
\code{\link{azimuth_angle}()},
\code{\link{day_of_year}()},
\code{\link{daylength}()},
\code{\link{dec_angle}()},
\code{\link{degree_to_radian}()},
\code{\link{fahrenheit_to_kelvin}()},
\code{\link{kelvin_to_celsius}()},
\code{\link{solar_noon}()},
\code{\link{zenith_angle}()}
}
\concept{utility functions}
