% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{fahrenheit_to_kelvin}
\alias{fahrenheit_to_kelvin}
\title{Converts Fahrenheit to Kelvin}
\usage{
fahrenheit_to_kelvin(T)
}
\arguments{
\item{T}{Temperature in Fahrenheit.}
}
\value{
Temperature in Kelvin
}
\description{
This function allows you to convert temperature from Fahrenheit to Kelvin. Source: https://swcarpentry.github.io.
}
\details{
Converts Fahrenheit to Kelvin.
}
\examples{
\dontrun{
fahrenheit_to_kelvin(85)
}
}
\keyword{Fahrenheit}
\keyword{Kelvin}
