% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{fahrenheit_to_kelvin}
\alias{fahrenheit_to_kelvin}
\title{Converts Fahrenheit to Kelvin}
\usage{
fahrenheit_to_kelvin(T)
}
\arguments{
\item{T}{\code{numeric} temperature in Fahrenheit.}
}
\value{
\code{numeric}  temperature in Kelvin.
}
\description{
Convert temperature from Fahrenheit to Kelvin. Source: https://swcarpentry.github.io.
}
\examples{
  fahrenheit_to_kelvin(85)

}
\seealso{
Other utility functions: 
\code{\link{airpressure_from_elev}()},
\code{\link{azimuth_angle}()},
\code{\link{daylength}()},
\code{\link{dec_angle}()},
\code{\link{degree_to_radian}()},
\code{\link{fahrenheit_to_celsius}()},
\code{\link{kelvin_to_celsius}()},
\code{\link{solar_noon}()},
\code{\link{zenith_angle}()}
}
\concept{utility functions}
