% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEffectiveTrenchDOF.R
\name{getEffectiveTrenchDOF}
\alias{getEffectiveTrenchDOF}
\title{Effective degrees of freedom of autocorrelated trench profiles}
\usage{
getEffectiveTrenchDOF(
  lambda = NULL,
  a1 = NULL,
  positions = NULL,
  N = length(positions),
  delta = 1
)
}
\arguments{
\item{lambda}{horizontal decorrelation length of the assumed AR1 process
measured in the same units as \code{positions}; if specified used to
calculate the corresponding lag-1 autocorrelation via \code{a1 = exp(-1 /
lambda)}.}

\item{a1}{horizontal autocorrelation of the trench data at lag 1, where lag 1
is measured relative to unit profile distance; only used if \code{lambda}
is not specified, and specifying both \code{lambda} and \code{a1} gives an
error.}

\item{positions}{a vector of horizontal profile positions.}

\item{N}{integer number of profiles; if \code{positions} is not specified,
one can specify the profile set via the number of profiles and a constant
inter-profile distance \code{delta}.}

\item{delta}{inter-profile distance; if \code{positions} is not specified,
one can specify the profile set via the number of profiles \code{N} and a
constant inter-profile distance.}
}
\value{
number of effective degrees of freedom
}
\description{
Calculate the effective degrees of freedom for a set of trench profiles,
either given by the number of profiles and their (constant) inter-profile
distance, or given by their specific horizontal profile positions. This
calculation assumes the horizontal data variations along the trench to follow
a first-order autoregressive (AR1) process.
}
\details{
The effective degrees of freedom, `N_eff`, returned by this function are
the inverse of the quantity `sigma^{*2}_{i}` in Münch et al. (2016),
Eq. (A12): `sigma^{*2}_{i} = 1 / N_eff`; see also Eq. (1) in Münch et
al. (2017).
}
\examples{

getEffectiveTrenchDOF(a1 = 0.5, N = 10, delta = 2)
# is equivalent:
getEffectiveTrenchDOF(a1 = 0.5, positions = seq(0, 18, 2))

# alternatively, one can specify a decorrelation length directly;
# above a1 (measured at unit distace) corresponds to lambda = -1 / log(a1):
getEffectiveTrenchDOF(lambda = -1 / log(0.5), N = 10, delta = 2)

# for zero autocorrelation Neff = N:
getEffectiveTrenchDOF(a1 = 0, N = 10, delta = 2) # or
getEffectiveTrenchDOF(lambda = 0, N = 10, delta = 2)

}
\references{
Münch, T., et al., Regional climate signal vs. local noise: a two-dimensional
view of water isotopes in Antarctic firn at Kohnen Station, Dronning Maud
Land, Clim. Past, 12(7), 1565-1581, doi:10.5194/cp-12-1565-2016, 2016.

Münch, T., et al., Constraints on post-depositional isotope modifications
in East Antarctic firn from analysing temporal changes of isotope profiles,
The Cryosphere, 11(5), 2175-2188, doi:10.5194/tc-11-2175-2017, 2017.
}
\author{
Thomas Münch
}
