% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{Grashof_number}
\alias{Grashof_number}
\title{Calculate Grashof Number}
\usage{
Grashof_number(Ta, Tg, D, nu)
}
\arguments{
\item{Ta}{Air temperature (C).}

\item{Tg}{Ground (surface) temperature (C).}

\item{D}{is characteristic dimension (e.g., body diameter) (m)}

\item{nu}{is the kinematic viscosity, ratio of dynamic viscosity to density of the fluid (m2 s-1), can calculate from DRYAIR or WETAIR}
}
\value{
Grashof number
}
\description{
Calculate Grashof Number
}
\details{
This function allows you to estimate the Grashof Number, which describes the abilty of a parcel of fluid warmer or colder than the surrounding fluid to rise against or fall with the attractive force of gravity (Gates 1980 Biophysical Ecology). Ratio of a buoyant force times an inertial force to the square of a viscous force.
}
\examples{
\dontrun{
Grashof_number(Ta=30, Tg=35, D=0.001, nu=1.2)
}

}
\keyword{Grashof}
\keyword{number}
