% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{heat_transfer_coefficient}
\alias{heat_transfer_coefficient}
\title{Calculate heat transfer coefficient (based on Mitchell 1976)
(Uses Table 1 which is Convective Heat Trasfer Relations to Animal Shapes)}
\usage{
heat_transfer_coefficient(A_v, length, taxa = "lizard")
}
\arguments{
\item{A_v}{Air velocity m/s.}

\item{length}{Characteristic dimension (trunk diameter or SVL) in meter.}

\item{taxa}{Which class of organism, current choice: lizard,frog or general 'sphere'}
}
\value{
heat transfer coefficient(W m^-2 K^-1)
}
\description{
Calculate heat transfer coefficient (based on Mitchell 1976)
(Uses Table 1 which is Convective Heat Trasfer Relations to Animal Shapes)
}
\details{
This function allows you get heat transfer coefficient for various taxa(based Mitchell 1976)
}
\examples{
\dontrun{
heat_transfer_coefficient(3,.05, "lizard")
}

}
\keyword{coefficient}
\keyword{heat}
\keyword{lizard}
\keyword{transfer}
