% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{heat_transfer_coefficient}
\alias{heat_transfer_coefficient}
\title{Calculate heat transfer coefficient (based on Mitchell 1976)
(Uses Table I: Convective Heat Transfer Relations for Animal Shapes)}
\usage{
heat_transfer_coefficient(V, D, K, nu, taxa = "cylinder")
}
\arguments{
\item{V}{Air velocity m/s.}

\item{D}{Characteristic dimension (e.g., diameter or snout-vent length) in meters.}

\item{K}{Thermal conductivity of air, W m^-1 K^-1, can calculate using DRYAIR or WETAIR in NicheMapR}

\item{nu}{Kinematic Viscocity of air, m^2 s^-1, can calculate using DRYAIR or WETAIR in NicheMapR}

\item{taxa}{Which class of organism, current choices: sphere,cylinder,frog,lizard_surface,lizard_elevated,flyinginsect,spider}
}
\value{
heat transfer coefficient, H_L (W m^-2 K^-1)
}
\description{
Calculate heat transfer coefficient (based on Mitchell 1976)
(Uses Table I: Convective Heat Transfer Relations for Animal Shapes)
}
\details{
This function allows you estimate the heat transfer coefficient for various taxa (based on Mitchell 1976).  Based on empirical measurements.
}
\examples{
\dontrun{
heat_transfer_coefficient(V=0.5,D=0.05,K= 25.7 * 10^(-3),nu= 15.3 * 10^(-6), "cylinder")
}

}
\keyword{coefficient}
\keyword{heat}
\keyword{transfer}
