% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{heat_transfer_coefficient_approximation}
\alias{heat_transfer_coefficient_approximation}
\title{Calculate heat transfer coefficient using a sphereical approximation (based on Mitchell 1976)
(Uses Table III: Convective Heat Transfer Relations for Animal Shapes)}
\usage{
heat_transfer_coefficient_approximation(V, D, K, nu, taxa = "sphere")
}
\arguments{
\item{D}{Characteristic dimension (e.g., diameter or snout-vent length) in meters.}

\item{K}{Thermal conductivity of air, W m^-1 K^-1, can calculate using DRYAIR or WETAIR in NicheMapR}

\item{nu}{Kinematic Viscocity of air, m^2 s^-1, can calculate using DRYAIR or WETAIR in NicheMapR}

\item{taxa}{Which class of organism, current choices: sphere,frog,lizard,flyinginsect,spider}

\item{u}{Air velocity m/s.}
}
\value{
heat transfer coefficient, H_L (W m^-2 K^-1)
}
\description{
Calculate heat transfer coefficient using a sphereical approximation (based on Mitchell 1976)
(Uses Table III: Convective Heat Transfer Relations for Animal Shapes)
}
\details{
This function allows you estimate the heat transfer coefficient for various taxa (based on Mitchell 1976).  Approximates forced convective heat transfer for animal shapes using the convective relationship for a sphere.
}
\examples{
\dontrun{
heat_transfer_coefficient_approximation(V=3,D=0.05,K= 25.7 * 10^(-3),nu= 15.3 * 10^(-6), "sphere")
}

}
\keyword{coefficient}
\keyword{heat}
\keyword{transfer}
