% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{heat_transfer_coefficient_lizard}
\alias{heat_transfer_coefficient_lizard}
\title{Calculate heat transfer coefficient for lizard}
\usage{
heat_transfer_coefficient_lizard(A_v, orientation = "parallel")
}
\arguments{
\item{A_v}{Air velocity m/s.}

\item{orientation}{Orientation with respect to air flow}
}
\value{
heat transfer coefficient(W m^-2 K^-1)
}
\description{
Calculate heat transfer coefficient for lizard
}
\details{
This function allows you get heat transfer coefficient for Lizard(Based on Porter et al. 1973)
}
\examples{
\dontrun{
heat_transfer_coefficient_lizard(3,"parallel")
}

}
\keyword{coefficient}
\keyword{heat}
\keyword{lizard}
\keyword{transfer}
