% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{heat_transfer_coefficient_lizard}
\alias{heat_transfer_coefficient_lizard}
\title{Calculate heat transfer coefficient for lizard}
\usage{
heat_transfer_coefficient_lizard(A_v, taxa = "lizard")
}
\arguments{
\item{A_v}{Air velocity m/s.}

\item{taxa}{Which class of organism, current choice: lizard}
}
\value{
heat transfer coefficient(W m^-2 K^-1)
}
\description{
Calculate heat transfer coefficient for lizard
}
\details{
This function allows you get heat transfer coefficient for Lizard(Based on Porter et al. 1973)
}
\examples{
\dontrun{
heat_transfer_coefficient_lizard(3,"lizard")
}

}
\keyword{coefficient}
\keyword{heat}
\keyword{transfer}
