% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{heat_transfer_coefficient_simple}
\alias{heat_transfer_coefficient_simple}
\title{Calculate heat transfer coefficient (based on Mitchell 1976 in Spotila 1992)}
\usage{
heat_transfer_coefficient_simple(V, D)
}
\arguments{
\item{V}{Air velocity m/s.}

\item{D}{Characteristic dimension (e.g., diameter or snout-vent length) in meters.}
}
\value{
heat transfer coefficient, H_L (W m^-2 K^-1)
}
\description{
Calculate heat transfer coefficient (based on Mitchell 1976 in Spotila 1992)
}
\details{
This function allows you estimate the heat transfer coefficient for various taxa (based on Mitchell 1976 and using relationship in Spotila et al 1992).
}
\examples{
\dontrun{
heat_transfer_coefficient_simple(V=0.5,D=0.05)
}

}
\keyword{coefficient}
\keyword{heat}
\keyword{transfer}
