% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{heat_transfer_coefficient_simple}
\alias{heat_transfer_coefficient_simple}
\title{Calculate heat transfer coefficient (based on Mitchell 1976 in Spotila 1992)}
\usage{
heat_transfer_coefficient_simple(V, D)
}
\arguments{
\item{V}{Air velocity m/s.}

\item{D}{Characteristic dimension (e.g., diameter or snout-vent length) in meters.}
}
\value{
heat transfer coefficient, H_L (W m^-2 K^-1)
}
\description{
Calculate heat transfer coefficient (based on Mitchell 1976 in Spotila 1992)
}
\details{
This function allows you to estimate the heat transfer coefficient for various taxa (based on Mitchell 1976 and using relationship in Spotila et al 1992). References: Spotila et al. 1992. Biophysics of Heat and Mass Transfer. In Feder and Burggren. Environmental Physiology of the Amphibians; Mitchell. 1976. Heat transfer from spheres and other animal forms. Biophysical Journal 16(6): 561–569.
}
\examples{
\dontrun{
heat_transfer_coefficient_simple(V=0.5,D=0.05)
}

}
\seealso{
Other biophysical models: \code{\link{Free_or_forced_convection}},
  \code{\link{Grashof_number_Gates}},
  \code{\link{Grashof_number}}, \code{\link{Nu_from_Gr}},
  \code{\link{Nu_from_Re}}, \code{\link{Nusselt_number}},
  \code{\link{Prandtl_number}},
  \code{\link{Qconduction_animal}},
  \code{\link{Qconduction_substrate}},
  \code{\link{Qconvection}},
  \code{\link{Qemitted_thermal_radiation}},
  \code{\link{Qevaporation}},
  \code{\link{Qmetabolism_from_mass_temp}},
  \code{\link{Qmetabolism_from_mass}},
  \code{\link{Qnet_Gates}},
  \code{\link{Qradiation_absorbed}},
  \code{\link{Qthermal_radiation_absorbed}},
  \code{\link{Reynolds_number}},
  \code{\link{Tb_CampbellNorman}}, \code{\link{Tb_Fei}},
  \code{\link{Tb_Gates}}, \code{\link{Tb_butterfly}},
  \code{\link{Tb_grasshopper}}, \code{\link{Tb_limpet}},
  \code{\link{Tb_lizard}}, \code{\link{Tb_mussel}},
  \code{\link{Tb_salamander_humid}}, \code{\link{Tsoil}},
  \code{\link{actual_vapor_pressure}},
  \code{\link{boundary_layer_resistance}},
  \code{\link{external_resistance_to_water_vapor_transfer}},
  \code{\link{heat_transfer_coefficient_approximation}},
  \code{\link{heat_transfer_coefficient}},
  \code{\link{saturation_vapor_pressure}},
  \code{\link{saturation_water_vapor_pressure}}
}
\concept{biophysical models}
\keyword{coefficient}
\keyword{heat}
\keyword{transfer}
