% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{kelvin_to_celsius}
\alias{kelvin_to_celsius}
\title{Converts Kelvin to Celsius}
\usage{
kelvin_to_celsius(T)
}
\arguments{
\item{T}{Temperature in Fahrenheit.}
}
\value{
Temperature in Celsius
}
\description{
(credit  https://swcarpentry.github.io)
}
\details{
This function allows you to convert temperature from Kelvin to Celsius.
}
\examples{
\dontrun{
kelvin_to_celsius(270)
}
}
\keyword{Celsius}
\keyword{Kelvin}
