% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{kelvin_to_celsius}
\alias{kelvin_to_celsius}
\title{Converts Kelvin to Celsius}
\usage{
kelvin_to_celsius(T)
}
\arguments{
\item{T}{Temperature in Kelvin.}
}
\value{
Temperature in Celsius
}
\description{
This function allows you to convert temperature from Kelvin to Celsius. Source: https://swcarpentry.github.io.
}
\details{
Converts Kelvin to Celsius.
}
\examples{
\dontrun{
kelvin_to_celsius(270)
}
}
\seealso{
Other utility functions: \code{\link{airpressure_from_elev}},
  \code{\link{azimuth_angle}}, \code{\link{daylength}},
  \code{\link{dec_angle}}, \code{\link{degree_to_radian}},
  \code{\link{fahrenheit_to_celsius}},
  \code{\link{fahrenheit_to_kelvin}},
  \code{\link{solar_noon}}, \code{\link{zenith_angle}}
}
\concept{utility functions}
\keyword{Celsius}
\keyword{Kelvin}
