% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoopParamSpace.R
\name{LoopParamSpace}
\alias{LoopParamSpace}
\title{Loop over parameter space to find minimum RMSD.}
\usage{
LoopParamSpace(TR, advSpace = seq(40, 60, 0.5)/TR$HiRes, sigmaSpace = seq(0,
  8, 0.1), densfSpace = seq(0, 10, 0.1))
}
\arguments{
\item{TR}{}

\item{advSpace}{}

\item{sigmaSpace}{}

\item{densfSpace}{}
}
\value{
value.
}
\description{
Description.
}
\details{
Details.
}
\author{
Thomas Münch
}
