% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTC17Figures.R
\name{MakeTC17Figures}
\alias{MakeTC17Figures}
\title{Produce TC2017 Figures.}
\usage{
MakeTC17Figures(save = FALSE, path = file.path(getwd(), "plots"),
  height = 6, width = 8)
}
\arguments{
\item{save}{if \code{TRUE}, the plots are saved as files in the
folder specified by \code{path}. Defaults to \code{FALSE} which results in
on-screen display of the plots.}

\item{path}{path to the directory in which to save the plots (for
\code{save = TRUE}). Defaults to \code{plots} in the current working
directory. If this folder does not exist, it is attempted to create it
with a warning (see also \code{\link{OpenDevice}}).}

\item{height}{height of the plotting area in inches. Default '6'.}

\item{width}{width of the plotting area in inches. Default '8'.}
}
\description{
This function is a wrapper to produce all figures in Münch et al. (2017)
using the \code{quartz} graphics device for Mac OS X systems.
}
\references{
Münch, T., et al., Constraints on post-depositional isotope modifications
in East Antarctic firn from analysing temporal changes of isotope profiles,
The Cryosphere, doi:10.5194/tc-11-2175-2017, 2017.
}
\author{
Thomas Münch
}
