% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllometricFunctions.R
\name{mass_from_length}
\alias{mass_from_length}
\title{Calculate mass from length}
\usage{
mass_from_length(l, taxa)
}
\arguments{
\item{l}{Length in meters (m), length is snout-vent length for amphibians and reptiles (excepting turtles where length is carapace length).}

\item{taxa}{Class of organism, current choices: \code{"insect"}, \code{"lizard"}, \code{"salamander"}, \code{"frog"}, \code{"snake"}, \code{"turtle"}.}
}
\value{
Mass in grams (g)
}
\description{
Estimate mass (g) from length (m) for a variety of taxa.
}
\details{
Relationships come from 
 \itemize{
   \item{insect: Sample et al. (1993)}
   \item{lizard: Meiri (2010)}
   \item{salamander: Pough (1980)}
   \item{frog: Pough (1980)}
   \item{snake: Pough (1980)}
   \item{turtle: Pough (1980)} 
 }
}
\examples{
 mass_from_length(l = 0.04, taxa = "insect")
 mass_from_length(l = 0.04, taxa = "lizard")
 mass_from_length(l = 0.04, taxa = "salamander")
 mass_from_length(l = 0.04, taxa = "frog")
 mass_from_length(l = 0.04, taxa = "snake")
 mass_from_length(l = 0.04, taxa = "turtle")

}
\references{
Meiri S. 2010. Length-weight allometries in lizards. Journal of Zoology 281: 218-226. 
 
 Pough. 1980. The Advantages of Ectothermy for Tetrapods. The American Naturalist 115: 92-112.
 
 Sample BE, Cooper RJ, Greer RD, and Whitmore RC. 1993. Estimation of insect biomass by length and width. The American Midland Naturalist 129:234-240.
}
\seealso{
Other allometric functions: 
\code{\link{prop_silhouette_area_shapes}()},
\code{\link{prop_silhouette_area}()},
\code{\link{sa_from_length}()},
\code{\link{sa_from_mass}()},
\code{\link{sa_from_volume}()},
\code{\link{volume_from_length}()}
}
\concept{allometric functions}
\keyword{length}
\keyword{mass}
