% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllometricFunctions.R
\name{mass_from_length}
\alias{mass_from_length}
\title{Calculate mass from length}
\usage{
mass_from_length(l, taxa)
}
\arguments{
\item{l}{Length in m, length is snout-vent length for amphibians and reptiles (excepting turtles where length is carapace length).}

\item{taxa}{Which class of organism, current choices: insect, lizard, salamander, frog, snake, turtle}
}
\value{
mass (g)
}
\description{
Calculate mass from length
}
\details{
This function allows you to estimate mass (g) from length (m) for a variety of taxa
}
\examples{
 \dontrun{
 mass_from_length(0.04,"insect")
}

}
\keyword{length}
\keyword{mass}
