% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinorTick.R
\name{MinorTick}
\alias{MinorTick}
\title{Make aesthetic minor ticks.}
\usage{
MinorTick(nx = 2, ny = 2, tick.ratio = 0.5, side = 1,
  col = "black")
}
\arguments{
\item{nx}{number of intervals in which to divide the area between major tick
marks on the X-axis. Set to 1 to suppress minor tick marks.}

\item{ny}{same as ‘nx’ but for the Y-axis.}

\item{tick.ratio}{ratio of lengths of minor tick marks to major tick
marks. The length of major tick marks is retrieved from ‘par("tck")’.}

\item{side}{an integer specifying on which axis the minor ticks are to be
drawn on. The axis is denoted as follows: 1=below, 2=left, 3=above and
4=right.}

\item{col}{color of the minor ticks.}
}
\description{
This is a modified version of the function \code{minor.tick} from the
package \code{Hmisc}. Main difference is preventing overplotting of existing
major ticks by the new minor ticks and allowing to use colored minor ticks.
}
\author{
Frank Harrell, Earl Bellinger, Viktor Horvath, Thomas Münch
}
