% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinorTick.R
\name{MinorTick}
\alias{MinorTick}
\title{Make aesthetic minor ticks.}
\usage{
MinorTick(nx = 2, ny = 2, tick.ratio = 0.5, side = 1, col = "black")
}
\arguments{
\item{nx}{}

\item{ny}{}

\item{tick.ratio}{}

\item{side}{}

\item{col}{}
}
\description{
Description.
}
\author{
Thomas Münch
}
