% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModifyT13.R
\name{ModifyT13}
\alias{ModifyT13}
\title{Model T13 temporal changes.}
\usage{
ModifyT13(TR = prepareTrenchData(na.treat = TRUE)$oxy, SIGMA = NULL,
  STRETCH = NULL, ADV = NULL)
}
\arguments{
\item{TR}{}
}
\value{
A list.
}
\description{
Modify the T13 mean isotope profile. (Details)
}
\author{
Thomas Münch
}
