% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModifyT13.R
\name{ModifyT13}
\alias{ModifyT13}
\title{Title}
\usage{
ModifyT13(TR = prepareTrenchData(na.treat = TRUE)$oxy, mod.param)
}
\arguments{
\item{TR}{}

\item{mod.param}{}
}
\description{
Details
}
\author{
Thomas Münch
}
