% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationFunctions.R
\name{monthly_solar_radiation}
\alias{monthly_solar_radiation}
\title{Estimate monthly solar radiation}
\usage{
monthly_solar_radiation(lat, lon, doy, elev, Temp, Hr, P)
}
\arguments{
\item{lat}{latitude in degrees}

\item{lon}{longitude in degrees}

\item{doy}{is the day of year}

\item{elev}{is elevation in m}

\item{Temp}{is mean monthly temp in degree Celcius}

\item{Hr}{is mean month relative humidity}

\item{P}{is total monthly precipitation (mm)}
}
\description{
Function to estimate average monthly solar radiation (W m^-2 per day) using basic topographic and climatic information for input. Based on Nikolov and Zeller. 1992. A solar radiation algorithm for ecosystem dynamic models. Ecological modelling 61: 149-168.
}
\details{
Estimate average monthly solar radiation (W m^-2 per day) using basic topographic and climatic information for input.
}
\examples{
\dontrun{
monthly_solar_radiation(lat=47.61,lon=-122.33,doy=112,elev=1500,Temp=15,Hr=50,P=50)
}
}
\keyword{monthly}
\keyword{radiation}
\keyword{solar}
