% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyLegend.R
\name{MyLegend}
\alias{MyLegend}
\title{Produce a custom legend.}
\usage{
MyLegend(x, y = NULL, legend, fill = NULL, col = par("col"),
  border = "black", lty, lwd, pch, angle = 45, density = NULL,
  bty = "o", bg = par("bg"), box.lwd = par("lwd"), box.lty = par("lty"),
  box.col = par("fg"), pt.bg = NA, cex = 1, pt.cex = cex,
  pt.lwd = lwd, xjust = 0, yjust = 1, x.intersp = 1, y.intersp = 1,
  adj = c(0, 0.5), text.width = NULL, text.col = par("col"),
  text.font = NULL, merge = do.lines && has.pch, trace = FALSE,
  plot = TRUE, ncol = 1, horiz = FALSE, title = NULL, inset = 0, xpd,
  title.col = text.col, title.adj = 0.5, seg.len = 2, end.pch = FALSE,
  pch.xoff = 0.2)
}
\arguments{
\item{end.pch}{The character symbol for the legend end points.}

\item{pch.xoff}{Horizontal offset of the legend end points.}
}
\description{
Description.
}
\author{
Thomas Münch
}
