% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenDevice.R
\name{OpenDevice}
\alias{OpenDevice}
\title{Open a graphics device.}
\usage{
OpenDevice(device = "quartz", path, file.name, type = "pdf", height = 6,
  width = 8, save.plot = FALSE)
}
\arguments{
\item{device}{Character string specifying the plotting device; defaults to
"quartz" which is suitable for macOS systems. Currently, no other devices
are implemented.}

\item{path}{The path specifying where to save the plot.}

\item{file.name}{The name of the file under which to save the plot.}

\item{type}{The type of output to use.}

\item{height}{The height of the graphics device; defaults to 6 inches.}

\item{width}{The width of the graphics device; defaults to 8 inches.}

\item{save.plot}{if \code{TRUE} a file stream is opened according to
\code{file.path} to save the plot as a pdf file. Defaults to \code{FALSE}
which simply opens a device window. Setting \code{save.plot} to \code{TRUE}
without specying a \code{file.path} results in an error.}
}
\description{
This is a wrapper function to open a graphics device for plotting, or a file
stream to save the plot into a pdf file.
}
\author{
Thomas Münch
}
