% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllometricFunctions.R
\name{prop_silhouette_area}
\alias{prop_silhouette_area}
\title{Calculate silhouette area}
\usage{
prop_silhouette_area(z, taxa, raz = 0, posture = "prostrate")
}
\arguments{
\item{z}{\code{numeric} zenith angle in degrees between \code{0} and \code{360}.}

\item{taxa}{\code{character} Organism taxon. Current choices are \code{"lizard"}, \code{"frog"}, and \code{"grasshopper"}.}

\item{raz}{\code{numeric} relative solar azimuth angle (in degrees). Required if \code{taxa = "lizard"}. Options currently include \code{0} (in front), \code{90} (to side), and \code{180} (behind) degrees.
\cr \cr 
This is the horizontal angle of the sun relative to the head and frontal plane of the lizard.}

\item{posture}{\code{character} value describing posture. Required if \code{taxa = "lizard"}. Options include \code{"prostrate"} (default) and \code{"elevated"}.}
}
\value{
\code{numeric} silhouette area as a proportion.
}
\description{
Estimate the projected (silhouette) area as a portion of the surface area of the organism. Estimates the projected area as a function of zenith angle.
}
\details{
Relationships come from \itemize{
  \item Lizards: \insertCite{Muth1977;textual}{TrenchR}.
  \item Frogs: \insertCite{Tracy1976}{TrenchR}.
  \item Grasshoppers: \insertCite{Anderson1979}{TrenchR}.
 }
}
\examples{
  prop_silhouette_area(z = 60, taxa= "frog")
  prop_silhouette_area(z = 60, taxa= "grasshopper")
  prop_silhouette_area(z = 60, taxa= "lizard", posture = "prostrate", raz = 90)
  prop_silhouette_area(z = 60, taxa= "lizard", posture = "elevated", raz = 180)

}
\references{
\insertAllCited{}
}
\seealso{
Other allometric functions: 
\code{\link{mass_from_length}()},
\code{\link{prop_silhouette_area_shapes}()},
\code{\link{sa_from_length}()},
\code{\link{sa_from_mass}()},
\code{\link{sa_from_volume}()},
\code{\link{volume_from_length}()}
}
\concept{allometric functions}
\keyword{area}
\keyword{silhouette}
