% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllometricFunctions.R
\name{prop_silhouette_area}
\alias{prop_silhouette_area}
\title{Calculate silhouette area}
\usage{
prop_silhouette_area(z, taxa, raz = 0, posture = "prostrate")
}
\arguments{
\item{z}{zenith angle in degrees}

\item{taxa}{Which class of organism, current choices: frog, lizard, grasshopper}

\item{raz}{if lizard, relative solar azimuth angle (degrees), the horizontal angle of the sun relative to the head and frontal plane of the lizard. Options are 0 (in front), 90 (to side), and 180 (behind) degrees.}

\item{posture}{if lizard, indicate posture as "prostrate" or "elevated"}
}
\value{
silhouette area as a proportion
}
\description{
Calculate silhouette area
}
\details{
This function allows you to estimate the projected (silhouette) area as a portion of the surface area of the organism. Estimates the projected area as a function of zenith angle.
}
\examples{
\dontrun{
prop_silhouette_area(z=60, taxa= "frog")
}

}
\seealso{
Other allometric functions: \code{\link{mass_from_length}},
  \code{\link{prop_silhouette_area_shapes}},
  \code{\link{sa_from_length}}, \code{\link{sa_from_mass}},
  \code{\link{sa_from_volume}},
  \code{\link{volume_from_length}}
}
\concept{allometric functions}
\keyword{area}
\keyword{silhouette}
