% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{Qemitted_thermal_radiation}
\alias{Qemitted_thermal_radiation}
\title{Calculate emitted thermal radiation}
\usage{
Qemitted_thermal_radiation(epsilon = 0.96, A, psa_dir, psa_ref, T_b, T_g,
  T_a, enclosed = FALSE)
}
\arguments{
\item{epsilon}{longwave infrared emissivity of skin (proportion), 0.95 to 1 for most animals (Gates 1980)}

\item{A}{surface area  in m^2}

\item{psa_dir}{proportion surface area exposed to sky (or enclosure)}

\item{psa_ref}{proportion surface area exposed to ground}

\item{T_b}{body surface temperatue in K}

\item{T_g}{ground surface temperatue in K}

\item{T_a}{ambient air temperature in K, only required if animal is in enclosed environment}

\item{enclosed}{TRUE or FALSE}
}
\value{
emitted thermal radiation, Qemit (W)
}
\description{
Calculate emitted thermal radiation
}
\details{
This function allows you to estimate thermal radiation (W) emitted by the surface of an animal. Follows Gates' Biophysical Ecology and Spotila et al. 1992.
}
\examples{
\dontrun{
Qemitted_thermal_radiation(epsilon=0.96, A=1, psa_dir=0.4, psa_ref=0.6, T_b=303, T_g=293, T_a=298, enclosed=FALSE)
}

}
\keyword{emitted}
\keyword{radiation}
\keyword{thermal}
