% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{Qevaporation}
\alias{Qevaporation}
\title{Calculate heat loss associated with evaporative water loss}
\usage{
Qevaporation(A, T_b, taxa, rho_s = NA, rho_a = NA, h = NA, H = NA,
  r_i = NA)
}
\arguments{
\item{A}{surface area  in m^2}

\item{T_b}{body temperatue in K}

\item{taxa}{taxa current choices: lizard, amphibian_wetskin (fully wet skin), amphibian (not fully wet skin)}

\item{rho_s}{saturation water vapor density at skin surface (kg/m^3) (needed if amphibian)}

\item{rho_a}{saturation water vapor density in ambient air (kg/m^3) (needed if amphibian)}

\item{h}{relative humidity (0-1) (needed if amphibian)}

\item{H}{convective heat transfer coefficient (W m^-2 K^-1) (needed if amphibian)}

\item{r_i}{internal (cutaneous) resistance to vapor transport (s/m) (needed if amphibian)}

\item{epsilon}{longwave infrared emissivity of skin (proportion), 0.95 to 1 for most animals (Gates 1980)}
}
\value{
evaporative heat loss (W)
}
\description{
Calculate heat loss associated with evaporative water loss
}
\details{
This function allows you to estimate heat loss associated with evaporative water loss by an amphibian (Spotila et al. 1992) or lizard (based on empirical measurements in Porter et al. 1973). Reference: Spotila et al. 1992. Biophysics of Heat and Mass Transfer. In Feder and Burggren. Environmental Physiology of the Amphibians.
}
\examples{
\dontrun{
Qevaporation(A=0.1, T_b=293, taxa="amphibian", rho_s=0.003, rho_a=0.002, h=0.5, H=20, r_i=50)
Qevaporation(A=0.1, T_b=293, taxa="lizard")
}

}
\seealso{
Other biophysical models: \code{\link{Free_or_forced_convection}},
  \code{\link{Grashof_number_Gates}},
  \code{\link{Grashof_number}}, \code{\link{Nu_from_Gr}},
  \code{\link{Nu_from_Re}}, \code{\link{Nusselt_number}},
  \code{\link{Prandtl_number}},
  \code{\link{Qconduction_animal}},
  \code{\link{Qconduction_substrate}},
  \code{\link{Qconvection}},
  \code{\link{Qemitted_thermal_radiation}},
  \code{\link{Qmetabolism_from_mass_temp}},
  \code{\link{Qmetabolism_from_mass}},
  \code{\link{Qnet_Gates}},
  \code{\link{Qradiation_absorbed}},
  \code{\link{Qthermal_radiation_absorbed}},
  \code{\link{Reynolds_number}},
  \code{\link{Tb_CampbellNorman}}, \code{\link{Tb_Fei}},
  \code{\link{Tb_Gates}}, \code{\link{Tb_butterfly}},
  \code{\link{Tb_grasshopper}}, \code{\link{Tb_lizard}},
  \code{\link{Tb_salamander_humid}}, \code{\link{Tsoil}},
  \code{\link{actual_vapor_pressure}},
  \code{\link{boundary_layer_resistance}},
  \code{\link{external_resistance_to_water_vapor_transfer}},
  \code{\link{heat_transfer_coefficient_approximation}},
  \code{\link{heat_transfer_coefficient_simple}},
  \code{\link{heat_transfer_coefficient}},
  \code{\link{saturation_vapor_pressure}},
  \code{\link{saturation_water_vapor_pressure}}
}
\concept{biophysical models}
\keyword{evaporative}
\keyword{heat}
\keyword{loss}
