% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{Qevaporation}
\alias{Qevaporation}
\title{Calculate Heat Loss Associated with Evaporative Water Loss}
\usage{
Qevaporation(A, T_b, taxa, rho_s = NA, rho_a = NA, h = NA, H = NA, r_i = NA)
}
\arguments{
\item{A}{\code{numeric} surface area ((\ifelse{html}{\out{m<sup>2</sup>}}{\eqn{m^2}{ASCII}}).}

\item{T_b}{\code{numeric} body temperature (K).}

\item{taxa}{\code{character} taxa current choices: lizard, amphibian_wetskin (fully wet skin), amphibian (not fully wet skin).}

\item{rho_s}{\code{numeric} saturation water vapor density at skin surface (kg/m^3) (needed if amphibian).}

\item{rho_a}{\code{numeric} saturation water vapor density in ambient air (kg/m^3) (needed if amphibian).}

\item{h}{\code{numeric} relative humidity (0-1) (needed if amphibian).}

\item{H}{\code{numeric} convective heat transfer coefficient (W m^-2 K^-1) (needed if amphibian).}

\item{r_i}{\code{numeric} internal (cutaneous) resistance to vapor transport (s/m) (needed if amphibian).}
}
\value{
\code{numeric} evaporative heat loss (W).
}
\description{
estimate heat loss associated with evaporative water loss by an amphibian \insertCite{Spotila1992}{TrenchR} or lizard (based on empirical measurements in \insertCite{Porter1973}{TrenchR}).
}
\examples{
  Qevaporation(A     = 0.1, 
               T_b   = 293, 
               taxa  = "amphibian", 
               rho_s = 0.003, 
               rho_a = 0.002, 
               h     = 0.5, 
               H     = 20, 
               r_i   = 50)
  Qevaporation(A    = 0.1,
               T_b  = 293, 
               taxa = "lizard")

}
\references{
\insertAllCited{}
}
\seealso{
Other biophysical models: 
\code{\link{Grashof_number_Gates}()},
\code{\link{Grashof_number}()},
\code{\link{Nusselt_from_Grashof}()},
\code{\link{Nusselt_from_Reynolds}()},
\code{\link{Nusselt_number}()},
\code{\link{Prandtl_number}()},
\code{\link{Qconduction_animal}()},
\code{\link{Qconduction_substrate}()},
\code{\link{Qconvection}()},
\code{\link{Qemitted_thermal_radiation}()},
\code{\link{Qmetabolism_from_mass_temp}()},
\code{\link{Qmetabolism_from_mass}()},
\code{\link{Qnet_Gates}()},
\code{\link{Qradiation_absorbed}()},
\code{\link{Qthermal_radiation_absorbed}()},
\code{\link{Reynolds_number}()},
\code{\link{Tb_CampbellNorman}()},
\code{\link{Tb_Fei}()},
\code{\link{Tb_Gates2}()},
\code{\link{Tb_Gates}()},
\code{\link{Tb_butterfly}()},
\code{\link{Tb_grasshopper}()},
\code{\link{Tb_limpetBH}()},
\code{\link{Tb_limpet}()},
\code{\link{Tb_lizard}()},
\code{\link{Tb_mussel}()},
\code{\link{Tb_salamander_humid}()},
\code{\link{Tb_snail}()},
\code{\link{Tbed_mussel}()},
\code{\link{Tsoil}()},
\code{\link{actual_vapor_pressure}()},
\code{\link{boundary_layer_resistance}()},
\code{\link{external_resistance_to_water_vapor_transfer}()},
\code{\link{free_or_forced_convection}()},
\code{\link{heat_transfer_coefficient_approximation}()},
\code{\link{heat_transfer_coefficient_simple}()},
\code{\link{heat_transfer_coefficient}()},
\code{\link{saturation_vapor_pressure}()},
\code{\link{saturation_water_vapor_pressure}()}
}
\concept{biophysical models}
