% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{Qradiation_absorbed}
\alias{Qradiation_absorbed}
\title{Calculate absorbed solar and thermal radiation}
\usage{
Qradiation_absorbed(
  a = 0.9,
  A,
  psa_dir,
  psa_ref,
  S_dir,
  S_dif,
  S_ref = NA,
  a_s = NA
)
}
\arguments{
\item{a}{solar absorptivity of animal surface (proportion), default value is for reptiles}

\item{A}{surface area in m^2}

\item{psa_dir}{proportion surface area exposed to solar radiation}

\item{psa_ref}{proportion surface area exposed to reflected solar radiation}

\item{S_dir}{direct solar radiation (W/m^2)}

\item{S_dif}{diffuse solar radiation (W/m^2)}

\item{S_ref}{reflected solar radiation (W/m^2), either provided or estimated if surface albedo is provided instead}

\item{a_s}{is surface albedo (proportion), optional (not used) if reflected radiation is provided, Values available in Gates Biophysical ecology Table 8.2.}
}
\value{
solar radiation absorbed (W)
}
\description{
Calculate absorbed solar and thermal radiation
}
\details{
This function allows you to estimate solar and thermal radiation (W) absorbed by the surface of an animal. 
         Follows Gates' Biophysical Ecology and Spotila et al. 1992. Reference: Spotila et al. 1992. Biophysics of Heat and Mass Transfer. In Feder and Burggren. Environmental Physiology of the Amphibians.
}
\examples{
\dontrun{
Qradiation_absorbed(a=0.9, A=1, psa_dir=0.4, psa_ref=0.4, S_dir=1000, S_dif=200, a_s=0.5)
}

}
\seealso{
Other biophysical models: 
\code{\link{Free_or_forced_convection}()},
\code{\link{Grashof_number_Gates}()},
\code{\link{Grashof_number}()},
\code{\link{Nu_from_Gr}()},
\code{\link{Nu_from_Re}()},
\code{\link{Nusselt_number}()},
\code{\link{Prandtl_number}()},
\code{\link{Qconduction_animal}()},
\code{\link{Qconduction_substrate}()},
\code{\link{Qconvection}()},
\code{\link{Qemitted_thermal_radiation}()},
\code{\link{Qevaporation}()},
\code{\link{Qmetabolism_from_mass_temp}()},
\code{\link{Qmetabolism_from_mass}()},
\code{\link{Qnet_Gates}()},
\code{\link{Qthermal_radiation_absorbed}()},
\code{\link{Reynolds_number}()},
\code{\link{Tb_CampbellNorman}()},
\code{\link{Tb_Fei}()},
\code{\link{Tb_Gates}()},
\code{\link{Tb_butterfly}()},
\code{\link{Tb_grasshopper}()},
\code{\link{Tb_limpetBH}()},
\code{\link{Tb_limpet}()},
\code{\link{Tb_lizard}()},
\code{\link{Tb_mussel}()},
\code{\link{Tb_salamander_humid}()},
\code{\link{Tb_snail}()},
\code{\link{Tbed_mussel}()},
\code{\link{Tsoil}()},
\code{\link{actual_vapor_pressure}()},
\code{\link{boundary_layer_resistance}()},
\code{\link{external_resistance_to_water_vapor_transfer}()},
\code{\link{heat_transfer_coefficient_approximation}()},
\code{\link{heat_transfer_coefficient_simple}()},
\code{\link{heat_transfer_coefficient}()},
\code{\link{saturation_vapor_pressure}()},
\code{\link{saturation_water_vapor_pressure}()}
}
\concept{biophysical models}
\keyword{Solar}
\keyword{absorbed}
\keyword{radiation}
