% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{radian_to_degree}
\alias{radian_to_degree}
\title{Converts angle in radians to degrees}
\usage{
radian_to_degree(rad)
}
\arguments{
\item{rad}{angle in radians}
}
\value{
angle in degrees
}
\description{
This function allows you to convert angle in radians to degrees
}
\details{
Converts angles in radians to degrees
}
\examples{
\dontrun{
radian_to_degree(0.831)
}
}
\keyword{degrees}
\keyword{radians}
\keyword{to}
