% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{radian_to_degree}
\alias{radian_to_degree}
\title{Convert Angle in Radians to Degrees}
\usage{
radian_to_degree(rad)
}
\arguments{
\item{rad}{\code{numeric} angle (radians).}
}
\value{
\code{numeric} angle (degrees).
}
\description{
Convert angle in radians to degrees.
}
\examples{
  radian_to_degree(0.831)

}
