% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{radians_to_degrees}
\alias{radians_to_degrees}
\title{Converts angle in radians to degrees}
\usage{
radians_to_degrees(rad)
}
\arguments{
\item{rad}{angle in radians}
}
\value{
angle in degrees
}
\description{
This function allows you to convert angle in radians to degrees
}
\details{
Converts angles in radians to degrees
}
\examples{
\dontrun{
radians_to_degrees(0.831)
}
}
\keyword{degrees}
\keyword{radians}
\keyword{to}
