% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllometricFunctions.R
\name{sa_from_mass}
\alias{sa_from_mass}
\title{Calculate surface area from mass}
\usage{
sa_from_mass(m, taxa)
}
\arguments{
\item{m}{mass in g}

\item{taxa}{Which class of organism, current choices: lizard, salamander, frog, insect}
}
\value{
sa (m^2)
}
\description{
Calculate surface area from mass
}
\details{
This function allows you to estimate surface area (m^2) from mass (g) for a variety of taxa
}
\examples{
 \dontrun{
 sa_from_mass(2,"insect")
}

}
\seealso{
Other allometric functions: 
\code{\link{mass_from_length}()},
\code{\link{prop_silhouette_area_shapes}()},
\code{\link{prop_silhouette_area}()},
\code{\link{sa_from_length}()},
\code{\link{sa_from_volume}()},
\code{\link{volume_from_length}()}
}
\concept{allometric functions}
\keyword{area}
\keyword{surface}
