% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllometricFunctions.R
\name{sa_from_mass}
\alias{sa_from_mass}
\title{Calculate surface area from mass}
\usage{
sa_from_mass(m, taxa)
}
\arguments{
\item{m}{Mass in g}

\item{taxa}{Which class of organism, current choices: lizard, salamander, frog, insect}
}
\value{
sa (m^2)
}
\description{
Calculate surface area from mass
}
\details{
This function allows you to estimate surface area (m^2) from mass (g) for a variety of taxa
}
\examples{
 \dontrun{
 sa_from_mass(2,"insect")
}

}
\keyword{area}
\keyword{surface}
