% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllometricFunctions.R
\name{sa_from_volume}
\alias{sa_from_volume}
\title{Calculate surface area from volume (Based on Mitchell 1976)}
\usage{
sa_from_volume(volume, taxa)
}
\arguments{
\item{volume}{Density in m^3.}

\item{taxa}{Which class of organism, current choices: lizard, frog, sphere.}
}
\value{
sa (m^2)
}
\description{
Calculate surface area from volume (Based on Mitchell 1976)
}
\details{
This function allows you to calculate surface area (m^2) from volume (m^3) for a variety of taxa. For use in estimating convection by approximating animal shape as a sphere.
}
\examples{
 \dontrun{
 sa_from_volume(volume=0.001,"lizard")
}

}
\keyword{area}
\keyword{surface}
