% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{saturation_vapor_pressure}
\alias{saturation_vapor_pressure}
\title{Calculate saturation vapor pressure}
\usage{
saturation_vapor_pressure(Ta)
}
\arguments{
\item{Ta}{air temperature (K)}
}
\value{
saturation vapor pressure, e_s (kPa)
}
\description{
Calculate saturation vapor pressure
}
\details{
Calculate saturation vapor pressure (kPa) based on the Clausius-Clapeyron equation (Stull 2000)
}
\examples{
\dontrun{
saturation_vapor_pressure(Ta=293)
}

}
\author{
Eric Riddell
}
\keyword{pressure}
\keyword{saturation}
\keyword{vapor}
