% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{saturation_water_vapor_pressure}
\alias{saturation_water_vapor_pressure}
\title{Approximate saturation water vapor pressure}
\usage{
saturation_water_vapor_pressure(T_a)
}
\arguments{
\item{T_a}{air temperature (°C)}
}
\value{
Saturation water vapor pressure, e_s (Pa)
}
\description{
Approximate saturation water vapor pressure
}
\details{
Approximate saturation water vapor pressure as a function of ambient temperature for temperatures from 0 to 40°C (Rosenberg 1974 in Spotila et al. 1992, see also NichMapR WETAIR and DRYAIR functions from Tracy et al. 1980).  Reference: Spotila et al. 1992. Biophysics of Heat and Mass Transfer. In Feder and Burggren. Environmental Physiology of the Amphibians.
}
\examples{
\dontrun{
saturation_water_vapor_pressure(T_a=20)
}

}
\seealso{
Other biophysical models: \code{\link{Free_or_forced_convection}},
  \code{\link{Grashof_number_Gates}},
  \code{\link{Grashof_number}}, \code{\link{Nu_from_Gr}},
  \code{\link{Nu_from_Re}}, \code{\link{Nusselt_number}},
  \code{\link{Prandtl_number}},
  \code{\link{Qconduction_animal}},
  \code{\link{Qconduction_substrate}},
  \code{\link{Qconvection}},
  \code{\link{Qemitted_thermal_radiation}},
  \code{\link{Qevaporation}},
  \code{\link{Qmetabolism_from_mass_temp}},
  \code{\link{Qmetabolism_from_mass}},
  \code{\link{Qnet_Gates}},
  \code{\link{Qradiation_absorbed}},
  \code{\link{Qthermal_radiation_absorbed}},
  \code{\link{Reynolds_number}},
  \code{\link{Tb_CampbellNorman}}, \code{\link{Tb_Fei}},
  \code{\link{Tb_Gates}}, \code{\link{Tb_butterfly}},
  \code{\link{Tb_grasshopper}}, \code{\link{Tb_limpetBH}},
  \code{\link{Tb_limpet}}, \code{\link{Tb_lizard}},
  \code{\link{Tb_mussel}},
  \code{\link{Tb_salamander_humid}},
  \code{\link{Tb_snail}}, \code{\link{Tbed_mussel}},
  \code{\link{Tsoil}}, \code{\link{actual_vapor_pressure}},
  \code{\link{boundary_layer_resistance}},
  \code{\link{external_resistance_to_water_vapor_transfer}},
  \code{\link{heat_transfer_coefficient_approximation}},
  \code{\link{heat_transfer_coefficient_simple}},
  \code{\link{heat_transfer_coefficient}},
  \code{\link{saturation_vapor_pressure}}
}
\concept{biophysical models}
\keyword{Saturation}
\keyword{pressure}
\keyword{vapor}
\keyword{water}
