% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TsoilFunctions.R
\name{soil_conductivity}
\alias{soil_conductivity}
\title{Estimate soil thermal conductivity}
\usage{
soil_conductivity(x, lambda, g_a)
}
\arguments{
\item{x}{is a vector of volume fractions of soil constituents (e.g., clay, quartz, minerals other than quartz, organic matter, water, air).  The volume fractions should sum to 1. Note that x and lambda values in the example correspond to these soil constituents.}

\item{lambda}{is a vector of the thermal conductivities (W m^-1 K^-1) of the soil constituents.}

\item{g_a}{is a shape factor on soil particles.  The soil particles are assumed to be ellipsoids with axes g_a, g_b, and g_c, where g_a +g_b +g_c=1.  de Vries 1952 suggests g_a=g_b=0.125.}
}
\value{
soil thermal conductivity (W m^-1 K^-1)
}
\description{
This function allows you to estimate soil thermal conductivity in W m^-1 K^-1 using the methods of de Vries (1963, The Physics of Plant Environments, Ch2 in Environmental Control of Plant Growth).
}
\details{
Estimate soil thermal conductivity in W m^-1 K^-1
}
\examples{
\dontrun{
soil_conductivity(x=c(0.10,0.40,0.11,0.01,0.2, 0.18), lambda=c(0.10,0.40,0.11,0.01,0.2, 0.18), g_a=0.125)
}
}
\seealso{
Other soil temperature functions: \code{\link{soil_specific_heat}},
  \code{\link{soil_temperature_equation}},
  \code{\link{soil_temperature_function}},
  \code{\link{soil_temperature_integrand}},
  \code{\link{soil_temperature}}
}
\author{
Joseph Grigg
}
\concept{soil temperature functions}
\keyword{soil}
\keyword{temperature}
