% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TsoilFunctions.R
\name{soil_specific_heat}
\alias{soil_specific_heat}
\title{Estimate soil specific heat}
\usage{
soil_specific_heat(x_o, x_m, x_w, rho_so)
}
\arguments{
\item{x_o}{\code{numeric} volume fraction of organic material}

\item{x_m}{\code{numeric} volume fraction of minerals}

\item{x_w}{\code{numeric} volume fraction of water}

\item{rho_so}{\code{numeric} particle density of soil in kg/m^3 (bulk density). Defaults to 1620.}
}
\value{
\code{numeric} soil specific heat (J kg^-1 K^-1)
}
\description{
Estimates soil specific heat (J kg^-1 K^-1) using the methods of \insertCite{deVries1963;textual}{TrenchR}. Uses the volume fraction of organic material, minerals, and water in soil.
}
\examples{
  soil_specific_heat(x_o = 0.01, x_m = 0.6, x_w = 0.2, rho_so = 1620)

}
\references{
\insertAllCited{}
}
\seealso{
Other soil temperature functions: 
\code{\link{soil_conductivity}()},
\code{\link{soil_temperature_equation}()},
\code{\link{soil_temperature_function}()},
\code{\link{soil_temperature_integrand}()},
\code{\link{soil_temperature}()}
}
\author{
Joseph Grigg
}
\concept{soil temperature functions}
\keyword{soil}
\keyword{temperature}
