% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TsoilFunctions.R
\name{soil_temperature_function}
\alias{soil_temperature_function}
\title{Function to calculate soil temperature.}
\usage{
soil_temperature_function(j, T_so, params)
}
\arguments{
\item{j}{is the number of the iteration of running the model}

\item{T_so}{is the initial soil temperature profile in C}

\item{params}{is a list containing the following parameters, which are described below: list(SSA, epsilon_so, sigma, k_so, c_so, dz, k, z_r, z0, H, T_a, u, rho_a, rho_so, c_a, TimeIn, dt, shade).}
}
\value{
soil temperature function
}
\description{
Function called to calculate soil temperature from Beckman et al. (1973, Thermal Model for Prediction of a Desert Iguana's Daily and Seasonal Behavior). Parameters are passed as a list to facilitating solving the equations. This function is not intended to be called directly.
}
\details{
Function to calculate soil temperature.
}
\examples{
\dontrun{
temp_vector= runif(96, min=-10, max=10)
wind_speed_vector= runif(96, min=0, max=0.4)
time_vector= rep(1:24,4)
solrad_vector= rep(c(rep(0,6),seq(10,700,length.out=6), seq(700,10,length.out=6),rep(0,6)),4)

params=list(
  SSA=0.7, 
  epsilon_so=0.98, 
  k_so=0.293, 
  c_so=800, 
  dz=0.05, 
  z_r=1.5, 
  z0=0.02, 
  H=solrad_vector, 
  T_a=temp_vector, 
  u=wind_speed_vector, 
  rho_a=1.177,
  rho_so=1620, 
  c_a=1006, 
  TimeIn=time_vector, 
  dt=60*60, 
  shade=FALSE)

soil_temperature_function(j=1,T_so= rep(20,13), params=params)

}
}
\seealso{
Other soil temperature functions: 
\code{\link{soil_conductivity}()},
\code{\link{soil_specific_heat}()},
\code{\link{soil_temperature_equation}()},
\code{\link{soil_temperature_integrand}()},
\code{\link{soil_temperature}()}
}
\author{
Joseph Grigg
}
\concept{soil temperature functions}
\keyword{soil}
\keyword{temperature}
