% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TsoilFunctions.R
\name{soil_temperature_function}
\alias{soil_temperature_function}
\title{Function to calculate soil temperature.}
\usage{
soil_temperature_function(j, T_so, params)
}
\arguments{
\item{j}{is the number of the iteration of running the model}

\item{T_so}{is the initial soil temperature profile in degrees C}

\item{params}{is a list containing the following parameters, which are described below: list(SSA, epsilon_so, sigma, k_so, c_so, dz, k, z_r, z0, H, T_a, u, rho_a, rho_so, c_a, TimeIn,dt,shade).}

\item{SSA}{is the solar absorbtivity of soil surface as a fraction}

\item{epsilon_so}{is the thermal absorbtivity of soil surface as a fraction}

\item{k_so}{is soil thermal conductivity in W m^-1 K^-1}

\item{c_so}{is the soil specific heat capacity in in J kg^-1 K-1.}

\item{dz}{is the vertical interval in m for the soil temperature profile}

\item{z_r}{is reference height in m}

\item{z0}{is surface roughness in m}

\item{H}{is solar radiation in W m^-2}

\item{T_a}{is air temperature in degrees C}

\item{u}{is wind speed (m/s)}

\item{rho_a}{is the density of air (kg/m^3)}

\item{rho_so=}{1620 particle density of soil}

\item{c_a}{is the specific heat of air (J/(kg*K))}

\item{TimeIn}{is a vector of time periods for the model}

\item{dt=}{60*60 is the time interval for running the model}

\item{shade}{is whether or not soil temperature should be calculated in the shade, TRUE or FALSE}
}
\value{
soil temperature function
}
\description{
Function called to calculate soil temperature from Beckman et al. (1973, Thermal Model for Prediction of a Desert Iguana's Daily and Seasonal Behavior). Parameters are passed as a list to facilitating solving the equations. This function is not intended to be called directly.
}
\details{
Function to calculate soil temperature.
}
\examples{
\dontrun{
temp_vector= runif(96, min=-10, max=10)
wind_speed_vector= runif(96, min=0, max=0.4)
time_vector= rep(1:24,4)
solrad_vector= rep(c(rep(0,6),seq(10,700,length.out=6), seq(700,10,length.out=6),rep(0,6)),4)

params=list(SSA=0.7, epsilon_so=0.98, k_so=0.293, c_so=800, dz=0.05, z_r=1.5, z0=0.02, H=solrad_vector, T_a=temp_vector, u=wind_speed_vector, rho_a=1.177,rho_so=1620, c_a=1006, TimeIn=time_vector, dt=60*60, shade=FALSE)

soil_temperature_function(j=1,T_so= rep(20,13), params=params)

#RUN USING ODE SOLVER
Tsoil_out<- ode(y = rep(20,13), func = soil_temperature, times = 1:length(solrad_vector), parms=params)
}
}
\author{
Joseph Grigg
}
\keyword{soil}
\keyword{temperature}
