% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{solar_noon}
\alias{solar_noon}
\title{Utility: Calculate time of solar noon}
\usage{
solar_noon(lon, doy, offset = NA)
}
\arguments{
\item{lon}{\code{numeric} longitude (decimal degrees)}

\item{doy}{\code{numeric} day of year (1-366). (This can be obtained from standard date via the TrenchR::day_of_year() function.)}

\item{offset}{\code{numeric} number of hours to add to UTC to get local time (improves accuracy but not always necessary). Defaults to NA.}
}
\value{
\code{numeric} time of solar noon (hours).
}
\description{
Calculate the time of solar noon in hours as a function of the day of year and longitude  \insertCite{Campbell1998}{TrenchR}.
}
\examples{
  solar_noon(lon = -122.335, doy = 112)

}
\references{
\insertAllCited{}
}
\seealso{
Other utility functions: 
\code{\link{airpressure_from_elev}()},
\code{\link{azimuth_angle}()},
\code{\link{day_of_year}()},
\code{\link{daylength}()},
\code{\link{dec_angle}()},
\code{\link{degree_to_radian}()},
\code{\link{fahrenheit_to_celsius}()},
\code{\link{fahrenheit_to_kelvin}()},
\code{\link{kelvin_to_celsius}()},
\code{\link{zenith_angle}()}
}
\concept{utility functions}
