% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{solar_radiation_absorbed}
\alias{solar_radiation_absorbed}
\title{Calculate absorbed solar radiation}
\usage{
solar_radiation_absorbed(mass, solar, species = "lizard")
}
\arguments{
\item{mass}{Mass in grams.}

\item{solar}{Downward solar radiation(W/m^2).}

\item{species}{Which species.}
}
\value{
solar radiation absorbed
}
\description{
Calculate absorbed solar radiation
}
\details{
This function allows you to calculate absorbed solar radiation of an ectotherm(Reptiles)
}
\examples{
\dontrun{
solar_radiation_absorbed(10.5,1000,"lizard")
}

}
\keyword{Solar}
\keyword{absorbed}
\keyword{radiation}
