% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempWindProfileFunctions.R
\name{surface_roughness}
\alias{surface_roughness}
\title{Estimate surface roughness from empirical measurements.}
\usage{
surface_roughness(u_r, zr)
}
\arguments{
\item{u_r}{\code{numeric} wind velocity at a vector of reference heights (m/s).}

\item{zr}{\code{numeric} vector of reference heights (meters)}
}
\value{
surface roughness (meters)
}
\description{
estimate surface roughness (m) from empirical wind speed (m/s) data collected at a vector of heights (m). Estimates surface roughness from empirical measurements. References: \insertCite{Kingsolver2015}{TrenchR}, \insertCite{Campbell1998}{TrenchR}, and \insertCite{Porter1979}{TrenchR}.
}
\examples{
surface_roughness(u_r = c(0.01, 0.025, 0.05, 0.1, 0.2), zr = c(0.05, 0.25, 0.5, 0.75, 1))

}
\references{
\insertAllCited{}
}
\seealso{
Other microclimate functions: 
\code{\link{air_temp_profile_neutral}()},
\code{\link{air_temp_profile_segment}()},
\code{\link{air_temp_profile}()},
\code{\link{degree_days}()},
\code{\link{direct_solar_radiation}()},
\code{\link{diurnal_radiation_variation}()},
\code{\link{diurnal_temp_variation_sineexp}()},
\code{\link{diurnal_temp_variation_sinesqrt}()},
\code{\link{diurnal_temp_variation_sine}()},
\code{\link{estimate_radiation}()},
\code{\link{monthly_solar_radiation}()},
\code{\link{partition_solar_radiation}()},
\code{\link{proportion_diffuse_solar_radiation}()},
\code{\link{wind_speed_profile_neutral}()},
\code{\link{wind_speed_profile_segment}()}
}
\concept{microclimate functions}
\keyword{profile}
\keyword{wind}
