% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempWindProfileFunctions.R
\name{surface_roughness}
\alias{surface_roughness}
\title{Estimate surface roughness from empirical measurements.}
\usage{
surface_roughness(u_r, zr)
}
\arguments{
\item{u_r}{is wind velocity at a vector of reference heights in m/s.}

\item{zr}{is the vector of reference heights in m.}
}
\value{
surface roughness (m)
}
\description{
This function allows you to estimate surface roughness in m from empirical wind speed (m/s) 
data collected at a vector of heights (m). Estimates surface roughness from empirical measurements. 
References: Kingsolver and Buckley. 2015. Climate variability slows evolutionary responses of Colias butterflies to recent climate change. PRSb;
G. S. Campbell, J. M. Norman, An introduction to environmental biophysics. (Springer Science, New York, 1998); W. P. Porter, J. W. Mitchell, W. A. Beckman, C. B. DeWitt, Behavioral implications of mechanistic
ecology. Thermal and behavioral modeling of desert ecotherms and their microenvironment. Oecologia 13, 1-54 (1973)
}
\details{
Estimate surface roughness in m
}
\examples{
\dontrun{
surface_roughness(u_r=c(0.01,0.025,0.05,0.1,0.2), zr=c(0.05,0.25,0.5,0.75,1))
}
}
\seealso{
Other microclimate functions: 
\code{\link{air_temp_profile_neutral}()},
\code{\link{air_temp_profile_segment}()},
\code{\link{air_temp_profile}()},
\code{\link{degree_days}()},
\code{\link{diurnal_radiation_variation}()},
\code{\link{diurnal_temp_variation_sineexp}()},
\code{\link{diurnal_temp_variation_sinesqrt}()},
\code{\link{diurnal_temp_variation_sine}()},
\code{\link{estimate_radiation}()},
\code{\link{monthly_solar_radiation}()},
\code{\link{partition_solar_radiation}()},
\code{\link{wind_speed_profile_neutral}()},
\code{\link{wind_speed_profile_segment}()}
}
\concept{microclimate functions}
\keyword{profile}
\keyword{wind}
