% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempWindProfileFunctions.R
\name{surface_roughness}
\alias{surface_roughness}
\title{Estimate surface roughness from empirical measurements.}
\usage{
surface_roughness(u_r, zr)
}
\arguments{
\item{u_r}{is wind velocity at a vector of reference heights in m/s.}

\item{zr}{is the vector of reference heights in m.}
}
\value{
surface roughness
}
\description{
This function allows you to estimate surface roughness in m from empirical wind speed (m/s) 
data collected at a vector of heights (m). Estimates surface roughness from empirical measurements.
}
\details{
Estimate surface roughness in m
}
\examples{
\dontrun{
surface_roughness(u_r=c(0.01,0.025,0.05,0.1,0.2), zr=c(0.05,0.25,0.5,0.75,1))
}
}
\keyword{profile}
\keyword{wind}
