% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T13AnnualMeans.R
\name{T13AnnualMeans}
\alias{T13AnnualMeans}
\title{T13 annual-mean isotope time series.}
\usage{
T13AnnualMeans(t1, t2, depth, i.max = c(5, 11, 19, 24, 33, 38),
  i.min = c(7, 14, 21, 26, 37), start.year = 2013, cheat = FALSE)
}
\arguments{
\item{t1}{numeric vector of the T13--1 isotope record.}

\item{t2}{numeric vector of the T13--2 isotope record.}

\item{depth}{numeric vector of the common T13 depth scale (i.e. after
the optimal shift of T13--2 to maximise the inter-trench correlation; see
Münch et al. (2016)).}

\item{i.max}{index positions of the summer maxima of the trench mean isotope
record; default indices are the ones used in Münch et al. (2016).}

\item{i.min}{index positions of the winter minima of the trench mean isotope
record; default indices are the ones used in Münch et al. (2016).}

\item{start.year}{assumed year of the first summer maximum where the maximum
is defined to occur in January of the year; defaults to \code{2013}.}

\item{cheat}{Originally, the annual mean isotope data have been erroneously
calculated such that the last value contributing to bin \code{i} was also
included in bin \code{i + 1}, due to an erroneous implementation of
\code{\link{AverageIndexBins}}. This bug has been fixed. For
\code{cheat = TRUE}, the original annual mean data using this erroneous bin
definition can be reproduced, while for \code{cheat = FALSE}, the correct
implementation is used. However, the difference in the annual mean data
between the two version is minor and thus does not influence any results or
conclusions of Münch et al. (2017).}
}
\value{
A list of three data frames:
\describe{
  \item{means:}{the annual-mean time series of the T13--1, T13--2 and mean
  T13 isotope records together with the range of annual means from the
  different binning methods;}
  \item{summer.max:}{the annual time series of summer maximum values for the
  T13--1, T13--2 and mean T13 isotope records;}
  \item{winter.min:}{the annual time series of winter minimum values for the
  T13--1, T13--2 and mean T13 isotope records.}
}
}
\description{
Calculate annual-mean time series of the T13 trench isotope records defined
by binning the records according to the isotopic maxima and minima. From
this, four different time series are created: averages from bins defined by
(1) the summer maxima, (2) the winter minima, (3) the midpoints of the
ascending slopes flanking the maxima and (4) the midpoints of the descending
slopes.
}
\references{
Münch, T., et al., Regional climate signal vs. local noise: a
two-dimensional view of water isotopes in Antarctic firn at Kohnen Station,
Dronning Maud Land, Clim. Past, doi:10.5194/cp-12-1565-2016, 2016.
}
\author{
Thomas Münch
}
