% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docu.R
\docType{data}
\name{t15.trench2}
\alias{t15.trench2}
\title{Trench T15-2 stable water isotopologue data and meta information}
\format{
A tibble with 1375 rows and 8 variables:
\describe{
  \item{profileName:}{character string of the name of the trench profile.}
  \item{profilePosition:}{the horizontal position of the trench profile in m
    relative to the first profile.}
  \item{surfaceHeight:}{the surface "height" in cm at each profile position
    (positive downwards) relative to the maximum observed surface height.}
  \item{sampleNumber:}{number enumerating the depth samples taken at each
    profile position; i.e. the profile name together with the sample number
    specify the exact position of the isotope sample within the entire
    trench.}
  \item{depth:}{depth in cm of the isotope sample on an absolute depth scale,
    i.e. relative to the maximum observed surface height along the profile
    positions.}
  \item{d18O:}{the sample's oxygen isotope composition in permil VSMOW/SLAP.}
  \item{dD:}{the sample's hydrogen isotope composition in permil VSMOW/SLAP.}
  \item{dxs:}{d-excess value in permil VSMOW/SLAP calculated from d18O and
    dD.}
}
}
\source{
\url{https://doi.pangaea.de/10.1594/PANGAEA.876638}
}
\usage{
t15.trench2
}
\description{
A dataset providing the stable water isotopologue data measured on Kohnen
Snow Trench T15-2 (oxygen isotopes, hydrogen isotopes, and deuterium excess)
together with meta information.
}
\details{
Note that isotope values close to the surface are NA when the surface height
at a profile position does not reach the maximum surface height. A few
depth samples (T15-2-W-39 and 52; T15-2-X-48, 50-52, 74 and 77; T15-2-Y-36
and 42) were lost in the measurement process and have been linearly
interpolated from the neighbouring values.

For data source files and for how to create this dataset from source, see the
\code{data-raw} folder in the R package source. The data are also available
in a similar format from PANGAEA, see the source information below, but note
that sample positions with NA values due to the variable surface height are
not included in the PANGAEA dataset and missing samples have not been
interpolated.
}
\keyword{datasets}
