% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperativeTemperatureFunctions.R
\name{Tb_CampbellNorman}
\alias{Tb_CampbellNorman}
\title{Predicts body temperatures (operative environmental temperature) of an ectotherm in K.}
\usage{
Tb_CampbellNorman(T_a, S, epsilon = 0.96, c_p = 29.3, D, V)
}
\arguments{
\item{T_a}{is air temperature in K.}

\item{S}{solar and thermal radiation absorbed (W m^-2)}

\item{epsilon}{longwave infrared emissivity of skin (proportion), 0.95 to 1 for most animals (Gates 1980)}

\item{c_p}{specific heat of air (J mol^-1 K^-1)}

\item{D}{characteristic dimension of the animal (m)}

\item{V}{is wind speed in m/s}
}
\value{
operative environmental temperature, T_e (K)
}
\description{
Predicts body temperatures (operative environmental temperature) of an ectotherm in K.
}
\details{
Predicts body temperatures (operative environmental temperature) of an ectotherm in K. Uses approximation in Campbell and Norman (1998).
}
\examples{
\dontrun{
Tb_CampbellNorman(T_a=303, S=823, epsilon=0.96, c_p=29.3, D=0.17, V=1)
}

}
\keyword{environmental}
\keyword{operative}
\keyword{temperature}
