% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_To.R
\name{Tb_Fei}
\alias{Tb_Fei}
\title{Calculate body temperature(Operative Temperature) of Lizard using model by Fei et al.}
\usage{
Tb_Fei(TAH, T_s, T_a, UDT, GLW, shade)
}
\arguments{
\item{TAH}{Temperature in canopy.}

\item{T_s}{Surface Temperature.}

\item{T_a}{Air Temperature.}

\item{UDT}{upper developmental threshold.}

\item{GLW}{Downward flux of near-infrared radiation}

\item{shade}{type of method being used}
}
\description{
[1] T. Fei, A. K. Skidmore, V. Venus, T. Wang, M. Schlerf, B. Toxopeus, 
S. van Overjijk, M. Bian, and Y. Liu, “A body temperature model for 
lizards as estimated from the thermal environment,” J. Therm. 
Biol., vol. 37, no. 1, pp. 56–64, 2012.
TAH, T_s, T_a, UDT, GLW, time_step, shade
}
\details{
Calculate Operative Temperature of Lizard.
}
\examples{
\dontrun{
Tb_Fei(254, 265, 267, 15, 307,  0)
}

}
\author{
Ofir Levy
}
\keyword{Fei}
\keyword{Operative}
\keyword{Temperature}
