% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_To.R
\name{Tb_Fei}
\alias{Tb_Fei}
\title{Calculate body temperature(Operative Temperature) of Lizard using model by Fei et al.}
\usage{
Tb_Fei(T_a, T_g, H, lw, shade, m, Acondfact, Agradfact)
}
\arguments{
\item{T_a}{is air temperature at lizard height in K}

\item{T_g}{is surface temperature in K}

\item{H}{is total (direct + diffuse) solar radiation flux in W/m^2}

\item{lw}{Downward flux of near-infrared radiation (W/m^2)}

\item{shade}{proportion of shade}

\item{m}{is lizard mass in g}

\item{Acondfact}{is the proportion of the lizard projected area that is in contact with the ground, Acondfact=0.05 for standing and Acondfact=0.35 for lying on ground}

\item{Agradfact}{is the proportion of the lizard projected area exposed to radiation from the ground, Agradfact=0.4 for standing and Agradfact=0.0 for lying on ground}
}
\description{
Transient model for predicting operative environmental temperature using model in the following publication: T. Fei, A. K. Skidmore, V. Venus, T. Wang, M. Schlerf, B. Toxopeus, 
S. van Overjijk, M. Bian, and Y. Liu, “A body temperature model for 
lizards as estimated from the thermal environment,” J. Therm. 
Biol., vol. 37, no. 1, pp. 56–64, 2012.
}
\details{
Predicts body temperature (operative environmental temperature) of a lizard in K.
}
\examples{
\dontrun{
Tb_Fei(T_a=293, T_g=300, H=600, lw=30, shade=0.5, m=10.5, Acondfact=0.05, Agradfact=0.4)
}

}
\author{
Ofir Levy
}
\keyword{Fei}
\keyword{Operative}
\keyword{Temperature}
