% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_To.R
\name{Tb_Fei}
\alias{Tb_Fei}
\title{Calculate body temperature(Operative Temperature) of Lizard using model by Fei et al.}
\usage{
Tb_Fei(T_a, T_g, H, lw, shade, m, Acondfact, Agradfact)
}
\arguments{
\item{T_a}{is air temperature at lizard height in K}

\item{T_g}{is surface temperature in K}

\item{H}{is total (direct + diffuse) solar radiation flux in W/m^2}

\item{lw}{Downward flux of near-infrared radiation (W/m^2)}

\item{shade}{proportion of shade}

\item{m}{is lizard mass in g}

\item{Acondfact}{is the proportion of the lizard projected area that is in contact with the ground, Acondfact=0.1 for standing and Acondfact=0.4 for lying on ground}

\item{Agradfact}{is the proportion of the lizard projected area exposed to radiation from the ground, Agradfact=0.3 for standing and Agradfact=0.0 for lying on ground}
}
\value{
Body temperature of a lizard in K
}
\description{
Transient model for predicting operative environmental temperature using model in the following publication: T. Fei, A. K. Skidmore, V. Venus, T. Wang, M. Schlerf, B. Toxopeus, 
S. van Overjijk, M. Bian, and Y. Liu, “A body temperature model for lizards as estimated from the thermal environment,” J. Therm. Biol., vol. 37, no. 1, pp. 56–64, 2012.
}
\details{
Predicts body temperature (operative environmental temperature) of a lizard in K.
}
\examples{
\dontrun{
Tb_Fei(T_a=293, T_g=300, H=1300, lw=60, shade=0.5, m=10.5, Acondfact=0.1, Agradfact=0.3)
}

}
\seealso{
Other biophysical models: \code{\link{Free_or_forced_convection}},
  \code{\link{Grashof_number_Gates}},
  \code{\link{Grashof_number}}, \code{\link{Nu_from_Gr}},
  \code{\link{Nu_from_Re}}, \code{\link{Nusselt_number}},
  \code{\link{Prandtl_number}},
  \code{\link{Qconduction_animal}},
  \code{\link{Qconduction_substrate}},
  \code{\link{Qconvection}},
  \code{\link{Qemitted_thermal_radiation}},
  \code{\link{Qevaporation}},
  \code{\link{Qmetabolism_from_mass_temp}},
  \code{\link{Qmetabolism_from_mass}},
  \code{\link{Qnet_Gates}},
  \code{\link{Qradiation_absorbed}},
  \code{\link{Qthermal_radiation_absorbed}},
  \code{\link{Reynolds_number}},
  \code{\link{Tb_CampbellNorman}}, \code{\link{Tb_Gates}},
  \code{\link{Tb_butterfly}}, \code{\link{Tb_grasshopper}},
  \code{\link{Tb_limpetBH}}, \code{\link{Tb_limpet}},
  \code{\link{Tb_lizard}}, \code{\link{Tb_mussel}},
  \code{\link{Tb_salamander_humid}},
  \code{\link{Tb_snail}}, \code{\link{Tbed_mussel}},
  \code{\link{Tsoil}}, \code{\link{actual_vapor_pressure}},
  \code{\link{boundary_layer_resistance}},
  \code{\link{external_resistance_to_water_vapor_transfer}},
  \code{\link{heat_transfer_coefficient_approximation}},
  \code{\link{heat_transfer_coefficient_simple}},
  \code{\link{heat_transfer_coefficient}},
  \code{\link{saturation_vapor_pressure}},
  \code{\link{saturation_water_vapor_pressure}}
}
\author{
Ofir Levy
}
\concept{biophysical models}
\keyword{Fei}
\keyword{Operative}
\keyword{Temperature}
