% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biophysmodel_Grasshopper.R
\name{Tb_grasshopper}
\alias{Tb_grasshopper}
\title{Predicts body temperatures (operative environmental temperature) of a grasshopper in °C.}
\usage{
Tb_grasshopper(T_a, T_g, u, H, K_t, psi, L, Acondfact = 0.25,
  z = 0.001, abs = 0.7, r_g = 0.3)
}
\arguments{
\item{T_a}{is air temperature in °C}

\item{T_g}{is surface temperature in °C, Kingsolver (1983) assumes T_g-T_a=8.4}

\item{u}{is wind speed in m/s}

\item{H}{is total (direct + diffuse) solar radiation flux in W/m^2}

\item{K_t}{is the clearness index (dimensionless), which is the ratio of the global solar radiation measured at the surface to the total solar radiation at the top of the atmosphere.}

\item{psi}{is solar zenith angle in degrees}

\item{L}{in grasshopper length in m}

\item{Acondfact}{is the proportion of the grasshopper surface area that is in contact with the ground}

\item{z}{is grasshopper's distance from the ground in m}

\item{abs}{is absorptivity of grasshopper to solar radiation (proportion), See Anderson et al. (1979).}

\item{r_g}{is substrate solar reflectivity (proportion), see Kingsolver (1983)}
}
\value{
predicted body (operative environmental) temperature (°C)
}
\description{
Predicts body temperature (operative environmental temperature) of a grasshopper in °C. Described in Buckleyet al. (2014, Phenotypic clines, energy balances, and ecological responses to climate change. Journal of Animal Ecology 83:41-50.) See also a related model by Anderson et al. (1979, Habitat selection in two species of short-horned grasshoppers. Oecologia 38:359–74.)
}
\details{
Predicts body temperatures (operative environmental temperature) of a grasshopper in °C.
}
\examples{
\dontrun{
Tb_grasshopper(T_a=25, T_g=25, u=0.4, H=400, K_t=0.7, psi=30, L=0.02, Acondfact=0.25, z=0.001, abs=0.7, r_g=0.3)
}
}
\seealso{
Other biophysical models: \code{\link{Free_or_forced_convection}},
  \code{\link{Grashof_number_Gates}},
  \code{\link{Grashof_number}}, \code{\link{Nu_from_Gr}},
  \code{\link{Nu_from_Re}}, \code{\link{Nusselt_number}},
  \code{\link{Prandtl_number}},
  \code{\link{Qconduction_animal}},
  \code{\link{Qconduction_substrate}},
  \code{\link{Qconvection}},
  \code{\link{Qemitted_thermal_radiation}},
  \code{\link{Qevaporation}},
  \code{\link{Qmetabolism_from_mass_temp}},
  \code{\link{Qmetabolism_from_mass}},
  \code{\link{Qnet_Gates}},
  \code{\link{Qradiation_absorbed}},
  \code{\link{Qthermal_radiation_absorbed}},
  \code{\link{Reynolds_number}},
  \code{\link{Tb_CampbellNorman}}, \code{\link{Tb_Fei}},
  \code{\link{Tb_Gates}}, \code{\link{Tb_butterfly}},
  \code{\link{Tb_limpetBH}}, \code{\link{Tb_limpet}},
  \code{\link{Tb_lizard}}, \code{\link{Tb_mussel}},
  \code{\link{Tb_salamander_humid}},
  \code{\link{Tb_snail}}, \code{\link{Tbed_mussel}},
  \code{\link{Tsoil}}, \code{\link{actual_vapor_pressure}},
  \code{\link{boundary_layer_resistance}},
  \code{\link{external_resistance_to_water_vapor_transfer}},
  \code{\link{heat_transfer_coefficient_approximation}},
  \code{\link{heat_transfer_coefficient_simple}},
  \code{\link{heat_transfer_coefficient}},
  \code{\link{saturation_vapor_pressure}},
  \code{\link{saturation_water_vapor_pressure}}
}
\concept{biophysical models}
\keyword{biophysical}
\keyword{body}
\keyword{model}
\keyword{temperature,}
