% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biophysmodel_Grasshopper.R
\name{Tb_grasshopper}
\alias{Tb_grasshopper}
\title{Operative Environmental Temperature of a Grasshopper}
\usage{
Tb_grasshopper(
  T_a,
  T_g,
  u,
  H,
  K_t,
  psi,
  L,
  Acondfact = 0.25,
  z = 0.001,
  abs = 0.7,
  r_g = 0.3
)
}
\arguments{
\item{T_a}{\code{numeric} air temperature in C.}

\item{T_g}{\code{numeric} surface temperature in C. \insertCite{Kingsolver1983;textual}{TrenchR} assumes T_g - T_a = 8.4.}

\item{u}{\code{numeric} wind speed in m / s.}

\item{H}{\code{numeric} total (direct + diffuse) solar radiation flux in W / \ifelse{html}{\out{m<sup>2</sup>}}{\eqn{m^2}{ASCII}}.}

\item{K_t}{\code{numeric} clearness index (dimensionless), which is the ratio of the global solar radiation measured at the surface to the total solar radiation at the top of the atmosphere.}

\item{psi}{\code{numeric} solar zenith angle in degrees.}

\item{L}{\code{numeric} grasshopper length in m.}

\item{Acondfact}{\code{numeric} the proportion of the grasshopper surface area that is in contact with the ground.}

\item{z}{\code{numeric} distance from the ground to the grasshopper in m}

\item{abs}{\code{numeric} absorptivity of the grasshopper to solar radiation (proportion). See \insertCite{Anderson1979;textual}{TrenchR}.}

\item{r_g}{\code{numeric} substrate solar reflectivity (proportion), see \insertCite{Kingsolver1983;textual}{TrenchR}.}
}
\value{
\code{numeric} predicted body (operative environmental) temperature (C).
}
\description{
Predicts body temperatures (operative environmental temperatures) of a grasshopper in C.
 \cr \cr
 Based on \insertCite{Swinbank1963;textual}{TrenchR}, using the regression in  \insertCite{Lactin1998;textual}{TrenchR} following \insertCite{Gates1962;textual}{TrenchR} in  \insertCite{Kingsolver1983;textual}{TrenchR}.
}
\details{
Total radiative flux is calculated as thermal radiative heat flux plus convective heat flux, following \insertCite{Kingsolver1983;textual}{TrenchR}, with the \insertCite{Erbs1982;textual}{TrenchR} model from \insertCite{Wong2001;textual}{TrenchR}.
  \cr \cr
  Energy balance is based on \insertCite{Kingsolver1983;textual}{TrenchR}.
  \cr \cr
  W is calculated without area dependence \insertCite{Anderson1979}{TrenchR}.
  \cr \cr
  The body of a grasshopper female is approximated by a rotational ellipsoid with half the body length as the semi-major axis  \insertCite{Samietz2005}{TrenchR}.
  \cr \cr
  The diffuse fraction is corrected following \insertCite{Olyphant1984;textual}{TrenchR}.
}
\examples{
  Tb_grasshopper(T_a       = 25, 
                 T_g       = 25,      
                 u         = 0.4, 
                 H         = 400, 
                 K_t       = 0.7, 
                 psi       = 30, 
                 L         = 0.02, 
                 Acondfact = 0.25, 
                 z         = 0.001, 
                 abs       = 0.7, 
                 r_g       = 0.3)

}
\references{
\insertAllCited{}
}
\seealso{
Other biophysical models: 
\code{\link{Free_or_forced_convection}()},
\code{\link{Grashof_number_Gates}()},
\code{\link{Grashof_number}()},
\code{\link{Nu_from_Gr}()},
\code{\link{Nu_from_Re}()},
\code{\link{Nusselt_number}()},
\code{\link{Prandtl_number}()},
\code{\link{Qconduction_animal}()},
\code{\link{Qconduction_substrate}()},
\code{\link{Qconvection}()},
\code{\link{Qemitted_thermal_radiation}()},
\code{\link{Qevaporation}()},
\code{\link{Qmetabolism_from_mass_temp}()},
\code{\link{Qmetabolism_from_mass}()},
\code{\link{Qnet_Gates}()},
\code{\link{Qradiation_absorbed}()},
\code{\link{Qthermal_radiation_absorbed}()},
\code{\link{Reynolds_number}()},
\code{\link{Tb_CampbellNorman}()},
\code{\link{Tb_Fei}()},
\code{\link{Tb_Gates2}()},
\code{\link{Tb_Gates}()},
\code{\link{Tb_butterfly}()},
\code{\link{Tb_limpetBH}()},
\code{\link{Tb_limpet}()},
\code{\link{Tb_lizard}()},
\code{\link{Tb_mussel}()},
\code{\link{Tb_salamander_humid}()},
\code{\link{Tb_snail}()},
\code{\link{Tbed_mussel}()},
\code{\link{Tsoil}()},
\code{\link{actual_vapor_pressure}()},
\code{\link{boundary_layer_resistance}()},
\code{\link{external_resistance_to_water_vapor_transfer}()},
\code{\link{heat_transfer_coefficient_approximation}()},
\code{\link{heat_transfer_coefficient_simple}()},
\code{\link{heat_transfer_coefficient}()},
\code{\link{saturation_vapor_pressure}()},
\code{\link{saturation_water_vapor_pressure}()}
}
\concept{biophysical models}
