% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biophysmodel_Mussel.R
\name{Tb_mussel}
\alias{Tb_mussel}
\title{Calculate Operative Environmental Temperature of a Mussel in C.}
\usage{
Tb_mussel(L, H, T_a, T_g, S, k_d, u, psi, cl, evap = FALSE, group = "solitary")
}
\arguments{
\item{L}{\code{numeric} mussel length (anterior/posterior axis) (m).}

\item{H}{\code{numeric} mussel height (dorsal/ventral axis) (m). It is reasonable to assume \code{H = 0.5 * L}.}

\item{T_a}{\code{numeric} air temperature (C).}

\item{T_g}{\code{numeric} ground temperature (C).}

\item{S}{\code{numeric} direct solar flux density (W / \ifelse{html}{\out{m<sup>2</sup>}}{\eqn{m^2}{ASCII}}).}

\item{k_d}{\code{numeric} diffuse fraction, proportion of solar radiation that is diffuse.}

\item{u}{\code{numeric} wind speed (m/s).}

\item{psi}{\code{numeric} solar zenith angle (degrees): can be calculated from \code{\link{zenith_angle function}}.}

\item{cl}{\code{numeric} fraction of the sky covered by cloud.}

\item{evap}{\code{logical} if mussels gaping to evaporatively cool. If \code{TRUE}, assumes constant mass loss rate of 5 percent of initial body mass per hour.}

\item{group}{\code{character}; options are \code{"aggregated"}: mussels living in beds; \code{"solitary"}: mussels individuals, anterior or posterior end facing upwind; and \code{"solitary_valve"}: solitary individuals, valve facing upwind.}
}
\value{
\code{numeric} predicted body temperature (C).
}
\description{
Predicts body temperature of a mussel in C. Implements a steady-state model, which assumes unchanging environmental conditions.
}
\details{
Thermal radiative flux is calculated following \insertCite{Helmuth1998;textual}{TrenchR}, \insertCite{Helmuth1999;textual}{TrenchR}, and \insertCite{Idso1969;textual}{TrenchR}.
}
\examples{

Tb_mussel(L     = 0.1, 
          H     = 0.05, 
          T_a   = 25, 
          T_g   = 30, 
          S     = 500, 
          k_d   = 0.2, 
          u     = 2, 
          psi   = 30, 
          evap  = FALSE, 
          cl    = 0.5, 
          group = "solitary")

}
\references{
\insertAllCited{}
}
\seealso{
Other biophysical models: 
\code{\link{Grashof_number_Gates}()},
\code{\link{Grashof_number}()},
\code{\link{Nusselt_from_Grashof}()},
\code{\link{Nusselt_from_Reynolds}()},
\code{\link{Nusselt_number}()},
\code{\link{Prandtl_number}()},
\code{\link{Qconduction_animal}()},
\code{\link{Qconduction_substrate}()},
\code{\link{Qconvection}()},
\code{\link{Qemitted_thermal_radiation}()},
\code{\link{Qevaporation}()},
\code{\link{Qmetabolism_from_mass_temp}()},
\code{\link{Qmetabolism_from_mass}()},
\code{\link{Qnet_Gates}()},
\code{\link{Qradiation_absorbed}()},
\code{\link{Qthermal_radiation_absorbed}()},
\code{\link{Reynolds_number}()},
\code{\link{Tb_CampbellNorman}()},
\code{\link{Tb_Fei}()},
\code{\link{Tb_Gates2}()},
\code{\link{Tb_Gates}()},
\code{\link{Tb_butterfly}()},
\code{\link{Tb_grasshopper}()},
\code{\link{Tb_limpetBH}()},
\code{\link{Tb_limpet}()},
\code{\link{Tb_lizard}()},
\code{\link{Tb_salamander_humid}()},
\code{\link{Tb_snail}()},
\code{\link{Tbed_mussel}()},
\code{\link{Tsoil}()},
\code{\link{actual_vapor_pressure}()},
\code{\link{boundary_layer_resistance}()},
\code{\link{external_resistance_to_water_vapor_transfer}()},
\code{\link{free_or_forced_convection}()},
\code{\link{heat_transfer_coefficient_approximation}()},
\code{\link{heat_transfer_coefficient_simple}()},
\code{\link{heat_transfer_coefficient}()},
\code{\link{saturation_vapor_pressure}()},
\code{\link{saturation_water_vapor_pressure}()}
}
\concept{biophysical models}
