% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biophysmodel_Mussel.R
\name{Tb_mussel}
\alias{Tb_mussel}
\title{Predicts body temperature (operative environmental temperature) of a mussel in °C.}
\usage{
Tb_mussel(L, H, T_a, T_g, S, k_d, u, psi, evap = FALSE, cl,
  group = "solitary")
}
\arguments{
\item{L}{mussel length (anterior/posterior axis) (m)}

\item{H}{mussel height (dorsal/ventral axis) (m), reasonable to assume 0.5*L}

\item{T_a}{air temperature (°C)}

\item{T_g}{ground temperature (°C)}

\item{S}{direct solar flux density (W/m2)}

\item{k_d}{diffuse fraction, proportion of solar radiation that is diffuse}

\item{u}{wind speed (m/s)}

\item{psi}{solar zenith angle (degrees): can be calculated from zenith_angle function}

\item{evap}{Are mussels gaping to evaporatively cool? TRUE of FALSE (default), If TRUE, assumes constant mass loss rate of 5% of initial body mass per hour}

\item{cl}{fraction of the sky covered by cloud}

\item{group}{options are "aggregated": mussels living in beds, "solitary": mussels individuals, anterior or posterior end facing upwind, 
and "solitary_valve": solitary individuals, valve facing upwind}
}
\value{
predicted body temperature (°C)
}
\description{
Predicts body temperature of a mussel in °C. Implements a steady‐state model, which assumes unchanging environmental conditions. Based on Helmuth 1998, INTERTIDAL MUSSEL MICROCLIMATES: PREDICTING THE BODY TEMPERATURE OF A SESSILE INVERTEBRATE
}
\details{
Predicts body temperature of a mussel in °C.
}
\examples{
\dontrun{
Tb_mussel(L = 0.1, H = 0.05, T_a = 25, T_g = 30, S=500, k_d=0.2, u = 2, psi = 30, evap = FALSE, cl = 0.5, group = "solitary")
}
}
\seealso{
Other biophysical models: \code{\link{Free_or_forced_convection}},
  \code{\link{Grashof_number_Gates}},
  \code{\link{Grashof_number}}, \code{\link{Nu_from_Gr}},
  \code{\link{Nu_from_Re}}, \code{\link{Nusselt_number}},
  \code{\link{Prandtl_number}},
  \code{\link{Qconduction_animal}},
  \code{\link{Qconduction_substrate}},
  \code{\link{Qconvection}},
  \code{\link{Qemitted_thermal_radiation}},
  \code{\link{Qevaporation}},
  \code{\link{Qmetabolism_from_mass_temp}},
  \code{\link{Qmetabolism_from_mass}},
  \code{\link{Qnet_Gates}},
  \code{\link{Qradiation_absorbed}},
  \code{\link{Qthermal_radiation_absorbed}},
  \code{\link{Reynolds_number}},
  \code{\link{Tb_CampbellNorman}}, \code{\link{Tb_Fei}},
  \code{\link{Tb_Gates}}, \code{\link{Tb_butterfly}},
  \code{\link{Tb_grasshopper}}, \code{\link{Tb_limpetBH}},
  \code{\link{Tb_limpet}}, \code{\link{Tb_lizard}},
  \code{\link{Tb_salamander_humid}},
  \code{\link{Tb_snail}}, \code{\link{Tbed_mussel}},
  \code{\link{Tsoil}}, \code{\link{actual_vapor_pressure}},
  \code{\link{boundary_layer_resistance}},
  \code{\link{external_resistance_to_water_vapor_transfer}},
  \code{\link{heat_transfer_coefficient_approximation}},
  \code{\link{heat_transfer_coefficient_simple}},
  \code{\link{heat_transfer_coefficient}},
  \code{\link{saturation_vapor_pressure}},
  \code{\link{saturation_water_vapor_pressure}}
}
\concept{biophysical models}
\keyword{biophysical}
\keyword{body}
\keyword{model}
\keyword{temperature,}
