% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{Tb_salamander_humid}
\alias{Tb_salamander_humid}
\title{Calculate humid operative temperature}
\usage{
Tb_salamander_humid(r_i, r_b, D, T_a, elev, e_a, e_s, Qabs,
  epsilon = 0.96)
}
\arguments{
\item{r_i}{internal (skin) resistance (s cm^-1)}

\item{r_b}{boundary layer resistance (s cm^-1)}

\item{D}{body diameter (m), ##diameter = 0.0016*log(mass) + 0.0061 for mass(g) #empirical formula for diameter, Riddell et al. 2017}

\item{T_a}{ambient temperature (C)}

\item{elev}{elevation (m)}

\item{e_a}{actual vapor pressure (kPa)}

\item{e_s}{saturation vapor pressure (kPa)}

\item{Qabs}{Solar and thermal radiation absorbed (W)}

\item{epsilon}{emissivity of salamander skin, default epsilon=0.96}
}
\value{
humid operative temperature (C)
}
\description{
Calculate humid operative temperature
}
\details{
This function allows you to calculate humid operative temperature (adaptation of Campbell and Norman 1998).
}
\examples{
\dontrun{
Tb_salamander_humid(r_i=4,r_b=1,D=0.007,T_a=20,elev=500,e_a=2.5,e_s=2.3,Qabs=400, epsilon=0.96)
}

}
\author{
Eric Riddell
}
\keyword{humid}
\keyword{operative}
\keyword{temperature}
