% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TC17.Fig01.R
\name{TC17.Fig01}
\alias{TC17.Fig01}
\title{Produce TC17 Figure 01.}
\usage{
TC17.Fig01()
}
\description{
This function makes all necessary calculations and plots the results for
Figure 01 shown in Muench et al. (2017).
}
\references{
Muench, T., et al., Constraints on post-depositional isotope modifications
in East Antarctic firn from analysing temporal changes of isotope profiles,
The Cryosphere, doi:10.5194/tc-11-2175-2017, 2017.
}
\author{
Thomas Münch
}
