% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TC17.Fig01.R
\name{TC17.Fig01}
\alias{TC17.Fig01}
\title{Produce TC17 Figure 01.}
\usage{
TC17.Fig01(cheat = TRUE)
}
\arguments{
\item{cheat}{Originally, the annual mean isotope data have been erroneously
calculated such that the last value contributing to bin \code{i} was also
included in bin \code{i + 1}, due to an erroneous implementation of
\code{\link{AverageIndexBins}}. This bug has been fixed. For
\code{cheat = TRUE}, the original paper figure with the erroneous bin
definition can be reproduced, while for \code{cheat = FALSE}, the correct
implementation is used. However, the difference of the annual mean data
between the two figure versions is minor and it thus does not influence any
results or conclusions of M端nch et al. (2017).}
}
\description{
This function makes all necessary calculations and plots the results for
Figure 01 shown in M端nch et al. (2017).
}
\references{
M端nch, T., et al., Constraints on post-depositional isotope modifications
in East Antarctic firn from analysing temporal changes of isotope profiles,
The Cryosphere, 11(5), 2175-2188, doi:10.5194/tc-11-2175-2017, 2017.
}
\author{
Thomas M端nch
}
