% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TC17.Fig01.R
\name{TC17.Fig01}
\alias{TC17.Fig01}
\title{TC17 Figure 01.}
\usage{
TC17.Fig01(TR, path = file.path(getwd(), "plots"),
  file.name = "tc17_fig_01", device = "quartz", save.plot = FALSE)
}
\arguments{
\item{TR}{}

\item{path}{}

\item{device}{}

\item{save.plot}{}

\item{"plots")}{}
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Thomas Münch
}
