% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc17-plots.R
\name{TC17.Fig02}
\alias{TC17.Fig02}
\title{Produce TC17 Figure 02.}
\usage{
TC17.Fig02()
}
\description{
This function makes all necessary calculations and plots the results for
Figure 02 shown in Münch et al. (2017).
}
\references{
Münch, T., et al., Constraints on post-depositional isotope modifications
in East Antarctic firn from analysing temporal changes of isotope profiles,
The Cryosphere, 11(5), 2175-2188, doi:10.5194/tc-11-2175-2017, 2017.
}
\author{
Thomas Münch
}
