% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TC17.Fig03.R
\name{TC17.Fig03}
\alias{TC17.Fig03}
\title{TC17 Figure 03.}
\usage{
TC17.Fig03(TR, path = file.path(getwd(), "plots"),
  file.name = "tc17_fig_03", device = "quartz", dev.size = list(h = 6, w =
  8), save.plot = FALSE)
}
\arguments{
\item{TR}{}

\item{path}{}

\item{file.name}{}

\item{device}{}

\item{dev.size}{}

\item{save.plot}{}

\item{"plots")}{}
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Thomas Münch
}
