% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TC17.Fig04.R
\name{TC17.Fig04}
\alias{TC17.Fig04}
\title{Produce TC17 Figure 04.}
\usage{
TC17.Fig04()
}
\description{
This function makes all necessary calculations and plots the results for
Figure 04 shown in Münch et al. (2017).
}
\references{
Münch, T., et al., Constraints on post-depositional isotope modifications
in East Antarctic firn from analysing temporal changes of isotope profiles,
The Cryosphere, doi:10.5194/tc-11-2175-2017, 2017.
}
\author{
Thomas Münch
}
