% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TC17.Fig06.R
\name{TC17.Fig06}
\alias{TC17.Fig06}
\title{Produce TC17 Figure 06.}
\usage{
TC17.Fig06(mod.param = NULL)
}
\arguments{
\item{mod.param}{list of the parameters according to which the original
T13 mean profile is modified to model the temporal changes. It must containt
the following elements: \code{ADV.ind}, \code{ADV.opt}, \code{ADV.only},
\code{SIGMA.ind}, \code{SIGMA.opt}, \code{STRETCH.ind}, and
\code{STRETCH.opt} (see also \code{\link{SetModificationPar}}). If
\code{NULL} (the default), the original data from M端nch et al. (2017) is
used for plotting.}
}
\description{
This function makes all necessary calculations and plots the results for
Figure 06 shown in M端nch et al. (2017).
}
\references{
M端nch, T., et al., Constraints on post-depositional isotope modifications
in East Antarctic firn from analysing temporal changes of isotope profiles,
The Cryosphere, doi:10.5194/tc-11-2175-2017, 2017.
}
\seealso{
\code{\link{SetModificationPar}}
}
\author{
Thomas M端nch
}
