% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TC17.Fig07.R
\name{TC17.Fig07}
\alias{TC17.Fig07}
\title{Produce TC17 Figure 07.}
\usage{
TC17.Fig07(TR, path = file.path(getwd(), "plots"),
  file.name = "tc17_fig_07", device = "quartz", save.plot = FALSE,
  mod.param)
}
\arguments{
\item{TR}{The results from a call to \code{\link{prepareTrenchData()}}.}

\item{path}{The path to the directory in which to save the plot (for
\code{save.plot = TRUE}). Defaults to the folder \code{plots} in the current
working directory. If this folder does not exist, it is attempted to create
with a warning (see also \code{\link{OpenDevice}}).}

\item{file.name}{The name of the file (excluding extension) to save the
plot in.}

\item{device}{The graphics device to be used to display and save the
plot. Defaults to the \code{quartz} device which is the only currently
implemented device option.}

\item{save.plot}{if \code{TRUE}, the plot is saved as a pdf file in the
folder specified by \code{path}. Defaults to \code{FALSE} which results in
on-screen display of the plot.}

\item{mod.param}{A list with the parameters according to which the T13 mean
profile shall be modified. It must containt the following elements:
\code{ADV}, \code{ADVopt}, \code{SIGMA}, \code{SIGMAopt}, \code{stretch},
and \code{stretchOPT}.}
}
\description{
This function makes all necessary calculations and plots the results for
Figure 07 shown in Muench et al. (2017).
}
\references{
Muench, T., et al., Constraints on post-depositional isotope modifications
in East Antarctic firn from analysing temporal changes of isotope profiles,
The Cryosphere Discussions, doi:10.5194/tc-2017-35, 2017.
}
\author{
Thomas Münch
}
