% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{temperature conversions}
\alias{temperature conversions}
\alias{fahrenheit_to_kelvin}
\alias{kelvin_to_celsius}
\alias{celsius_to_kelvin}
\alias{fahrenheit_to_celsius}
\title{Convert Among Temperature Scales}
\usage{
fahrenheit_to_kelvin(temperature)

kelvin_to_celsius(temperature)

celsius_to_kelvin(temperature)

fahrenheit_to_celsius(temperature)
}
\arguments{
\item{temperature}{\code{numeric} temperature (Celsius, Fahrenheit, or Kelvin).}
}
\value{
\code{numeric} temperature (Celsius, Fahrenheit, or Kelvin).
}
\description{
The function converts temperatures among Celsius, Fahrenheit, and Kelvin \insertCite{Blischak}{TrenchR}.
}
\examples{
  kelvin_to_celsius(temperature = 270)
  fahrenheit_to_kelvin(temperature = 85)
  fahrenheit_to_celsius(temperature = 85)
  celsius_to_kelvin(temperature = -10)

}
\references{
\insertAllCited{}
}
\seealso{
Other utility functions: 
\code{\link{airpressure_from_elev}()},
\code{\link{azimuth_angle}()},
\code{\link{day_of_year}()},
\code{\link{daylength}()},
\code{\link{dec_angle}()},
\code{\link{solar_noon}()},
\code{\link{zenith_angle}()}
}
\concept{utility functions}
