% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTRFunctions.R
\name{Thour.sine}
\alias{Thour.sine}
\title{Diurnal temperature across hours
From Campbell and Norman 1998 - Uses sine interpolation}
\usage{
Thour.sine(Tmx, Tmn, Hr)
}
\arguments{
\item{Tmx}{maximum daily temperature in degree celsius}

\item{Tmn}{minimum daily temperature in degree celsius}

\item{Hr}{hour for temperature estimate}
}
\description{
Diurnal temperature across hours
From Campbell and Norman 1998 - Uses sine interpolation
}
\details{
This function allows you to calculate temperature across hours.
}
\examples{
\dontrun{
Thour.sine()
}
}
\keyword{Temperature}
