% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTRfunction.R
\name{Thour.sineexp}
\alias{Thour.sineexp}
\title{Estimate temperature across hours using a diurnal temperature variation function incorporating sine and exponential components}
\usage{
Thour.sineexp(Tmx, Tmn, Hr, tr, ts, alpha = 2.59, beta = 1.55,
  gamma = 2.2)
}
\arguments{
\item{Tmx}{maximum daily temperature (C)}

\item{Tmn}{minimum daily temperature (C)}

\item{Hr}{hour for temperature estimate}

\item{tr}{time of sunset (hour)}

\item{ts}{time of sunrise (hour)}

\item{alpha}{time difference between tx(time of maximum temperature) and noon}

\item{beta}{time difference between tx and sunrise}

\item{gamma}{decay parameter for rate of t change from sunset to tn(time of minimum temp)}
}
\description{
Estimate temperature across hours using a diurnal temperature variation function incorporating sine and exponential components
}
\details{
This function allows you to calculate temperature across hours using a diurnal temperature variation function incorporating sine and exponential components.
}
\examples{
\dontrun{
Thour.sineexp()
}
}
\keyword{Temperature}
