% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPCfunctions.R
\name{TPC}
\alias{TPC}
\title{Construct a Gaussian-Quadratic Thermal Performance Curve}
\usage{
TPC(T, Topt, CTmin, CTmax)
}
\arguments{
\item{T}{\code{numeric} vector of temperature range (C).}

\item{Topt}{\code{numeric} thermal optima (C), the temperature at which peak performance occurs.}

\item{CTmin}{\code{numeric} critical thermal minima (C), the lower temperature limit for performance.}

\item{CTmax}{\code{numeric} critical thermal maxima (C), the upper temperature limit for performance.}
}
\value{
performance
}
\description{
Constructs a thermal performance curve by combining as a Gaussian function to describe the rise in performance up to the optimal temperature and a quadratic decline to zero performance at critical thermal maxima and higher temperatures \insertCite{Deutsch2008}{TrenchR}.
}
\examples{
  TPC(T     = 0:60, 
      Topt  = 30, 
      CTmin = 10, 
      CTmax = 40)

}
\references{
\insertAllCited{}
}
