% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenchR.R
\docType{package}
\name{TrenchR}
\alias{TrenchR}
\title{Translating Environmental Change Into Organismal Responses}
\description{
The TrenchR package aids in translating environmental change into organismal responses. The package includes microclimate models to vertically scale weather station data to organismal heights. The biophysical modeling tools include both general models for heat flows and specific models to predict body temperatures for a variety of ectothermic taxa. Additional functions model and temporally partition air and soil temperatures and solar radiation. We offer utility functions to aid in estimating the organismal and environmental parameters needed for biophysical ecology. TrenchR focuses on relatively simple and modular functions so users can create transparent and flexible biophysical models.
}
\keyword{package}
