% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance_functions.R
\name{Tsoil}
\alias{Tsoil}
\title{Statistical approximation of soil temperature}
\usage{
Tsoil(Tg_max, Tg_min, hour, depth)
}
\arguments{
\item{Tg_max}{daily maximum soil surface temperature (C)}

\item{Tg_min}{daily minimum soil surface temperature (C)}

\item{depth}{depth (cm) ???}
}
\value{
soil temperature (C)
}
\description{
Statistical approximation of soil temperature
}
\details{
This function allows you to estimate soil temperature at a given depth and hour approximating diurnal variation as sinusoidal (adapted from Campbell and Norman 1998).
}
\examples{
\dontrun{
Tsoil(Tg_max=30, Tg_min=15, hour=12, depth=5)
}

}
\author{
Eric Riddell
}
\keyword{soil}
\keyword{temperature}
