% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRYWETAIR.R
\name{VAPPRS}
\alias{VAPPRS}
\title{Calculate Saturation Vapor Pressure}
\usage{
VAPPRS(db = db)
}
\arguments{
\item{db}{\code{numeric} Dry bulb temperature (C)}
}
\value{
esat \code{numeric} Saturation vapor pressure (Pa)
}
\description{
Calculates saturation vapor pressure for a given air temperature.
}
\examples{
  VAPPRS(db = 30)

}
\references{
\insertAllCited{}
}
