% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRYWETAIR.R
\name{VAPPRS}
\alias{VAPPRS}
\title{VAPPRS}
\usage{
VAPPRS(db = db)
}
\arguments{
\item{db}{Dry bulb temperature (C)}
}
\value{
esat Saturation vapour pressure (Pa)
}
\description{
Calculates saturation vapour pressure for a given air temperature.
}
\examples{

VAPPRS(db=30)

}
