% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRYWETAIR.R
\name{VAPPRS}
\alias{VAPPRS}
\title{Saturation Vapor Pressure}
\usage{
VAPPRS(db)
}
\arguments{
\item{db}{\code{numeric} Dry bulb temperature (C)}
}
\value{
esat \code{numeric} Saturation vapor pressure (Pa)
}
\description{
The function calculates saturation vapor pressure for a given air temperature. The program is based on equations from \insertCite{List1971;textual}{TrenchR} and code implementation from NicheMapR \insertCite{Kearney2017,Kearney2020}{TrenchR}.
}
\examples{
  VAPPRS(db = 30)

}
\references{
\insertAllCited{}
}
