% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRYWETAIR.R
\name{WETAIR}
\alias{WETAIR}
\title{WETAIR}
\usage{
WETAIR(db = db, wb = db, rh = 0, dp = 999, bp = 101325)
}
\arguments{
\item{db}{\code{numeric} Dry bulb temperature (C)}

\item{wb}{\code{numeric} Wet bulb temperature (C)}

\item{rh}{\code{numeric} Relative humidity (\%)}

\item{dp}{\code{numeric} Dew point temperature (C)}

\item{bp}{\code{numeric} Barometric pressure (pascal)}
}
\value{
e \code{numeric} Vapour pressure (Pa)

esat \code{numeric} Saturation vapour pressure (Pa)

vd \code{numeric} Vapour density (kg m-3)

rw \code{numeric} Mixing ration (kg kg-1)

tvir \code{numeric} Virtual temperature (K)

tvinc \code{numeric} Virtual temperature increment (K)

denair \code{numeric} Hourly predictions of the soil moisture under the maximum specified shade

cp \code{numeric} Specific heat of air at constant pressure (J kg-1 K-1)

wtrpot \code{numeric} Water potential (Pa)

rh \code{numeric} Relative humidity (%)
}
\description{
Calculates several properties of humid air as output variables below. The program is based on equations from \insertCite{List1971}{TrenchR}. WETAIR must be used in conjunction with function VAPPRS. Input variables are shown below. The user must supply known values for DB and BP (BP at one standard atmosphere is 101,325 pascals). Values for the remaining variables are determined by whether the user has either (1) psychrometric data (WB or RH), or (2) hygrometric data (DP)

(1) Psychrometric data:
If WB is known but not RH, then set RH = -1 and DP = 999.
If RH is known but not WB then set WB = 0 and DP = 999.
(2) Hygrometric data:
If DP is known then set WB = 0 and RH = 0.
}
\examples{
WETAIR(db=30, wb=28, rh=60, bp=100*1000)

}
\references{
\insertAllCited{}
}
