% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.peaks.R
\name{which.peaks}
\alias{which.peaks}
\title{Find Maxima or Minima.}
\usage{
which.peaks(x, minima = FALSE, partial = FALSE)
}
\arguments{
\item{x}{Numeric vector for which maxima (minima) shall be analysed.}

\item{minima}{if \code{FALSE} (the default) return the positions of the
maxima of \code{x}, else return the positions of the minima.}

\item{partial}{if \code{TRUE}, the first and last data points of \code{x}
are considered to be maxima or minima as well; defaults to \code{FALSE}.}
}
\value{
Numeric vector with the index positions of the maxima (minima) of
\code{x}.
}
\description{
Find the positions of the maxima or minima in a data vector.
}
\author{
Thomas Münch
}
