% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempWindProfileFunctions.R
\name{wind_speed_profile}
\alias{wind_speed_profile}
\title{Estimate wind speed profile as in NicheMapR.}
\usage{
wind_speed_profile(u_r, zr, z0, z)
}
\arguments{
\item{u_r}{is wind velocity at reference height in m/s.}

\item{zr}{is initial reference height in m.}

\item{z0}{is surface roughness in m.}

\item{z}{is height to scale to in m.}
}
\value{
wind speed (m/s)
}
\description{
This function allows you to estimate wind speed (m/s) at a specified height (m).  
Estimates a single, unsegmented wind velocity using the MICRO routine from NicheMapR as described in Kearney and Porter 2016.
Source: Kearney MR and Porter WP. 2016. NicheMapR – an R package for biophysical modelling: the microclimate model. Ecography 40:664-674.
Section 5. Vertical air temperature and wind speed profiles, p11
}
\details{
Calculates wind speed at a specified height
}
\examples{
\dontrun{
wind_speed_profile(u_r=0.1, zr=0.1, z0=0.2, z=0.15)
}
}
\seealso{
Other microclimate functions: \code{\link{air_temp_profile_neutral}},
  \code{\link{air_temp_profile_segment}},
  \code{\link{air_temp_profile}},
  \code{\link{degree_days}},
  \code{\link{diurnal_radiation_variation}},
  \code{\link{diurnal_temp_variation_sineexp}},
  \code{\link{diurnal_temp_variation_sinesqrt}},
  \code{\link{diurnal_temp_variation_sine}},
  \code{\link{estimate_radiation}},
  \code{\link{monthly_solar_radiation}},
  \code{\link{partition_solar_radiation}},
  \code{\link{surface_roughness}},
  \code{\link{wind_speed_profile_neutral}},
  \code{\link{wind_speed_profile_segment}}
}
\concept{microclimate functions}
\keyword{profile}
\keyword{wind}
