% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempWindProfileFunctions.R
\name{wind_speed_profile_neutral}
\alias{wind_speed_profile_neutral}
\title{Estimate Wind Speed at a Specific Height Under Neutral Conditions}
\usage{
wind_speed_profile_neutral(u_r, zr, z0, z)
}
\arguments{
\item{u_r}{\code{numeric} wind velocity at reference height (m/s)}

\item{zr}{\code{numeric} initial reference height (meters).}

\item{z0}{\code{numeric} surface roughness (meters).}

\item{z}{\code{numeric} height to scale (meters).}
}
\value{
windspeed (m/s)
}
\description{
Calculate wind speed (m/s) at a specified height (m) within a boundary layer near the surface.  The profile assumes neutral conditions. The velocity profile is the neutral profile described by \insertCite{Sellers1965}{TrenchR}. Function is equations (2) and (3) of \insertCite{Porter1973}{TrenchR}.
}
\examples{
  wind_speed_profile_neutral(u_r = 0.1, 
                             zr  = 0.1, 
                             z0  = 0.2, 
                             z   = 0.15)

}
\references{
\insertAllCited{}
}
\seealso{
Other microclimate functions: 
\code{\link{air_temp_profile_neutral}()},
\code{\link{air_temp_profile_segment}()},
\code{\link{air_temp_profile}()},
\code{\link{degree_days}()},
\code{\link{direct_solar_radiation}()},
\code{\link{diurnal_radiation_variation}()},
\code{\link{diurnal_temp_variation_sineexp}()},
\code{\link{diurnal_temp_variation_sinesqrt}()},
\code{\link{diurnal_temp_variation_sine}()},
\code{\link{estimate_radiation}()},
\code{\link{monthly_solar_radiation}()},
\code{\link{partition_solar_radiation}()},
\code{\link{proportion_diffuse_solar_radiation}()},
\code{\link{surface_roughness}()},
\code{\link{wind_speed_profile_segment}()}
}
\concept{microclimate functions}
\keyword{profile}
\keyword{wind}
