% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YearMean.R
\name{YearMean}
\alias{YearMean}
\title{Annual means from binning.}
\usage{
YearMean(x, ind, na.rm = TRUE)
}
\arguments{
\item{x}{}
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Thomas Münch
}
