% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{zenith_angle}
\alias{zenith_angle}
\title{Calculate Zenith Angle}
\usage{
zenith_angle(doy, lat, lon, hour)
}
\arguments{
\item{doy}{is day of year.}

\item{lat}{is latitude in degrees.}

\item{lon}{is longitude in degrees.}

\item{hour}{is hour of the day.}
}
\value{
Zenith angle in degrees
}
\description{
This function allows you to calculate the zenith angle, the location of the sun as an angle (in degrees) measured from vertical
}
\details{
Calculate Zenith angle
}
\examples{
\dontrun{
zenith_angle(112, 47.61, -122.33, 12)
}
}
\keyword{Zenith}
\keyword{angle}
