% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{zenith_angle}
\alias{zenith_angle}
\title{Calculate Zenith Angle}
\usage{
zenith_angle(doy, lat, lon, hour, offset = NA)
}
\arguments{
\item{doy}{is day of year.}

\item{lat}{is latitude in decimal degrees.}

\item{lon}{is longitude in decimal degrees.}

\item{hour}{is hour of the day.}

\item{offset}{is the number of hours to add to UTC to get local time (to improve accuracy but not always necessary)}
}
\value{
Zenith angle in degrees
}
\description{
This function allows you to calculate the zenith angle, the location of the sun as an angle (in degrees) measured from vertical. Source: Campbell and Norman. 1998. An Introduction to Environmental Biophysics.
}
\details{
Calculate Zenith angle
}
\examples{
\dontrun{
zenith_angle(doy=112, lat=47.61, lon=-122.33, hour=12)
}
}
\seealso{
Other utility functions: \code{\link{airpressure_from_elev}},
  \code{\link{azimuth_angle}}, \code{\link{daylength}},
  \code{\link{dec_angle}}, \code{\link{degree_to_radian}},
  \code{\link{fahrenheit_to_celsius}},
  \code{\link{fahrenheit_to_kelvin}},
  \code{\link{kelvin_to_celsius}}, \code{\link{solar_noon}}
}
\concept{utility functions}
\keyword{Zenith}
\keyword{angle}
